% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abund.manycensus}
\alias{abund.manycensus}
\title{Collect abundances of all species across several censuses.}
\usage{
abund.manycensus(allcns = list(bci.full1, bci.full2, bci.full3), mindbh,
  dbhunit = "mm", type = "abund", excludespp = NULL,
  excludestatus = NULL)
}
\arguments{
\item{allcns}{Full R census, submitted as a list (as many as desired).}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{dbhunit}{'cm' or 'mm', only used for basal area}

\item{type}{Either 'abund' (default) for abundance, 'ba' for basal area, or
'agb'.}

\item{excludespp}{A character vector of species codes can be submitted as
\code{excludespp}, for instance those for unidentified trees.}

\item{excludestatus}{By default, any tree ever given code M is not counted in
any census, but set \code{excludestatus = NULL} to include them.}
}
\description{
Collect abundances of all species across several censuses.
}
\examples{
\dontrun{
N = abund.manycensus(
  allcns = list(
    bci::bci12full1, bci::bci12full2, bci::bci12full3, bci::bci12full4
  ),
  mindbh = 10,
  type = 'abund',
  excludespp = 'uniden',
  excludestatus = 'M'
)
head(N)
colSums(N)
apply(N,2,countspp)

N = abund.manycensus(
  allcns = list(bci::bci12full5, bci::bci12full6),
  mindbh = 10,
  type = 'abund',
  excludespp = c('uniden', 'tremxx'),
  excludestatus = NULL
)
}

}
