% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{addBinParam}
\alias{addBinParam}
\title{addBinParam}
\usage{
addBinParam(x, best, bin)
}
\description{
Given parameters for a model with N linear bins, creates parameters for N+1 bins which produce the same model.
}
