% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.CTFSdb.R
\name{agb.model}
\alias{agb.model}
\title{Calculates biomass from density, height, and dbh.}
\usage{
agb.model(dbh, density, height, param)
}
\arguments{
\item{dbh}{Stem diameter, usually at breast height (see data dictionary at 
http://ctfs.si.edu/Public/DataDict/data_dict.php).}
}
\description{
Calculates biomass from density, height, and dbh. Requires just two
parameters, following Chave (2005). The parameters can be changed, but the
formula cannot be. Returns a vector of biomass as long as vector of dbh
submitted.

This is called by Chave.AGB in the standard calculation of biomass for CTFS R
tables.
}
\examples{
\dontrun{
agb.model(
  dbh = c(1, 1, 2),
  density = c(.6, .6, .5),
  height = c(2, 3, 4),
  param = c(.0501, 1)
)
}



}
