% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.CTFSdb.R
\name{AGB.tree}
\alias{AGB.tree}
\title{Computes AGB of each tree in a table, grouping all stems of one tre...}
\usage{
AGB.tree(df, dbhunit = "mm", plot = "bci", wsgdata = wsg.ctfs2,
  forest = "moist", ht.param = NULL, htmodel = predht.asym)
}
\description{
Computes AGB of each tree in a table, grouping all stems of one tree and adding there agbs.

The submitted table, df, must have dbh, species name (sp),
and a treeID to identify which tree every stem belong to. There must be just one dbh for each stem.  Returns
a dataframe with one row per tree, including the treeID and total agb per tree. Note that it will have fewer rows
than the table submitted. This is called by biomass.CTFSdb in the standard calculation of biomass for CTFS R tables.

biomasstbl=AGB.tree(df=bci.stem1)
dim(bci.stem1)
dim(biomasstbl)
head(biomasstbl)
}
