% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography_slope.R
\name{allquadratslopes}
\alias{allquadratslopes}
\title{allquadratslopes}
\usage{
allquadratslopes(elev, gridsize = 20, plotdim = c(1000, 500),
  edgecorrect = TRUE)
}
\description{
Calculating slope of all quadrats in a plot
calcslope takes 3 elevations and finds the slope of the plane through them
quadslope divides the 4 corners of a quadrat into 4 different groups of 3 stakes,
takes the slope of each, then averages
these were first written in C++
see slopeelev.cpp for more on the geometry
allquadratslopes goes through all 20x20 m quadrats in a plot and finds
the slope, mean elevation, and convexity of each
convexity is the mean elevation of one 20x20 m quadrat relative (minus) the mean of its
immediate neighbors
Helene Muller-Landau added a section to correct convexity in edge quadrats
}
