% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_geometry.R
\name{are.ptsinside}
\alias{are.ptsinside}
\title{are.ptsinside}
\usage{
are.ptsinside(pts, coord)
}
\description{
Checks many points (dataframe pt with x and y) against a single quadrat whose corners are given by as xlo, ylo, xhi, yhi.
It returns a logical vector, TRUE for the points inside. This is same as insideRectange, but accepting input as a matrix pts
and a single vector of the four corners of the rectange.
}
