% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-utils.R
\name{args_compare}
\alias{args_compare}
\alias{args_intersect}
\alias{args_setdiff}
\alias{args_intersect}
\alias{args_setdiff}
\title{Compare arguments between two functions.}
\usage{
args_intersect(x, y)

args_setdiff(x, y)
}
\arguments{
\item{x, y}{String giving function name.}
}
\value{
Intersection or difference in arguments between functions.
}
\description{
Compare arguments between two functions.
}
\examples{
args_intersect("AGB.tree", "biomass.CTFSdb")
args_setdiff("AGB.tree", "biomass.CTFSdb") 
}
\keyword{internal}
