% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{arrangeParam.llike}
\alias{arrangeParam.llike}
\title{arrangeParam.llike}
\usage{
arrangeParam.llike(testparam, allparam, whichtest)
}
\description{
Used in likelihood function of a Gibbs sampler. Allows any of a set of parameters to be submitted to metrop1step; 
whichtest is the index of the parameter to test. If NULL, zero, or NA, it simply returns allparam.
}
