% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{assembleBinOutput}
\alias{assembleBinOutput}
\title{Use the list output from piecewise regression (growthfit.bin) and c...}
\usage{
assembleBinOutput(inputtable, fulldata, sitename)
}
\description{
Use the list output from piecewise regression (growthfit.bin) and converts to a flat table. The single argument
inputdata is the output compare.growthbinmodel.

Written by Adrian Das.
(excluding burn in)
}
\examples{
\dontrun{
bciresult=compare.growthbinmodel(linearbin.fit.trim3,export=NULL,makegraph=FALSE)
bcitable=assembleBinOutput(bciresult,sitename='bci_12_spreadSD')
write.table(bcitable,file='growth/BCIpiecewise.txt',quote=FALSE,sep='\\t',row.names=FALSE) }

}
