% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{asymp.ht.fixmax}
\alias{asymp.ht.fixmax}
\title{asymp.ht.fixmax}
\usage{
asymp.ht.fixmax(x, param, asymp)
}
\description{
Same formulation, but the asymptote is fixed, so only two parameters fitted.
}
