% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{asymp.ht}
\alias{asymp.ht}
\title{asymp.ht}
\usage{
asymp.ht(x, param, ...)
}
\description{
The function from Sean Thomas which produces an asymptote for y as a function of x.

Original version: y=ymax*(1-exp(-a*x^b))

This is the centered version, with x normalized by dividing by parameter k, which is the x value at which
y is half ymax. This eliminates correlation between the a and b parameters in the above version, but
not the correlation between parameters 1 and 2.
}
