% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{asymptote.exp}
\alias{asymptote.exp}
\title{asymptote.exp}
\usage{
asymptote.exp(x, param, ...)
}
\description{
A 3-parameter function which asymptotes as x->infinity. The 3rd param must be >=0 and x>=0. The asymptote is a, the intercept a-b.
}
