% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{attach_if_needed}
\alias{attach_if_needed}
\title{attach_if_needed}
\usage{
attach_if_needed(datafiles)
}
\description{
Deprecated. See section \emph{Good practice} in \code{?attach}. To use <DATA> from BCI
with bci::<DATA>, see
\url{https://github.com/forestgeo/bci}.
}
\details{
Attach one or more datafiles, checking first whether the file is already
attached. If it is attached, it is not reattached, and the search position
where attached is returned.
}
