% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{autoregression}
\alias{autoregression}
\title{Autocorrelation with a given lag of a vector y.autoregression regre...}
\usage{
autoregression(y, lag, xlog = FALSE, ylog = FALSE, graphit = TRUE)
}
\description{
Autocorrelation with a given lag of a vector y.
}
