% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogchange_abund.fit.CTFS.R
\name{bad.asymexp.param}
\alias{bad.asymexp.param}
\title{bad.asymexp.param}
\usage{
bad.asymexp.param(hyperparam)
}
\description{
For either the Gaussian, or asymexp, the SD parameters must be > 0. Only used as a subroutine of the main modeling function, model.littleR.Gibbs.
}
