% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund.fit.CTFS.R
\name{bad.asympower.param}
\alias{bad.asympower.param}
\title{bad.asympower.param}
\usage{
bad.asympower.param(hyperparam)
}
\description{
The 3 parameters submitted to hyper.abundGibbs have to be checked, in case dasympower is used. The second and third, the
two rate parameters, have to be < (-1). Since the parameters are inverted, they must be in (-1,0). 
Only used as a subroutine of the main modeling function, model.littleR.Gibbs.
}
