% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{bad.binsdpar}
\alias{bad.binsdpar}
\title{bad.binsdpar calculateBinModel.BIC  Calculate Bayes Information Cri...}
\usage{
bad.binsdpar(x, param, ...)
}
\description{
bad.binsdpar calculateBinModel.BIC  Calculate Bayes Information Cri...
}
