% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{bad.paretopar}
\alias{bad.paretopar}
\title{Test whether parameters for the Pareto distribution are acceptable....}
\usage{
bad.paretopar(param)
}
\description{
Test whether parameters for the Pareto distribution are acceptable.
}
