% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{BadParam}
\alias{BadParam}
\title{This is a default for model.xy, never returning TRUE. To use model....}
\usage{
BadParam(x, param, pred)
}
\description{
This is a default for model.xy, never returning TRUE. To use model.xy, another function must be created to
return TRUE for any illegal parameter combination (any that would, for instance, produce a predicted y out-of-range, or a would
create an error in the likelihood function.)
}
