% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerBayes.R
\name{badSD}
\alias{badSD}
\title{badSD arrangeParam.llike.2D  Used in likelihood function of a Gibbs...}
\usage{
badSD(x, param, ...)
}
\arguments{
\item{...}{Unused.}
}
\description{
badSD arrangeParam.llike.2D  Used in likelihood function of a Gibbs...
}
