% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{beta.normalized}
\alias{beta.normalized}
\title{Normalzing beta.total. No longer used. beta.normalized beta.total  ...}
\usage{
beta.normalized(x, par1, xmin, par2, xmax)
}
\description{
Normalzing beta.total. No longer used.
}
