% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_distributions.R
\name{beta.total}
\alias{beta.total}
\title{beta.total}
\usage{
beta.total(x, par1, xmin, par2, xmax)
}
\description{
A beta distribution on the interval xmin to xmax, instead of 0 to 1. No longer used.
}
