% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.graph.R
\name{binGraphManySpecies.Panel}
\alias{binGraphManySpecies.Panel}
\title{binGraphManySpecies.Panel}
\usage{
binGraphManySpecies.Panel(fulldataname, sitename, darken = 8, xrange = c(-3,
  3), yrange = c(0, 0.1), export = pdf,
  outfile = "growth/linearbin.multi.pdf", h = 8, w = 10)
}
\description{
Make a graph, 4 panels, of AGB growth and model fit of many species overlaid, predicted functions only. Must submit names of four data objects
(not the object, just the name in quote marks!).
}
\examples{
\dontrun{
binGraphManySpecies.Panel(c('linearbin.fit.allspp','linearbin.fit.trim3','linearbin.fit.edoro3','linearbin.fit.lenda3'),
                         export=pdf,yrange=c(0,.25),sitename=c('Fushan','BCI','Edoro','Lenda'),darken=5,xrange=c(-3,3.5))}


}
