% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_growthfit.graph.R
\name{binGraphManySpecies}
\alias{binGraphManySpecies}
\title{binGraphManySpecies}
\usage{
binGraphManySpecies(fulldataname, darken = 8, xrange = c(-3, 3),
  yrange = c(0, 0.1), xtitle = NULL, ytitle = NULL, export = pdf,
  outfile = "growth/linearbin.multi.pdf", h = 8, w = 10)
}
\description{
Make a graph of AGB growth and model fit of many species overlaid, predicted functions only. Must submit name of data objects 
(not the object, just the name in quote marks!).
}
