% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.graph.R
\name{binGraphSampleSpecies}
\alias{binGraphSampleSpecies}
\title{binGraphSampleSpecies}
\usage{
binGraphSampleSpecies(fulldataname, species, whichbin, export = pdf,
  outfile = "growth/linearbin.summary.pdf", h = 8, w = 10)
}
\description{
Make a single graph, 4 panels, of AGB growth and model fit. Must submit four species names (though they could be
the same. Enter names of data objects (not the object, just the name in quote marks!) either one data object, or a list of four such objects;
if only one, all four graphs are based on the single one. Likewise, enter either one bin number, or four.
}
\examples{
\dontrun{
attach('growth/linearbin.fittrim3.bci.rdata')
attach('growth/growthfitYiching/linearbin.fit.allspp.rdata')
attach('growth/linearbin.fit.edoro.rdata')
attach('growth/linearbin.fit.lenda.rdata')
binGraphSampleSpecies(fulldata='linearbin.fit.allspp',species=c('pri2co','tri2tu','brosal','jac1co'),whichbin=2,export=NULL)
binGraphSampleSpecies(fulldata=c('linearbin.fit.allspp','linearbin.fit.trim3','linearbin.fit.edoro3','linearbin.fit.lenda3'),
                      species=c('PYRESH','pri2co','JULBSE','GILBDE'),whichbin=3,export=NULL)
binGraphSampleSpecies(fulldata='linearbin.fit.allspp',species=c('pri2co','pri2co','brosal','brosal'),whichbin=c(2,3,2,3),export=NULL)}


}
