% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.CTFSdb.R
\name{biomass.CTFSdb}
\alias{biomass.CTFSdb}
\title{Biomass for trees and stems using dbh allometry.}
\usage{
biomass.CTFSdb(RStemTable, RTreeTable, whichtable = "tree", dbhunit = "mm",
  plot = "bci", wsgdata = wsg.ctfs2, forest = "moist", ht.param = NULL,
  htmodel = predht.asym, useChave = TRUE, cnsno = NULL, dbname = NULL,
  fullname = NULL, plotcode = NULL)
}
\arguments{
\item{dbhunit}{'cm' or 'mm', only used for basal area}

\item{plot}{Character string giving the name of the plot where the data comes
from (as it appears in the names of the R Analytical Tables).}
}
\value{
The same table as submitted (either tree or stem), with a column agb added;
if the agb column was already present, it is replaced.
}
\description{
Calculate biomass from existing R-formatted tables for trees and stems using
dbh allometry. By default, it uses the Chave (2005) equations.
}
\details{
Note that the standard downloads of R Analytical Tables, already has the agb
column filled, calculated with this routine using the default parameters for
moist forest.

This program can be used to repeat or redo the calculation. It can also be
used for other tables, always requiring two tables:  one with trees and one
with all stems.

Calculations are done by AGB.ind and the subroutines it call.

An alternative option is to have AGB calculations already stored in the
server's AGB database, setting \code{useChave = FALSE}. This function then looks
up the AGB for each stem from the table named for the plot.

RStemTable: Name of table with one row per stem; must have dbh, species
(column sp), treeID

RTreeTable: Name of table with one row per tree; must have dbh, species
(column sp), treeID whichtable: Set to "tree" to return the entire tree table
with agb, or to "stem" to return stem table dbhunit: Set to "mm", "cm", or
"inch" plot: Name of plot, matching the plot name used in the wood-density
table wsgdata: Name of R object having wood-density for species in all CTFS
plots forest: Set to "moist", "dry", or "wet" to use the equations publshed
in Chave (2005) for the 3 forest types ht.param: A vector of parameters for a
formula relating tree height to dbh; if NULL, the biomass formula does not
use height htmodel: Name of an R function that returns tree height giving a
dbh and any number of parameters; if ht.param is NULL, htmodel is ignored.

If you need a dummy wood density table to feed the wsg argument see
?\code{\link[=wsgdata_dummy]{wsgdata_dummy()}}.
}
\examples{
\dontrun{
CTFSplot("bci","full",census=1) 

CTFSplot("bci","stem",census=1) 
attach("biomass/wsg.ctfs.Rdata") 
newtable = biomass.CTFSdb(
  RStemTable = bci::bci12stem1, 
  RTreeTable = bci::bci12full1
)
}

}
