% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{bootconf}
\alias{bootconf}
\title{A simple calculation of confidence limits based on the SD of a vect...}
\usage{
bootconf(x)
}
\description{
A simple calculation of confidence limits based on the SD of a vector.
}
