% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{bootstrap.corr}
\alias{bootstrap.corr}
\title{bootstrap.corr}
\usage{
bootstrap.corr(dataset, xcol = (-1), ycol = (-1), xcolname = "no",
  ycolname = "no", boot = 100)
}
\description{
Running bootstrap on a correlation. Any columsn can be chosen from the submitted dataset, by number or name.
}
