% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{calc.gradient}
\alias{calc.gradient}
\title{Calculate flow using Seibert & McGlynn algorithm.}
\usage{
calc.gradient(elev, grid = 20)
}
\arguments{
\item{elev}{A 3x3 matrix of elevations; works on central point.

This calculates the gradient for the 8 triangular facets around the center
point, following Seibert & McGlynn

The output is a data.frame of direction and slope for the 8 facets, starting
with the lower left and moving clockwise}
}
\description{
Calculate flow using Seibert & McGlynn algorithm.
}
\examples{
\dontrun{
data <- c(268.7, 275.9, 283.2, 275.9, 282.8, 290.0, 283.2, 290.0, 297)
z <- matrix(
  data,
  nrow = 3,
  byrow = TRUE
)
calc.gradient(z)
}
}
