% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcalpha.R
\name{calcalpha}
\alias{calcalpha}
\title{Calculates alpha from a stem and species count submitted using Newt...}
\usage{
calcalpha(n, s)
}
\description{
Calculates alpha from a stem and species count submitted using Newton's method.
}
