% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogChange.R
\name{calcMortIndivTable}
\alias{calcMortIndivTable}
\title{calcMortIndivTable}
\usage{
calcMortIndivTable(mtable, by = "species")
}
\description{
Calculate mortality rate per species per census interval using the output of individual_mort.table. 
Formerly named calcMortlmerTable.
}
