% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcalpha.R
\name{calcS.alpha}
\alias{calcS.alpha}
\title{Number of species in a sample of N individuals, given Fisher's alpha.}
\usage{
calcS.alpha(n, a)
}
\arguments{
\item{n}{Individuals count.}

\item{a}{Fisher's alpha.}
}
\description{
Estimated number of species in a sample of N individuals, given Fisher's alpha.
}
