% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{calcslope}
\alias{calcslope}
\title{Given the z-coordinate of 3 points in 3D space whose x-y coordinate...}
\usage{
calcslope(z, gridsize = 20)
}
\arguments{
\item{z}{numeric vector of length 3}

\item{gridsize}{Side of the square quadrate.}
}
\description{
Given the z-coordinate of 3 points in 3D space whose x-y coordinates are
corners of a right, isoceles triangle whose short side = gridsize, this
calculates the slope in degrees of the plane through the points.

\code{calcslope()} takes 3 elevations and finds the slope of the plane through
them.

The second point is the one between the two short sides.
}
