% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{calculateBinModel.AIC}
\alias{calculateBinModel.AIC}
\title{Calculate AIC of the model, using various log(likelihood) estimator...}
\usage{
calculateBinModel.AIC(fit, type = "optim")
}
\description{
Calculate AIC of the model, using various log(likelihood) estimators:
with optim, the highest likelihood found by optim
with mean, the mean llikelihood from the Gibbs sampler
with gibbs, the maximum llikelihood from the Gibbs sampler
with none, just return the mean Gibbs likelihood
}
