% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{calculateBinModel.bestpred}
\alias{calculateBinModel.bestpred}
\title{Calculate mean predicted value at every x using every one of the Gi...}
\usage{
calculateBinModel.bestpred(fit)
}
\arguments{
\item{fit}{A fitted model.}
}
\description{
Calculate mean predicted value at every x using every one of the Gibbs sampler
parameter combinations (excluding burn in)
}
