% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{cartesian.to.polar}
\alias{cartesian.to.polar}
\title{Convert Cartesian coordinates to polar. Returns a dataframe of two ...}
\usage{
cartesian.to.polar(x, y)
}
\arguments{
\item{x}{Coordinates.}

\item{y}{input vector of y coordinates (same length as x)}
}
\description{
Convert Cartesian coordinates to polar. Returns a dataframe of two columns named r and theta. This always
returns a theta between -pi/2 and pi/2. Note that polar.to.cartesian may not return the starting x, y submitted
to cartesian.to.polar due to problems with signs. It will work if theta is kept positive (ie, quadrant 1).
}
\examples{
\dontrun{
cartesian.to.polar(2,4)}


}
