% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{center.predictors}
\alias{center.predictors}
\title{center.predictors}
\usage{
center.predictors(x, by = "mean")
}
\description{
Transform all data by subtracting a constant, either the mean, median value, or a submitted constant.

The input may be a vector or a matrix. If a matrix, each column is centered on its mean (median), or by passing a
vector of constants. Note that setting by=0 amounts to no change.
}
\examples{
\dontrun{
center.predictors(x=c(1,4,7,0),by='median')
center.predictors(x=c(1,4,7,0),by=2)}


}
