% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{charlocate}
\alias{charlocate}
\title{Finds position of a substring needle inside a longer string haystac...}
\usage{
charlocate(needle, haystack)
}
\description{
Finds position of a substring needle inside a longer string haystack. Can return more than one position if the needle appears more
than once.  Returns 0 if no matches. The input are atomic: this is not vectorized.
}
\examples{
\dontrun{
charlocate("19","x190019xxx")}


}
