% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{CI}
\alias{CI}
\title{CI}
\usage{
CI(x, prob = c(0.025, 0.975), na.rm = FALSE)
}
\description{
Confidence limits (quantiles) from a vector at specified probabilities. Default is 95% confidence interval.
}
