% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesCTFS.R
\name{clean.demography}
\alias{clean.demography}
\title{Keep or clean rows outputed by mortality, growth, or recruitment?}
\usage{
clean.demography(demogdata, type = "mort", Ncol = NULL, Scol = NULL,
  timecol = NULL, ratecol = NULL, excludespp = NULL)
}
\arguments{
\item{type}{"mort" or "abund" to refer to columns by name or number.}
}
\value{
A logical vector indicating with TRUE which rows to keep, not the cleaned
table itself.
}
\description{
This takes a CTFS demography table, output by functions mortality, growth, or
recruitment, and removes rows where N == 0, or key data are NA.
}
\details{
The rownames are assumed to refer to species names, and some codes can be
excluded using the argument excludespp. The four columns can be submitted by
name or number using default column headers, or by setting type to 'mort'or
'abund'.
}
\examples{
\dontrun{
rows.include = clean.demography(demogtable,
  type = 'growth',
  Ncol = 'N',
  ratecol = 'ratecol')

result = demogtable[rows.include, ]
}

}
