% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{colMedians}
\alias{colMedians}
\title{For convenient medians, like colMeans.colMedians midPoint  Midpoint...}
\usage{
colMedians(mat, na.rm = TRUE)
}
\description{
For convenient medians, like colMeans.
}
