% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{complete.plotmap}
\alias{complete.plotmap}
\title{Map every individual.}
\usage{
complete.plotmap(cns = bci.full6, spnames = NULL, mindbh = 10,
  export = "no", nospp = 3, plotdim = c(1000, 500), clrlist = c("blue",
  "green", "red", "yellow", "gray"), ptsize = c(0.45, 0.3), xrange = c(0,
  100), yrange = c(0, 100), wd = 1100, ht = 850, side = 6,
  labsize = 1.75, axisdiv = 10, filepath = "/home/condit/data/maps/",
  outfile = "fullplotmap")
}
\arguments{
\item{cns}{a full census dataset (all species)}

\item{spnames}{names of species to map, using the mnemonic in the R tables}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{export}{* If `export = "no"`, does nothing, so the next graph uses the default R
device. With this option, the size of the graph cannot have been set larger
than the default size, which is usually 7 inches. 
* If `export = "unix"`, a new graphics window in unix, size `h` and width `w`,
will be opened allowing you to alter the size of the output. To do the same
in Windows, set `export = "windows"`, or for Mac, `export = "mac"`.
* To export the graph to a file, export can be "pdf", "png", "jpg", "bmp",
"emf".}

\item{nospp}{number of species to overlay; can be NULL or 0 for none}

\item{plotdim}{The x and y dimensions of the plot.}

\item{clrlist}{colors to use for the species to be overlaid}

\item{ptsize}{A vector of two numbers giving the point size for the
background of all species (first one) and for the individual species
(second one).}

\item{xrange}{Minimum and maximum x and y coordinates of area graphed.}

\item{yrange}{Minimum and maximum x and y coordinates of area graphed.}

\item{wd}{Height and width of the overall graph.
The default work for pdf export or mapping to the screen, and ht and wd are
inches. But if export is png, jpg, emf, height and width are pixels and
need to be 500-1000.}

\item{ht}{Height and width of the overall graph.
The default work for pdf export or mapping to the screen, and ht and wd are
inches. But if export is png, jpg, emf, height and width are pixels and
need to be 500-1000.}

\item{side}{the side in inches of the graph; see map()}

\item{labsize}{Size of axis labels.}

\item{axisdiv}{the division between tick marks; must be a vector of 2, first
for x ticks then for y ticks}

\item{filepath}{The folder to which map will be exported.}

\item{outfile}{Filename for output.}
}
\description{
This creates a map with every individual shown. The area to be mapped can be
chosen, allowing maps of small areas.
}
\details{
On top of the map of every individual, individual species can be overlaid in
different colors. If nospp is set NULL, no species are added. If spnames is
set NULL, then the most abundant species in the plot are chosen, up to the
number nospp.
}
