% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{constant.linear}
\alias{constant.linear}
\title{constant.linear}
\usage{
constant.linear(x, param, xcenter = NULL)
}
\description{
A model which is constant for x<lim, and linear for x>lim. The first parameter is the slope,
second the x value of break point, third the lower limit.
}
