% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert.factor}
\alias{convert.factor}
\title{Converts from factors to character.}
\usage{
convert.factor(r)
}
\arguments{
\item{r}{A dataframe.}
}
\description{
Converts any character fields in a dataframe from factors to character
fields.
}
\section{Reuse good code}{

"The single biggest way to improve both the quality of your code and your
productivity is to reuse good code (...) Many algorithms have already been
invented, tested, discussed in the trade literature, reviewed, and improved".

--Code complete 2, by Steve McConnell.
}

\section{Alternative}{

A more powerful and general alternative is \code{\link[dplyr:mutate_if]{dplyr::mutate_if()}}:\preformatted{lett <- letters[1:3]
x <- tibble::tibble(
  was_fctr = as.factor(lett), 
  was_chr = lett, 
  was_int = as.integer(1:length(lett)),
  was_dbl = as.numeric(1:length(lett)) 
  )

x
#> # A tibble: 3 x 4
#>   was_fctr was_chr was_int was_dbl
#>     <fctr>   <chr>   <int>   <dbl>
#> 1        a       a       1       1
#> 2        b       b       2       2
#> 3        c       c       3       3

dplyr::mutate_if(x, is.factor, as.character)
#> # A tibble: 3 x 4
#>   was_fctr was_chr was_int was_dbl
#>      <chr>   <chr>   <int>   <dbl>
#> 1        a       a       1       1
#> 2        b       b       2       2
#> 3        c       c       3       3

dplyr::mutate_if(x, is.double, as.integer)
#> # A tibble: 3 x 4
#>   was_fctr was_chr was_int was_dbl
#>     <fctr>   <chr>   <int>   <int>
#> 1        a       a       1       1
#> 2        b       b       2       2
#> 3        c       c       3       3
}
}

\seealso{
\code{\link[dplyr:mutate_if]{dplyr::mutate_if()}}.
}
