% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{convert.rowcol}
\alias{convert.rowcol}
\title{Convert an integer to a character, with a single leading zero if th...}
\usage{
convert.rowcol(num)
}
\description{
Convert an integer to a character, with a single leading zero if the integer
is < 10. Does not handle integers >99
}
