% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_utilities.R
\name{CountByGroup}
\alias{CountByGroup}
\title{CountByGroup}
\usage{
CountByGroup(data, groupcol, count_to_data = FALSE, countname = "count",
  groupname = "groupID")
}
\description{
Groups a dataframe by one or more columns (named by groupcol). This does exactly what
COUNT(*) GROUP BY does in SQL.
}
