% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{countEmpty}
\alias{countEmpty}
\title{Counts vector elements that are NA or a string of no length.countEm...}
\usage{
countEmpty(x)
}
\description{
Counts vector elements that are NA or a string of no length.
}
