% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{countpresent}
\alias{countpresent}
\title{Counts vector elements > 0. See countspp as well.countpresent find....}
\usage{
countpresent(x)
}
\description{
Counts vector elements > 0. See countspp as well.
}
