% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_lmerBayes.R
\name{covTocorr}
\alias{covTocorr}
\title{covTocorr}
\usage{
covTocorr(cov)
}
\description{
Convert covariance matrix to correlation matrix. Each element is divided by the square root of the product of the corresponding diagonal terms.
}
