% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create.fulldate}
\alias{create.fulldate}
\title{create.fulldate}
\usage{
create.fulldate(datestr, format = "\%Y-\%m-\%d")
}
\description{
Converts a vector of date character strings in any format to a dataframe with year, month, day, yday (day of the year) and julian.

Submitted datestr cannot be an array.
}
\examples{
\dontrun{

create.fulldate(c('23Oct2010','29Mar1956'),'\%d\%b\%Y')}


}
