% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{create.neighbordata}
\alias{create.neighbordata}
\title{Calculates the mean density in neighboring quadrats for every quadr...}
\usage{
create.neighbordata(abundperquad)
}
\description{
Calculates the mean density in neighboring quadrats for every quadrat, given
a vector of abundances per quadrat. The vector of abundances must be ordered
by quadrat index.
}
