\name{CTFS.manual}
\alias{CTFS.manual}
\concept{overview}
\concept{summary}
\concept{list of functions}

\title{CTFS Manuals : Overview}
\description{
	THIS PAGE, PACKAGE AND MANUALS ARE UNDER DEVELOPMENT
	
	Included in the CTFS package are chapters of a manual that provide a 
	brief introduction to R, instructions on how to manage and use CTFS 
	datasets, and lessons on use of R and CTFS functions with examples 
	from a CTFS dataset.
	}
	
\details{
	The CTFS manual is under development.  The chapters can be found under
	\code{overview} in the CTFS R package help pages.  Comments and
	suggestions are most welcome!  Please contact Pamela Hall,
	phall@alum.mit.edu.
	
	1.  Installing R Enviroment and Loading the CTFS R Package
	CTFS.Chpt1.InstallingR.pdf
	
	2.  R Objects and how to "address" them
	CTFS.Chpt2.AddressingObjects.pdf
	
	3.  How to Read R Help Pages
	CTFS.Chpt3.HelpPages.pdf
	
	4.  File Organization for CTFS datasets and analysis
	CTFS.Chpt4.FileStructure.pdf
	
	5.  Learning how to Write R Functions for CTFS projects
	To be implemented in next version of CTFS package
	
	6.  Reading and Writing CTFS R datasets
	CTFS.Chpt6.ReadWrite.pdf
	
	7.  Lessons on use of useful R Functions
	CTFS.Chpt7.UsefulRfunctions.pdf
	
	8.  Graphing in R for CTFS data
	To be implemented in next version of CTFS package
		
	9.  Documentation (R man pages) of CTFS specific functions
	
		abundance : \code{\link{CTFS.abundance}}\cr
		mortality : \code{\link{CTFS.mortality}}\cr
		recruitment : \code{\link{CTFS.recruitment}}\cr
		growth : \code{\link{CTFS.growth}}\cr
		quadrates : \code{\link{CTFS.quadfunctions}}\cr
		basal area : \code{\link{CTFS.basalarea}}\cr
		density dependence : \code{\link{CTFS.densdepend}}\cr
		biomass : \code{\link{CTFS.biomass}}\cr
		diversity : \code{\link{CTFS.diversity}}\cr
		mapping : \code{\link{CTFS.maps}}\cr
		data manipulation : \code{\link{CTFS.readwrite}}\cr
		data files : \code{\link{CTFS.datafiles}}\cr
}
\author{Pamela Hall, Suzanne Lao, Ellen Connell, Rick Condit and 
Others}
\references{\url{http://www.ctfs.si.edu}}

\seealso{
	\code{\link{CTFS.main}}
	}
}
\keyword{manip}
