% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctfs-package.r
\docType{package}
\name{ctfs}
\alias{ctfs}
\alias{ctfs-package}
\title{A Mimimal Package Containing the Original CTFS R Package (Richard Condit)}
\description{
ctfs is to be programatically imported by other packages
developed by \href{http://www.forestgeo.si.edu/}{ForestGEO}. If you need the
original CTFS R Package \href{conditr@gmail.com}{Richard Condit}) visit
\url{http://ctfs.si.edu/Public/CTFSRPackage/}, were not only functions but
also documentation is available.

\strong{Acknowledgement}

The BCI forest dynamics research project was founded by S.P. Hubbell and R.B.
Foster and is now managed by R. Condit, S. Lao, and R. Perez under the Center
for Tropical Forest Science and the Smithsonian Tropical Research in Panama.
Numerous organizations have provided funding, principally the U.S. National
Science Foundation, and hundreds of field workers have contributed.
}
\references{
Anderson-Teixeira, K. J., Davies, S. J., Bennett, A. C.,
Gonzalez-Akre, E. B., Muller-Landau, H. C., Joseph Wright, S., . . .
Zimmerman, J. 2015. CTFS-ForestGEO: a worldwide network monitoring forests
in an era of global change. Global Change Biology, 21(2), 528-549.
doi:10.1111/gcb.12712 Hubbell, S.P., Condit, R., and Foster, R.B. 2010.
Barro Colorado Forest Census Plot Data.
\href{http://ctfs.si.edu/webatlas/datasets/bci}{URL}.

Condit, R. 1998. Tropical Forest Census Plots. Springer-Verlag and R. G.
Landes Company, Berlin, Germany, and Georgetown, Texas.

Hubbell, S.P., R.B. Foster, S.T. O'Brien, K.E. Harms, R. Condit, B.
Wechsler, S.J. Wright, and S. Loo de Lao. 1999. Light gap disturbances,
recruitment limitation, and tree diversity in a neotropical forest. Science
283: 554-557.
}
\author{
Richard Condit \email{conditr@gmail.com}, Mauro Lepore
\email{maurolepore@gmail.com} (maintainer) and Stuart Davies
\email{daviess@si.edu}.
}
