% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dasymexp}
\alias{dasymexp}
\title{dasymexp}
\usage{
dasymexp(x, center, rate1, rate2, log = FALSE)
}
\description{
Probability distributions for a folded but asymmetrical exponential. When
\code{x >= center}, it's a standard exponential. When \code{x < center}, it's the
mirror image of a different exponential; \code{rate1} refers to the right half,
\code{rate2} to the left. The center is not the median: the section \code{x > center}
has integral \code{rate2 / (rate1 + rate2)}, and the section
\code{x < center rate1 / (rate1 + rate2)}.
}
