% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dbeta.reparam}
\alias{dbeta.reparam}
\title{This reparameterizes the beta distribution as a function of its mea...}
\usage{
dbeta.reparam(x, mu, sd)
}
\description{
This reparameterizes the beta distribution as a function of its mean and
standard deviation. The mean must be between 0 and 1, and sd>0.
}
