% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{defineSDpar}
\alias{defineSDpar}
\title{defineSDpar}
\usage{
defineSDpar(x, nopar)
}
\description{
For default SD parameters, if nothing else works. Choose the midpoint of x for the middle of 3 parameters.
}
