% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{detachfiles}
\alias{detachfiles}
\title{detachs from the searchpath files matching a submitted vector of na...}
\usage{
detachfiles(filename)
}
\description{
detachs from the searchpath files matching a submitted vector of names.
}
