% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dexp.sin}
\alias{dexp.sin}
\title{Four-parameter exponential sin, as a probability distributiondexp.s...}
\usage{
dexp.sin(x, lo, hi, skew, tail)
}
\description{
Four-parameter exponential sin, as a probability distribution
}
