% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dgamma.meansd}
\alias{dgamma.meansd}
\title{Probability distribution of gamma, parameterized with mean and sd.}
\usage{
dgamma.meansd(x, mean, sd, log = FALSE)
}
\arguments{
\item{mean}{Mean.}

\item{sd}{sd.}
}
\description{
Probability distribution of gamma, parameterized with mean and sd instead of
shape and scale.
}
