% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dgamma.scale}
\alias{dgamma.scale}
\title{Like above, but with scale as the first parameter.dgamma.scale dpow...}
\usage{
dgamma.scale(k, x, mu)
}
\description{
Like above, but with scale as the first parameter.
}
