% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{discrete.model}
\alias{discrete.model}
\title{discrete.model}
\usage{
discrete.model(x, param)
}
\description{
A model for a numeric response to a single discrete predictor. The predictor must be a factor, so the total number of levels is known. The response is modeled as a different value for each distinct predictor. The number of parameters must exceed the maximum value of the predictor.
}
