% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageJ.R
\name{distance.to.side}
\alias{distance.to.side}
\title{distance.to.side}
\usage{
distance.to.side(pt, target, side1, side2)
}
\description{
Called from imageJ.to.lxly where for each point, distance to each side is calculated in 3 different ways: perpendicular, parallel to one adjacent side,
and parallel to other adjacent. These 3 distances are averaged, and the discrepancy used as a measure of error, taken as the CV of the three measures.
}
