% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ditch}
\alias{ditch}
\title{Detaches all files at one or more search positions; v can be a vect...}
\usage{
ditch(v)
}
\description{
Detaches all files at one or more search positions; v can be a vector.
}
