% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dmixnorm}
\alias{dmixnorm}
\title{Mixed normal distribution. The parameter f is the probability  of f...}
\usage{
dmixnorm(x, mean1, mean2, sd1, sd2, f, log = FALSE)
}
\description{
Mixed normal distribution. The parameter f is the probability
of following the first, with mean1 and sd1; 1-f is the probability
for the second normal
}
