% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dnormprod0}
\alias{dnormprod0}
\title{Probability density of product of two normal variates, both with me...}
\usage{
dnormprod0(z, sx, sy)
}
\arguments{
\item{z}{A numeric vector.}
}
\description{
Probability density of product of two normal variates, both with mean 0, SD
sx and sy
}
