% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_distributions.R
\name{dnormrev}
\alias{dnormrev}
\title{dnormrev}
\usage{
dnormrev(m, x, s, log = F)
}
\description{
This reverses the order of parameters to dnorm, so that outer can be used
with a vector of x, and two vectors for mean and sd (the latter two equal in
length).
}
