% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{doublenormal}
\alias{doublenormal}
\title{Simulate draws from a double normaldoublenormal dnormprod0  Probabi...}
\usage{
doublenormal(N = 10000, m = 0, sd1 = 1, sd2 = seq(0, 2, by = 0.1))
}
\description{
Simulate draws from a double normal
}
