% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dpois.maxtrunc}
\alias{dpois.maxtrunc}
\title{dpois.maxtrunc}
\usage{
dpois.maxtrunc(x, lambda, maxx, log = FALSE)
}
\description{
A zero-truncated Poisson distribution with a ceiling (combining dpois.max and dpois.trunc).
}
