% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dpois.trunc}
\alias{dpois.trunc}
\title{A zero-truncated Poisson distribution. dpois.trunc dpois.maxtrunc  ...}
\usage{
dpois.trunc(x, lambda, log = FALSE)
}
\description{
A zero-truncated Poisson distribution.
}
