% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dpower}
\alias{dpower}
\title{dpower}
\usage{
dpower(x, beta, log = FALSE)
}
\description{
A probability distribution defined by a power function. There is a dpareto in R, but quite different, with
two parameters. In this, the exponent is the only parameter, and it must be < (-1); x must be >= 0.
}
