% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_geometry.R
\name{drawrectangle}
\alias{drawrectangle}
\title{drawrectangle}
\usage{
drawrectangle(corners, add = TRUE, clr = "black", lwidth = 1)
}
\description{
Draw a rectangle given a matrix or dataframe of 4 x-y coordinates. The columns must have x first then y coordinates.
}
