% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{drp}
\alias{drp}
\title{drp}
\usage{
drp(x)
}
\description{
A version of drop which includes as.matrix. Without it, drop does not serve its purpose. This is necessary in many many situations
where a single row is taken out of a dataframe, but must be passed as a vector.
}
