% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dweibull.3param}
\alias{dweibull.3param}
\title{PDF of three-parameter Weibulldweibull.3param weibull.median.3param...}
\usage{
dweibull.3param(x, x0, shape, scale)
}
\description{
PDF of three-parameter Weibull
}
