% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{ellipse}
\alias{ellipse}
\title{Equation for (half) a canonical ellipse fullellipse  Creates a data...}
\usage{
ellipse(x, center = c(0, 0), radius = c(7, 5), half = "top")
}
\arguments{
\item{x}{Coordinates.}

\item{center}{x-y coordinates of the shape center (vector of 2)}

\item{radius}{Radius (scalar)}
}
\description{
Equation for (half) a canonical
}
