% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{enoughSamplePerBin}
\alias{enoughSamplePerBin}
\title{Test whether the number of elements in a vector x between successiv...}
\usage{
enoughSamplePerBin(x, b, ...)
}
\arguments{
\item{x, b}{Passed to \code{x}, \code{breaks} in \code{\link[base:cut]{base::cut()}}.}

\item{...}{Unused.}
}
\description{
Test whether the number of elements in a vector x between successive breaks
exceeds a minimum. If any bins have too few, it returns FALSE.
}
\seealso{
\code{\link[base:cut]{base::cut()}}.
}
