% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{exp_2par}
\alias{exp_2par}
\title{exp_2par}
\usage{
exp_2par(x, param, asymp)
}
\description{
An exponential distribution with an asymptote.
}
\details{
Name exp.2par clashed with an S3 method, so it was replaced by exp_2par.
}
