% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{explode}
\alias{explode}
\title{Split a single (atomic) character variable into sections, separated...}
\usage{
explode(str, sep = "")
}
\description{
Split a single (atomic) character variable into sections, separated on sep. With the default, sep
, it divides the string into it's individual letters.
}
