% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{expon.model}
\alias{expon.model}
\title{expon.model}
\usage{
expon.model(x, param)
}
\description{
Exponential model, y = a exp(b1\emph{x1 + b2}x2) for any number of predictors x.
Compare to linear.model.
}
