% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{exponential.sin}
\alias{exponential.sin}
\title{Five-parameter exponential sinexponential.sin pexp.sin  CDF of four...}
\usage{
exponential.sin(x, asymp, b, c, d, e)
}
\description{
Five-parameter exponential sin
}
