% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{extract.growthdata}
\alias{extract.growthdata}
\title{Extract growth rates from plot databases and 2 censuses.}
\usage{
extract.growthdata(census1, census2, growcol = "incgr", mingrow = 0.1,
  logit = "x", growthfunc = growth.biomass.indiv, pomcut = 10000,
  rounddown = FALSE, mindbh = 10, dbhunit = "mm", err.limit = 4,
  maxgrow = 75, exclude.stem.change = TRUE, returnfull = FALSE)
}
\arguments{
\item{census1}{The R Analytical Table for a single census, either tree or
stem.}

\item{census2}{The matching R Analytical Table for a later census.}

\item{logit}{Defaults to return log-transformed growth, with negative and
zero growth set to a mingrow, but with \code{logit = TRUE}, growth and dbh are
not log-transformed.}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{dbhunit}{'cm' or 'mm', only used for basal area}
}
\value{
Returns a table with growth, size (ie dbh), and species name.
}
\description{
Extract data for growth rates from plot databases and 2 censuses in CTFS R
format.
}
