% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fill.1dimension}
\alias{fill.1dimension}
\title{fill.1dimension}
\usage{
fill.1dimension(dataarray, class1, fill = 0)
}
\description{
Filling a 1D array resulting from tapply. Same as fill.dimension, but for a vector.
}
