% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fill.dimension}
\alias{fill.dimension}
\title{fill.dimension}
\usage{
fill.dimension(dataarray, class1, class2, fill = 0)
}
\description{
This function fills out an array of 2 dimensions, adding zeroes (or other values) for extra columns
and rows as named in class1 and class2. If a column (or row) is
missing, it will be filled with the value given by fill. It is useful for results of table
or tapply when some elements had no records.
}
