% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find.nonNA}
\alias{find.nonNA}
\title{Returns the first value of a vector x which is not NAfind.nonNA whi...}
\usage{
find.nonNA(x)
}
\description{
Returns the first value of a vector x which is not NA
}
