% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_quadfunc.R
\name{findneighborabund}
\alias{findneighborabund}
\title{findneighborabund}
\usage{
findneighborabund(abundvect, gridsize = 20, plotdim = c(1000, 500))
}
\description{
For every quadrat, finds neighboring quadrats and then returns a vector of abundances in those
neighbors, as well as the number of neighboring quadrats. A subroutine used by create.neighbordata.
}
