% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_distributions.R
\name{fit.beta.normal}
\alias{fit.beta.normal}
\title{fit.beta.normal}
\usage{
fit.beta.normal(x, par1, xmin, par2, xmax)
}
\description{
Finding a normal distribution which most closely fits a given beta distribution.
Parameters for a beta function are submitted, and the best fit mean and SD of a
normal distribution returned.
}
