% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_distributions.R
\name{fit.logistic}
\alias{fit.logistic}
\title{fit.logistic}
\usage{
fit.logistic(x, y, start.param, tech = "Nelder-Mead")
}
\description{
A function to fit a set of data y, observed at the vector x, to a generalized
logistic function, using least squares.
}
