% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_distributions.R
\name{fitnorm}
\alias{fitnorm}
\title{fitnorm}
\usage{
fitnorm(start.param = c(-1, -1, 1), x, y)
}
\description{
Fitting a normal distribution to data
Parameters are a mean and sd of the normal being fitted, a scaling parameter k
and the last SD is for the likelihood of the deviations.

y is vector data to be fitted, at points x
}
