% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund.fit.CTFS.R
\name{fitSeveralAbundModel}
\alias{fitSeveralAbundModel}
\title{Run model.littleR.Gibbs for a series of census databases.}
\usage{
fitSeveralAbundModel(allcns = list(bci.full1, bci.full2, bci.full3),
  sptable = bci.spptable, mindbh, excludespp = NULL,
  useIDlevel = TRUE, abundrange = c(1, 1e+06), start = c(-3, 0.8,
  0.01, -0.5), modeltype = "asympower",
  bad.modelparam = bad.asympower.param, steps = 10000, burn = 1000,
  show = 250, debug = FALSE)
}
\arguments{
\item{allcns}{Full R census, submitted as a list (as many as desired).}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than \code{mindbh} are excluded. If \code{NULL}, all living trees are
included.}

\item{abundrange}{the default includes every species, but this can be set to
a minimum and maximum abundance (first census); species with abundances
outside the range are excluded}

\item{start}{A vector giving the starting set of parameters for the model. It
must be as long as the number of parameters required by the model.}

\item{modeltype}{Functional forms to fit to the distribution; it can be:
\itemize{
\item Gaussian \code{modeltype = "norm"}, with the quotes
\item Asymmetric Gaussian (a different standard deviation on left and right of
the mode) \code{modeltype = "asymnorm"}, with the quotes
\item Laplace (exponential distribution, with mirror image for negative values)
\code{modeltype = "symexp"}, with the quotes
\item Asymmetric Laplace (different rate constant for left and right of the
center) \code{modeltype = "asymexp"}, with the quotes
\item Asymmetric power distribution (different rate constant for left and right
of the center) `modeltype = "asympower", with the quotes.
}}

\item{bad.modelparam}{name of a function which checks the model parameters
for bad values; for modeltype asymexp, must be bad.asymexp.param, for
modeltype asympower, must be bad.asympower.param}

\item{steps}{The number of steps to run the Gibbs sampler.}

\item{burn}{number of steps of sampler to exclude as burn-in}

\item{debug}{Logical. If TRUE, call browser to debug.}
}
\description{
Run \code{\link[=model.littleR.Gibbs]{model.littleR.Gibbs()}} for a series of census databases, for every
successive pair, then the first to the last. Then repeat for 10 times the
initial mindbh.
}
\seealso{
\code{\link[=model.littleR.Gibbs]{model.littleR.Gibbs()}}.
}
