% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fromjulian}
\alias{fromjulian}
\title{Convert calendar dates from julian dates.}
\usage{
fromjulian(j, dateform = "\%m/\%d/\%Y")
}
\arguments{
\item{j}{Julian dates, given as a vector or array.}

\item{dateform}{Format for calendar date, see \code{\link[base:strptime]{base::strptime()}} for format
information.}
}
\value{
A character representation of a date.
}
\description{
Convert calendar dates from julian dates.
}
\examples{
\dontrun{
fromjulian(1000,'\%d\%B\%Y')
}

}
\seealso{
\code{\link[=tojulian]{tojulian()}}
}
