% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{full.xygrid}
\alias{full.xygrid}
\title{Create a complete of points x-y, given the sequence of unique x and...}
\usage{
full.xygrid(x, y)
}
\description{
Create a complete of points x-y, given the sequence of unique x and the sequence of unique y. So if x=y=0:2,
it creates all pairs: 0,0; 0,1; 0,2; 1,0; 1,1; 1,2; etc.
}
