% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{fullcircle}
\alias{fullcircle}
\title{Create a dataframe for a full circle, with x values repeated to get...}
\usage{
fullcircle(x, center = c(0, 0), radius = 5)
}
\description{
Create a dataframe for a full circle, with x values repeated to get top then bottom. NA is inserted
so this can be passed directly to graphing functions.
}
