% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{fullellipse}
\alias{fullellipse}
\title{Creates a dataframe for a full ellipse.}
\usage{
fullellipse(x, center = c(0, 0), radius = c(7, 5))
}
\arguments{
\item{x}{Coordinates.}

\item{center}{x-y coordinates of the shape center (vector of 2)}

\item{radius}{Radius (scalar)}
}
\description{
Creates a dataframe for a full ellipse.
}
