% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{getquadratname}
\alias{getquadratname}
\title{Convert x, y coordinates and plot dimensions into 4-character quadr...}
\usage{
getquadratname(x, y, plotdim)
}
\description{
Convert x, y coordinates and plot dimensions into 4-character quadrat names. If x or y are missing, the quadrat=9999.
}
