% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_quadfunc.R
\name{getsmallerquads}
\alias{getsmallerquads}
\title{getsmallerquads}
\usage{
getsmallerquads(index, gridlarge, gridsmall, plotdim)
}
\description{
Takes a vector of indices for a larger quadrat dimension, as created by gxgy.to.index, and for
each returns a vector of indices of smaller quadrats that would fit completely
within. Both larger and smaller quadrats must be square. Returns a matrix, each row being a 
vector of smaller quadrats inside a single larger quadrat.
}
