% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvetopo.R
\name{getTopoLinks}
\alias{getTopoLinks}
\title{getTopoLinks}
\usage{
getTopoLinks(index, labels, data, from, to, backward = FALSE)
}
\description{
This is solely for use by solve_topo. It finds all points linked via a sighting to a given point.
}
