% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Gibbs.normalmean}
\alias{Gibbs.normalmean}
\title{Gibbs.normalmean}
\usage{
Gibbs.normalmean(y, sigma)
}
\description{
The standard Gibbs sampler for a normal distribution with unknown mean and variance.

y is the vector of observations
sigma is the latest draw of the SD, using sqrt(Gibbs.normalvar)
}
