% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Gibbs.normalvar}
\alias{Gibbs.normalvar}
\title{Gibbs draw for the variance of a normal distribution (http://www.bi...}
\usage{
Gibbs.normalvar(y)
}
\description{
Gibbs draw for the variance of a normal distribution (http://www.biostat.jhsph.edu/~fdominic/teaching/BM/3-4.pdf). If all y are
identical, it returns a small positive number.
}
