% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Gibbs.regslope}
\alias{Gibbs.regslope}
\title{Updates the regression slope (used in regression.Bayes).Gibbs.regsl...}
\usage{
Gibbs.regslope(X, y, var)
}
\description{
Updates the regression slope (used in regression.Bayes).
}
