% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund.fit.CTFS.R
\name{graph.abundmodel}
\alias{graph.abundmodel}
\title{Output histograms of little.r across species, observed and fitted.}
\usage{
graph.abundmodel(fit, datafile = NULL, div = 0.01, tinydiv = 0.001,
  modeltype = "asympower", xrange = NULL, yrange = NULL,
  minabund = 50, conf = 25, returnextreme = TRUE,
  xname = "Rate of population change", yname = "Frequency",
  ltype = "solid", lwidth = 2, modelclr = "green", bartype = FALSE,
  addpts = TRUE, makeleg = TRUE, add = FALSE, newgraph = FALSE,
  ax = TRUE, mortcorr = FALSE, debug = FALSE)
}
\arguments{
\item{fit}{result of model.littleR.Gibbs}

\item{datafile}{optional name of file where the fitted result is saved}

\item{div}{width of bins for histogram of observed rate of population change}

\item{tinydiv}{width of bins used to draw the fitted distribution}

\item{modeltype}{Functional forms to fit to the distribution; it can be:
\itemize{
\item Gaussian \code{modeltype = "norm"}, with the quotes
\item Asymmetric Gaussian (a different standard deviation on left and right of
the mode) \code{modeltype = "asymnorm"}, with the quotes
\item Laplace (exponential distribution, with mirror image for negative values)
\code{modeltype = "symexp"}, with the quotes
\item Asymmetric Laplace (different rate constant for left and right of the
center) \code{modeltype = "asymexp"}, with the quotes
\item Asymmetric power distribution (different rate constant for left and right
of the center) `modeltype = "asympower", with the quotes.
}}

\item{xrange, yrange}{Range of graph's x-axis (or y-axis).}

\item{minabund}{minimum abundance of species to be used in histogram of
observed rates of population change}

\item{conf}{number of alternate fits to graph, as indication of confidence;
if conf = NULL, no confidence lines are added}

\item{returnextreme}{whether to print a list of the fastest increases and
decreases in abundance to the screen}

\item{xname, yname}{axis names}

\item{ltype}{Line type; see \code{ltype} in ?\code{\link[graphics:par]{graphics::par()}}.}

\item{lwidth}{Line width; see \code{width} in ?\code{\link[graphics:par]{graphics::par()}}.}

\item{modelclr}{Line color; see ?\code{\link[graphics:par]{graphics::par()}}.}

\item{bartype}{if TRUE, histogram is bar graph}

\item{addpts}{if TRUE, histogram is a point graph}

\item{makeleg}{whether to add legend}

\item{add}{Logical. Whether to add elements such as points or line to an
existing graph. if TRUE, there must be an appropriate existing graph,
otherwise, a new graph is drawn.}

\item{newgraph}{if FALSE, just graph to the screen and export is ignored; use
TRUE to create a new plot window or export to a file.}

\item{ax}{if FALSE, the axes are not added}

\item{mortcorr}{whether to graph the correlation between mortality and
population change across species}

\item{debug}{Logical. If TRUE, call browser to debug.}

\item{graphit}{xxxdocparam in graph.abundmodel() is truncated: "if set to
false," What follows?}
}
\description{
Output histograms of little.r across species, observed and fitted, using the
result of \code{\link[=model.littleR.Gibbs]{model.littleR.Gibbs()}}. The histogram of black points is all
species, blue points only those starting with \code{N >= minabund}.
}
\section{Arguments details}{

\code{mortcorr} If the argument \code{mortcorr = TRUE}, a graph of mortality rate vs.
population change for every species is also produced. Otherwise, a table of
the species with biggest increases and biggest decreases in abundance is
printed to the screen.
}

\section{Warning}{
 If you use the argument datafile, beware that it uses
\code{attach()}. After use you should remove the attached object from the serach
path. See \emph{Good practice} in \code{?attach()}. This argument may be deprecated
in future versions.
}

\seealso{
?\code{\link[graphics:plot]{graphics::plot()}}, ?\code{\link[graphics:par]{graphics::par()}}.
}
