% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.graph.R
\name{graph.growthmodel}
\alias{graph.growthmodel}
\title{Graph growth rates and model fit. graph.growthmodel overlay.growthb...}
\usage{
graph.growthmodel(spp, fitlist, whichbin = 1, regclr = "green",
  modelclr = "blue", graphdiv = 10, export = pdf,
  outfile = "growth/linearbin.fit.pdf", h = 8, w = 10)
}
\description{
Graph growth rates and model fit.
}
