% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.graph.R
\name{graph.outliers}
\alias{graph.outliers}
\title{Plot outliers overlaid on the model and full data for all species.}
\usage{
graph.outliers(full, trimmed, fit = NULL, allspp = NULL, size = "agb",
  export = NULL, wind = X11)
}
\arguments{
\item{fit}{A fitted model.}
}
\description{
Graph the outliers overlaid on the model and full data for all species in a
model fit result.

The argument export can be set to a window device (X11, win.graph, quartz) or
a graphics option (png, pdf).
}
