% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.graph.R
\name{graph.outliers.spp}
\alias{graph.outliers.spp}
\title{Pass the output of extract.growth with every individual's growth (f...}
\usage{
graph.outliers.spp(full, trimmed, spname = "gustsu", fit = NULL,
  size = "agb", export = NULL, xtitle = "log(agb)", ytitle = "growth")
}
\arguments{
\item{fit}{A fitted model.}
}
\description{
Pass the output of extract.growth with every individual's growth (full), and
another after outliers have been trimmed. This finds the records trimmed and
overlays them on the entire graph. If a model is submitted, then the curves
are graphed too.
}
\examples{
\dontrun{
full = extract.growthdata(
  census1 = bciex::bci12t5mini,
  census2 = bciex::bci12t6mini,
  growcol = 'incgr',
  growthfunc = growth.biomass.indiv,
  logit = 'x',
  rounddown = FALSE,
  mindbh = 100,
  dbhunit = 'mm',
  err.limit = 4000,
  maxgrow = 7500
)
trimmed = extract.growthdata(
  census1 = bciex::bci12t5mini,
  census2 = bciex::bci12t6mini,
  growcol = 'incgr',
  growthfunc = growth.biomass.indiv,
  logit = 'x',
  rounddown = FALSE,
  mindbh = 100,
  dbhunit = 'mm',
  err.limit = 4,
  maxgrow = 75
)
}

}
