% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{graphFilledBand}
\alias{graphFilledBand}
\title{Fill the area between two curves on a graph.}
\usage{
graphFilledBand(x, y1, y2, fillcol = "gray", add = TRUE,
  linecol = "black", ltype = "solid", lwidth = 1)
}
\arguments{
\item{x}{The x axis values.}

\item{y1, y2}{Two sets of y axis values, each of exactly the same length as
\code{x}.}

\item{fillcol}{The color filling the area between the two curves.}

\item{add}{Logical. Whether to add elements such as points or line to an
existing graph. if TRUE, there must be an appropriate existing graph,
otherwise, a new graph is drawn.}

\item{linecol}{Line colour.}

\item{ltype}{Line type; see `ltype` in ?[graphics::par()].}

\item{lwidth}{Line width; see `width` in ?[graphics::par()].}
}
\description{
Fill the area between two curves on a graph. Useful for confidence limits,
for example.
}
\details{
Typical use is to draw a graph first with some central y values, then add a
confidence band by filling the area between upper and lower confidence limits
(designated by variables lower.y and upper.y in the example below). The
central line should then be redrawn over the filled area.
}
\section{Customizing plot lines}{

If \code{add == TRUE}, \code{linecol}, \code{ltype} and \code{lwidth} are ignored.
}

\examples{
\dontrun{
plot(x, y, type = 'l')
graphFilledBand(x, lower.y, upper.y)
lines(x, y)
}

}
\seealso{
?\code{\link[graphics:plot]{graphics::plot()}}, ?\code{\link[graphics:par]{graphics::par()}}.
}
