% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth.eachspp}
\alias{growth.eachspp}
\title{Calculate growth for each species in given dbh categories.}
\usage{
growth.eachspp(census1, census2, classbreak = c(10, 100, 300),
  dbhunit = "mm", mindbh = 10, growthcol = "dbh", err.limit = 4,
  maxgrow = 75, rounddown = FALSE, method = "I", stdev = FALSE)
}
\arguments{
\item{census1}{The R Analytical Table for a single census, either tree or
stem.}

\item{census2}{The matching R Analytical Table for a later census.}

\item{dbhunit}{'cm' or 'mm', only used for basal area}

\item{mindbh}{the minimum diameter above which the counts are done; if NULL,
all (living) are included}

\item{growthcol}{defines how growth is measured, either 'dbh'or
'agb'(agb=biomass)}

\item{err.limit}{A number. Numbers such as 10000 are high and will
return all measures.}

\item{maxgrow}{A number. Numbers such as 10000 are high and will
return all measures.}

\item{rounddown}{If TRUE, all dbh < 55 are rounded down to the nearest
multiple of 5.}

\item{method}{Use 'I' to calculate annual dbh increment: (dbh2 - dbh1)/time,
or 'E' to calculate the relative growth rate (log(dbh2) - log(dbh1))/time.}

\item{stdev}{Logical. Default (FALSE) returns confidence limits, otherwise
returns the SD in growth rate per group.}
}
\description{
Calculates growth for each species in given dbh categories. It creates
the split variables then uses \code{growth()}.
}
\examples{
\dontrun{
growth.result <- growth.eachspp(
  bci12full5, 
  bci12full6, 
  classbreak = c(10, 50, 100, 300, 500)
)
}

}
\seealso{
\code{\link[=growth]{growth()}} and \code{\link[=abundance]{abundance()}}
}
