% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gsp}
\alias{gsp}
\title{gsp}
\usage{
gsp(searchpos, whichobj = 1)
}
\description{
Returns one of the objects at a given search position. This provides a way to write programs to
check multiple sets of data attached at different positions.
}
