% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_quadfunc.R
\name{gxgy.to.quad}
\alias{gxgy.to.quad}
\title{gxgy.to.quad}
\usage{
gxgy.to.quad(gx, gy, gridsize = 20, plotdim = c(1000, 500), digits = 2,
  start = "zero")
}
\description{
Calculate a quadrat name (column number then row number, as a 4-digit character string) from gy-gy. If start is set to zero, quadrats start with 0000, otherwise, 0101.
}
