% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{harmonic.mean}
\alias{harmonic.mean}
\title{Harmonic mean of a vector x.}
\usage{
harmonic.mean(x, add = 0, na.rm = TRUE)
}
\arguments{
\item{add}{xxxdocparam Here, add is different to that common @temp@template
add_plot}
}
\description{
Harmonic mean of a vector x. NAs and nonzero values can be ignored, and a
constant can be added to every x.
}
