% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{hist_compare}
\alias{hist_compare}
\title{hist_compare}
\usage{
hist_compare(x, y, div, breaks = NULL)
}
\description{
Compares two histograms with a Kolmogorov approach.
}
\details{
Name hist.compare clashed with a S3 method, so it was replaced by
hist_compare.
}
