% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund.fit.CTFS.R
\name{hyper.abundGibbs}
\alias{hyper.abundGibbs}
\title{Likelihood function for hyperparameters of abundance model, given t...}
\usage{
hyper.abundGibbs(test, Rsd, spmean, whichtest, type, badparam = NULL,
  returneach = FALSE)
}
\description{
Likelihood function for hyperparameters of abundance model, given the species
values of little.r (latter a vector). Simply calculates the pdf of whatever
modelfunc is requested. For symmetric models, norm and symexp, the third
parameter (second hyperSD) is not used.

Only used as a subroutine of the main modeling function, model.littleR.Gibbs.
}
