% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund.fit.CTFS.R
\name{hyper.mortGibbs}
\alias{hyper.mortGibbs}
\title{Likelihood function for logMu and logSD, given the species means (l...}
\usage{
hyper.mortGibbs(test, MuSD, spmean, whichtest)
}
\description{
Likelihood function for logMu and logSD, given the species means (latter a vector). Simply calculates
log-normal probability of observing the species means given logMu and logSD. This is modernized to use arrangeParam.llike,
and replaces the older mu.mortGibbs and sd.mortGibbs.
}
