% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapresponse.R
\name{imageGraph}
\alias{imageGraph}
\title{Maps a response variable with Rs contour or image functions, or both.}
\usage{
imageGraph(matrixdata, levels = NULL, breaks = c(-33000, 0, 150, 250, 500,
  1000, 2000), xname = "x", yname = "y", xrange = NULL, yrange = NULL,
  axisdiv = c(100, 100), img = TRUE, cntrs = FALSE, filled = TRUE,
  drawgrid = FALSE, addaxes = 1:4, clrs = NULL, cntcolor = "black",
  lwidth = 1, returnline = FALSE, newgraph = FALSE, add = FALSE,
  plotsize = 6, h = 11, w = 11, export = X11, graphfile = NULL)
}
\arguments{
\item{matrixdata}{a list having components x, y, and mat, exactly the output
of coldata.to.imagemat}

\item{levels}{if set, then contours are drawn (or colors assigned) to breaks
are at levels equally-spaced divisions of the response variable; can be
NULL to let breaks control}

\item{breaks}{if set, a vector of breaks for the image colors and the contour
lines; if NULL, levels controls the breaks}

\item{xname, yname}{names for the axes}

\item{xrange, yrange}{if \code{NULL}, all \code{x} (or \code{y}) are included, otherwise, the
graph will only include \code{x} (or \code{y}) inside \code{xrange} or (\code{yrange})}

\item{axisdiv}{the division between tick marks; must be a vector of 2, first
for x ticks then for y ticks}

\item{img}{TRUE or FALSE, whether to draw the color image}

\item{cntrs}{TRUE or FALSE, whether to add contour lines; both img and cntrs
can be TRUE}

\item{filled}{TRUE or FALSE, whether to draw filled contours; if TRUE,
neither image or standard contours are included}

\item{clrs}{colors for the image; if NULL, they are chosen by default,
otherwise, must be 1 + the number of breaks}

\item{returnline}{TRUE or FALSE, whether to return the contour lines; see R's
base function contourLines for details}

\item{newgraph}{if FALSE, just graph to the screen and export is ignored; use
TRUE to create a new plot window or export to a file}

\item{plotsize}{size of graph's x-dimension in inches; ignored if
newgraph=FALSE}

\item{h, w}{height and width of graph window in units which depend on the
export device chosen (some are pixels, some are inches); ignored if
newgraph=FALSE}

\item{export}{if newgraph=TRUE, this defines the graphics device; options
include X11, quartz, win.graph, win.metafile, png, bmp, jpeg, pdf (not
quoted!)}

\item{graphfile}{the file name if export is pdf, png, jpeg, win.metafile,
etc.}
}
\description{
Maps a response variable with R's contour or image functions, or both, using
the output of \code{\link[=coldata.to.imagemat]{coldata.to.imagemat()}}. Also can return contours as vector
data, using R's contourLines. The graph can be sent to the screen or
exported.
}
\examples{
\dontrun{

CTFSplot('bci', 6, 'full')

N = colSums(
  abundanceperquad(bci.full6, plotdim = c(1000, 500), grid = 20)$abund
)
matdata = coldata.to.imagemat(z = N,
  gridsize = 20,
  plotmax = c(1000, 500))
imageGraph(
  matrixdata = matdata,
  breaks = seq(100, 400, by = 25),
  img = TRUE,
  cntrs = FALSE,
  newgraph = FALSE,
  plotsize = 6,
  h = 11,
  w = 11
)
imageGraph(
  matrixdata = matdata,
  breaks = seq(100, 400, by = 25),
  img = TRUE,
  cntrs = TRUE,
  newgraph = FALSE,
  plotsize = 6,
  h = 11,
  w = 11
)
imageGraph(
  matrixdata = matdata,
  breaks = seq(100, 400, by = 25),
  img = TRUE,
  cntrs = FALSE,
  newgraph = TRUE,
  export = win.graph,
  h = 11,
  w = 11,
  plotsize = 8
)}


}
