% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageJ.R
\name{imageJ.to.lxly}
\alias{imageJ.to.lxly}
\title{Convert map pixels from digitzing program to plot coordinates. Uses...}
\usage{
imageJ.to.lxly(textfile = samplemapfile, delim = ",", lowerleft = "P2",
  upperleft = "P1", upperright = "P4", lowerright = "P3",
  gridsize = c(20, 20))
}
\arguments{
\item{textfile}{the complete name of the textfile to convert, including the path}

\item{delim}{the delimiter used to separate the fields in the map files}

\item{lowerleft, }{upperleft, upperright, lowerright the tags in each of the map files with the calibrated corners}

\item{gridsize}{size of each individual map}
}
\description{
Convert map pixels from digitzing program to plot coordinates. Uses 2 functions from geometry.r:
pts.to.interceptslope=function(pt1,pt2)
perpendicular.distance=function(b,m,x,y) return(sqrt((y-m*x-b)^2/(1+m^2)))
intersection.of.lines=function(b1,m1,b2,m2)
}
