% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogChange.R
\name{individual_grow.table}
\alias{individual_grow.table}
\title{Create a table of individual trees and their growth over two census...}
\usage{
individual_grow.table(cnsdata, powertransformation = 0.45, rnd = c(FALSE,
  TRUE, rep(FALSE, 4)), mingrow = 0.1, mindbh = 10, maxdbh = 10000,
  maxerrorSD = 4, maxerrorGrow = 75, center = 1992, debug = FALSE)
}
\arguments{
\item{cnsdata}{A list of census data sets, for example:
\code{list(bci::bci12full1, bci::bci12full2, bci::bci12full3)}.}

\item{rnd}{used to rounddown dbhs for certain intervals. This argument
indicates that dbhs < 50 mm are rounded down to 5-mm, necessary because
saplings at BCI in 1982 and 1985 were measuring in 5-mm increments. The
growth functions in the CTFS R Package handle this.}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{debug}{Logical. If TRUE, call browser to debug.}
}
\value{
A data frame where each row gives data from one individual, including:
\itemize{
\item \code{tag}: the tree’s tag number;
\item \code{gx}, \code{gy}: coordinates;
\item \code{species}: species;
\item \code{dbh1}, \code{dbh2}: tree diameter at census 1 and 2;
\item \code{census}, \code{censusfact}: census interval; 1 means growth between census 1
and 2, and 2 means growth from census 2 to 3, etc. \code{census} and \code{censusfact}
store the same information but they differ in that their values are integers
and factors;
\item \code{time}: mid-point of the interval over which growth was measured, centered
on 1992. This is the format needed by lmer. The reason for centering time is
that it is much better in linear regressions to have the x-axis near zero;
\item \code{LnSize}: log(dbh1);
\item \code{incgr}: untransformed growth increment; growth increment (difference in
dbh) per year;
\item \code{LnGrowth}: log of the growth rate, with negatives and zeroes converted to
0.1 (from the argument mingrowth in the function);
\item \code{CRGrowth}: cube root of growth rate; power transformation of the growth
rate, with negatives maintained negative (exponent = 0.45). Althouth you may
use \code{incgr} or \code{LnGrowth}, \code{CRGrowth} allows cleaner analyses, that need no
assumption about negative growths.
}
}
\description{
Create a table of individual trees and their growth over two censuses, with many species included.

The flag ctr is used to center the time variable, which is the number of
years since 1960 of the interval midpoint. There is a logarithmic
transformation, for which all growth <= 0 is converted to mingrow. There is
also a power transformation, where each growth rate is raised to the power
given by powertransformation.

In the latter, negative growths are transformed to negative, so do not need
to be corrected.
}
