% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{inside.rect}
\alias{inside.rect}
\title{Determines whether any of the 4 corners of one rectangle are within...}
\usage{
inside.rect(rect1, rect2)
}
\description{
Determines whether any of the 4 corners of one rectangle are within a second rectangle. Both rectangles are submitted as c(x0,x1,y0,y1). If just one
of the corners is inside, it returns true. See insideRectange(), which has a similar name but does something different.
}
