% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_geometry.R
\name{insideRectangle}
\alias{insideRectangle}
\title{insideRectangle}
\usage{
insideRectangle(x, y, xrange, yrange)
}
\description{
Checks a vector of coordinates x, y to return which are inside a rectangle. For a much more general function for checking whether
points are inside polygons, use the function inout() in the package splancs.
}
