% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{is.odd}
\alias{is.odd}
\title{A trivial function to test whether numbers (scalar or vector) are o...}
\usage{
is.odd(x)
}
\description{
A trivial function to test whether numbers (scalar or vector) are odd.
}
