% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-utils.R
\name{is_na_row}
\alias{is_na_row}
\title{Detect rows in a data frame full of NA (accross columns).}
\usage{
is_na_row(.data)
}
\arguments{
\item{.data}{A non empty matrix or data frame}
}
\value{
A logical vector of length equal to number of rows in data.
}
\description{
Detect rows in a data frame full of NA (accross columns).
}
\examples{
library(tibble)
df <- tribble(
~x,  ~y,    ~z,
 1, "a", FALSE,
NA,  NA,    NA,
5,  "b",  TRUE,
NA,  NA,    NA,
NA, "c",    NA
)
is_na_row(df)
}
\keyword{internal}
