% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{kurtosis}
\alias{kurtosis}
\title{Sample kurtosis. The biased portion is the population kurtosis; cor...}
\usage{
kurtosis(x)
}
\description{
Sample kurtosis. The biased portion is the population kurtosis; corrected is
for finite sample.
}
