% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{linear.model.ctr}
\alias{linear.model.ctr}
\title{A simple linear model, where the first parameter is intercept, seco...}
\usage{
linear.model.ctr(x, param, xcenter = NULL)
}
\description{
A simple linear model, where the first parameter is intercept, second the slope, and x can be centered on their median.
}
