% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{linear.model}
\alias{linear.model}
\title{linear.model}
\usage{
linear.model(x, param)
}
\description{
A simple linear model, where the first parameter is intercept, remaining parameters are slopes
}
