% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{linear.mortmodel}
\alias{linear.mortmodel}
\title{linear.mortmodel}
\usage{
linear.mortmodel(x, param)
}
\description{
A model for mortality as a function of one or more predictors, with the time
interval for each individual incorporated (as a last predictor).

The log(mortality parameter) is modeled as a linear function of
\code{x[ , -nopred]}. The return value is a survival probability. Nothing prevents
the output from being outside (0,1); that must be handled in the likelihood
function.
}
