% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{llike.GaussModelSD}
\alias{llike.GaussModelSD}
\title{llike.GaussModelSD}
\usage{
llike.GaussModelSD(testparam, allparam, whichtest, x, pred, obs, model, badsd,
  ...)
}
\description{
This is for model.xy. Take the function for the SD, its parameters, and both predicted and observed values of the 
dependent variable (pred,obs)
to generate a likelihood. One of the parameters is passed as testparam, for use with metrop1step. The predicted value
for each observation is included, and not calculated from the predicting function. 
MINIMUM_SD=.0001
MINIMUM_SD should be set in the program calling model.xy, to adjust it appropriately. If MINSD==0, then the sd can
collapse to a miniscule number and drive the likelihood very high, preventing parameter searches from ever escaping the sd.
}
