% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{llike.linearbin.optim}
\alias{llike.linearbin.optim}
\title{This is for optim, a single function taking all parameters at once,...}
\usage{
llike.linearbin.optim(param, x, y, predfunc, nomainpar, badpredpar,
  llikefunc, sdfunc, badsdpar, ...)
}
\description{
This is for optim, a single function taking all parameters at once, including sd, to get single likelihood.
}
