% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogchange_demogChange.R
\name{lmerMortLinear}
\alias{lmerMortLinear}
\title{lmerMortLinear}
\usage{
lmerMortLinear(x, param, ...)
}
\description{
A linear model of an annual mortality parameter [which is -log(annual survival)] as a function of N predictors x, which must be the first N columns of x. 
The parameters are standard slope and intercept for a linear model. There must be one additional column in x for the time interval t. The linear model predicts annual log(survival).
Return value is predicted survival rate (probability) over an interval of t years. Nothing prevents the output from being outside (0,1); that must be handled in the likelihood function.
}
