% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_utilitiesCTFS.R
\name{load.species}
\alias{load.species}
\title{load.species}
\usage{
load.species(spname, sppdatafile)
}
\description{
A function for extracting a single species'dataframe from the large spp dataset
(list of dataframes, one per species). The split data file must come
as a name, that is in quote marks.
}
