% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{log.model}
\alias{log.model}
\title{log.model}
\usage{
\method{log}{model}(x, param)
}
\description{
Logarithmic model, y = a + b1 log(x1) + b2 log(x2) for any number of predictors x. Compare to linear.model. All x should be positive.
}
