% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{log_model}
\alias{log_model}
\title{log_model}
\usage{
log_model(x, param)
}
\description{
Logarithmic model, y = a + b1 log(x1) + b2 log(x2) for any number of
predictors x. Compare to linear.model. All x should be positive.
}
\details{
Name log.model clashed with an S3 method, so it was replaced by log_model.
}
