% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logical.grep}
\alias{logical.grep}
\title{A form of grep returning logical instead of indices (numbers).logic...}
\usage{
logical.grep(needle, haystack)
}
\description{
A form of grep returning logical instead of indices (numbers).
}
