% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_distributions.R
\name{logistic.inter}
\alias{logistic.inter}
\title{logistic.inter}
\usage{
logistic.inter(x, param, log = FALSE, ...)
}
\description{
Logistic function with intercept parameterization (ie, first parameter is y when all x=0). The input x are all independent variables, in a matrix
with each column one of the variables. The number of rows is the number of datapoints. Just one inter, which is the value
at all x=0, and passed as param[1]. Slope parameters follow, one per column of x. 
This is identical to standard
}
