% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_distributions.R
\name{logistic.power}
\alias{logistic.power}
\title{logistic.power}
\usage{
logistic.power(x, param, log = FALSE, ...)
}
\description{
This is the Gaussian logistic function, where logit is a second-order polynomial of x; with asymptote and basement allowed. 
There must be 1+2*nopredictors parameters; the asympotote and basement are only implemented
if extra parameters are passed.
}
