% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{logistic.power_simple}
\alias{logistic.power_simple}
\title{This is a mixture of logistic and logistic-standard models. The pre...}
\usage{
logistic.power_simple(x, param, log = FALSE, N = 1, ...)
}
\arguments{
\item{...}{Unused.}
}
\description{
This is a mixture of logistic and logistic-standard models. The predictors n
get a power model, the remaining a simple model. So if nopredictors==8, and
n=c(1,7), then the first and seventh predictors use a power model, while the
rest a simple model.

There must be 1+length(n)+nopredictors parameters, plus additional 1 or 2 for
asymptote and basement.
}
