% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_distributions.R
\name{logistic.standard}
\alias{logistic.standard}
\title{logistic.standard}
\usage{
logistic.standard(x, param, log = FALSE, ...)
}
\description{
This is standard logistic function, but with asymptote and basement allowed. The latter are only implemented
if extra parameters are passed. Moved from calc.surviv.r on 25 July 2010 to provide the standard logistic.
}
