% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_distributions.R
\name{logistic.sum.squares}
\alias{logistic.sum.squares}
\title{logistic.sum.squares}
\usage{
logistic.sum.squares(param, x, obs)
}
\description{
Sets a prediction based on a generalized logistic, then returns the sum
of squared deviations
}
