% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{logisticmodel.bin}
\alias{logisticmodel.bin}
\title{logisticmodel.bin}
\usage{
logisticmodel.bin(x, param, ...)
}
\description{
Multiple bin model predicting y as a function of x, where each segment is modeled as a standard logistic.

A model with B bins has B-1 parameters for breaks points, B parameters as slopes, and one intercept (y at x=0).

Within each bin, y is a linear function of x.

Predictor can centered at medv.

This function accepts a set of parameters, submits to linearmodel.bin, then returns invlogit.
}
