% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{majoraxisreg.no.int}
\alias{majoraxisreg.no.int}
\title{majoraxisreg.no.int}
\usage{
majoraxisreg.no.int(x, y)
}
\description{
Major axis regression with no intercept. Only a slope
is returned. Below is the same for standard regression.
}
