% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_map.R
\name{map1species}
\alias{map1species}
\title{map1species}
\usage{
map1species(sppdata, plotdim = c(1000, 500), xrange = c(0, plotdim[1]),
  yrange = c(0, plotdim[2]), color = "black", bgcolor = "white",
  axspos = 0.5, xaxis = TRUE, yaxis = TRUE, symbol = 16,
  size = rep(-1, 3), cutoff = c(10, 100, 300, 3000), axisdiv = 100,
  labsize = 1.75, plotside = 4, add = FALSE)
}
\description{
This does the mapping for a single species, called from map() but also useful on its own.
With add=TRUE, points are added to an existing map. All other parameters are described with map().
}
