% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_utilities.R
\name{merge.data}
\alias{merge.data}
\title{merge.data}
\usage{
\method{merge}{data}(listdata, showOutput = NULL)
}
\description{
Combine many dataframes as elements of a list into a single large dataframe. Each individual dataframe must
have exactly the same columns. This is exactly the opposite operation as split.data.
}
