% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge_data}
\alias{merge_data}
\title{merge_data}
\usage{
merge_data(listdata, showOutput = NULL)
}
\description{
Combine many dataframes as elements of a list into a single large dataframe.
Each individual dataframe must have exactly the same columns. This is exactly
the opposite operation as split_data.
}
\details{
Name merge.data clashed with an S3 method, so it was replaced by merge_data.
}
