% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{metrop1step.discrete}
\alias{metrop1step.discrete}
\title{metrop1step.discrete}
\usage{
metrop1step.discrete(func, start.param, alter.param, ...)
}
\description{
A version for metrop1step where the alternative values are character states with no numeric meaning.
A random draw must be taken from all possible states, each with equal probability. There
is no step-size thus no adjustment.
}
