% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{minum.normal}
\alias{minum.normal}
\title{The likelihood function for use by fitnorm.minum.normal fitnorm  Fi...}
\usage{
minum.normal(param, x, y)
}
\description{
The likelihood function for use by fitnorm.
}
