% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{minum.perpdist}
\alias{minum.perpdist}
\title{The sum of squares used by majoraxisreg.minum.perpdist majoraxisreg...}
\usage{
minum.perpdist(param, x, y)
}
\description{
The sum of squares used by majoraxisreg.
}
