% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel.R
\name{mistaken_s3}
\alias{mistaken_s3}
\title{List functions mistakenly exported as S3 methods in NAMESPACE}
\usage{
mistaken_s3()
}
\value{
A tibble listing functions mistakenly exported as S3 methods
}
\description{
List functions mistakenly exported as S3 methods in NAMESPACE
}
