% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_monthnames.R
\docType{data}
\name{MONTHNAMES}
\alias{MONTHNAMES}
\title{Object assumed to be crucial. In `utilities.R, the original documen...}
\format{An object of class \code{character} of length 12.}
\usage{
MONTHNAMES
}
\description{
Object assumed to be crucial. In `utilities.R, the original documentation of
the CTFSRPackage states:
}
\details{
\preformatted{MONTHNAMES = c( 
'Jan','Feb','Mar','Apr','May','Jun','Jul','Aug','Sep','Oct','Nov','Dec' ) 
}
}
\keyword{datasets}
