% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mvrnormRC}
\alias{mvrnormRC}
\title{From a variance-covariance matrix, output normal variates with means 0.}
\usage{
mvrnormRC(N, Sigma)
}
\arguments{
\item{N}{The number to draw.}

\item{Sigma}{Must be square.}
}
\description{
Function that takes a variance-covariance matrix and produces normal variates
following it, but with means 0.
}
\details{
The R function mvrnorm does this too; this was a test of the algorithm from
Tommaso Zillio.
}
\seealso{
\code{\link[MASS:mvrnorm]{MASS::mvrnorm()}}.
}
