% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborDensityFun.R
\name{NDcount}
\alias{NDcount}
\title{Count of neighbors within radius r with edge effect correction.}
\usage{
NDcount(censdata, r = 20, plotdim = c(1000, 500))
}
\arguments{
\item{censdata}{A table of plot census data.}

\item{r}{Radius.}

\item{plotdim}{The x and y dimensions of the plot.}
}
\value{
A single vector with neighbor counts within radius r for each
individual in censdata.
}
\description{
Calculates the count of neighbors within radius \code{r} with edge effect
correction. Is a quick version of \code{\link[=NeighborDensities]{NeighborDensities()}} with limited
capabilities.
}
\details{
\code{censdata} should have the coordinates \code{gx}, \code{gy}.
}
\examples{
\dontrun{
# Count all the consp neighbors of one species 
one.sp = subset(bciex::bci12t7mini, sp == "ingasa" & status == "A")
neighbor.counts <- NDcount(one.sp)
}

}
\author{
Tania Brenes
}
