% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborDensityFun.R
\name{NeighborDensities}
\alias{NeighborDensities}
\title{Total number or basal area of conspecific and heterospecific neighbors.}
\usage{
NeighborDensities(censdata, censdata2 = NULL, r = 20, plotdim = c(1000,
  500), mindbh = 10, type = "count", include = c("A"))
}
\arguments{
\item{censdata}{A table of plot census data.}

\item{censdata2}{(NULL) an optional dataset for focal individuals, for
example a subset of only one species or a list of seedling coordinates that
are not part of census data. It must contain coordinates (gx,gy), species
id (sp) and tag. It will count the neighbors in censdata but only for these
focal coordinates.}

\item{r}{Radius.}

\item{plotdim}{The x and y dimensions of the plot.}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{type}{(count) calculates count of stems ('count') or sum of basal areas
('basal'). Note that for the sum of basal areas is better to use the plot
stem table, otherwise only one stem per tree will be included .}

\item{include}{(c("A")) a vector of status to include in the neighbors, by
default it counts only alive individuals, for all status use
c("A","P","D","M").}
}
\value{
A data.frame with two columns listing the density of consp and hetsp, and
rows equal to the size of censdata if \code{censdata2} is absent, or \code{censdata2}
if it is present.
}
\description{
Calculates the total number or basal area of conspecific and
heterospecific neighbors at a radius r with edge effect correction. Most
arguments aim to describe the type of neighbors to count.
\code{\link[=NeighborDensities]{NeighborDensities()}} counts neighbors for each of the individuals in
censdata.
}
\details{
\itemize{
\item Arguments \code{mindbh} and \code{r} refer to neighbours
\item \code{censdata} should have the coordinates gx, gy, species id \code{sp}, \code{status},
\code{tag} and \code{dbh}.
}

Dependencies: CalcRingArea and circlearea from the CTFS package, and splancs
package.
}
\examples{
\dontrun{
# sum consp and hetsp neighbors for all stems in the plot  
neighbor.counts <- NeighborDensities(bci::bci12full7, r=20, type='count') 

# sum consp and hetsp neighbors for only one species: 
one.sp = subset(bci::bci12full7, sp=="quaras") 
neighbor.counts <- NeighborDensities(bci::bci12full7, one.sp, type='count')
}

}
\author{
Tania Brenes
}
