% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nhd}
\alias{nhd}
\title{A version of head with only 6 columns shown.nhd TextToRdata  Reads ...}
\usage{
nhd(d, w = 1:6, h = 8)
}
\arguments{
\item{d}{An object (see x in \code{\link[=head]{head()}}).}

\item{w}{Sequence giving the range of columns to print}

\item{h}{Number giving the Number of rows to print.}
}
\description{
A version of head with only 6 columns shown.
}
