% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{normalproduct}
\alias{normalproduct}
\title{Product of 2 normal distributions, the first at x and the second at lag-x.}
\usage{
normalproduct(x, lag, mean1, sd1, meanint, meanslope, CV)
}
\arguments{
\item{x}{A normal distribution.}

\item{mean1, sd1, }{The mean and sd of the first distribution.}

\item{meanslope}{The slope}

\item{CV}{The coefficient of variation.}
}
\description{
A function which returns the product of 2 normal distributions, the first
at x (a vector), the second at lag-x (lag is a scalar). The mean and SD
of the second normal are linear functions of x, with meanint being the
intercept, meanslope the slope, and CV the coefficient of variation.

A convolution!
}
