% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{order.by.rowcol}
\alias{order.by.rowcol}
\title{Reorder rows and columns of a matrix so they are sorted as if colum...}
\usage{
order.by.rowcol(x)
}
\description{
Reorder rows and columns of a matrix so they are sorted as if column and row names are numeric. This allows labels from cut to
be ordered numerically. Otherwise, sort order is 1, 10, 11, 12, ... 18, 19, 2, 20, 21, etc.
}
