% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.graph.R
\name{overlay.growthbinmodel}
\alias{overlay.growthbinmodel}
\title{Show model fits for 1, 2, 3, and 4 bins on each species.}
\usage{
overlay.growthbinmodel(fit, bins = 1:4, regclr = "green",
  modelclr = "blue", graphdiv = 15, add = FALSE, newgraph = TRUE,
  export = pdf, outfile = "growth/linearbin.overlay.pdf", h = 8, w = 10)
}
\arguments{
\item{fit}{A fitted model.}

\item{add}{Logical. Whether to add elements such as points or line to an
existing graph. if TRUE, there must be an appropriate existing graph,
otherwise, a new graph is drawn.}

\item{outfile}{Filename for output.}

\item{h, w}{A number giving the height and width.}
}
\description{
Show model fits for 1, 2, 3, and 4 bins on each species.
}
