% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{parallel.line}
\alias{parallel.line}
\title{Slope and intercept of a line parallel to another known line.}
\usage{
parallel.line(b, m, x, y)
}
\arguments{
\item{b}{Intercept.}

\item{m}{Slope.}
}
\description{
Finds the slope and intercept of the line parallel to a line whose slope and
intercept are given, through the points \code{x, y}. Note that the intercept is
not needed. For any \code{m} that are infinite, the intercept is \code{x}.
}
