% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-utils.R
\name{paste_path}
\alias{paste_path}
\title{Paste the path with the content of a directory.}
\usage{
paste_path(path_to_dir)
}
\arguments{
\item{path_to_dir}{A string such as "R" or "man-roxygen".}
}
\value{
The string "path_to_dir/directory_content.extension".
}
\description{
Paste the path with the content of a directory.
}
\examples{
paste_path("R")
}
\keyword{internal}
