% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_block.analysis.R
\name{plot.wavelet}
\alias{plot.wavelet}
\title{plot.wavelet}
\usage{
\method{plot}{wavelet}(x)
}
\description{
Function to plot the wavelet variance from the output of the wavelet.allsp.
}
\author{
Tania Brenes  

 x (): output for wavelet.allsp;
}
