% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.analysis.R
\name{plot_wavelet}
\alias{plot_wavelet}
\title{plot_wavelet}
\usage{
plot_wavelet(x)
}
\arguments{
\item{x}{Output for wavelet.allsp}
}
\description{
Function to plot the wavelet variance from the output of the wavelet.allsp.
}
\details{
Name plot.wavelet clashed with an S3 method, so it was replaced by
plot_wavelet.
}
\author{
Tania Brenes
}
