% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{png.allplot}
\alias{png.allplot}
\title{Export species maps as png.}
\usage{
png.allplot(splitdata = bci.split6, spplist = plotspp$bci,
  elev = ctfs.elev$bci$mat, plotdim = c(1000, 500), topoint = 2,
  plotside = 9, h = 850, w = 1100, ptsizes = c(0.3, 0.45, 0.65),
  dbhcut = c(10, 100, 300, 3000), dead = FALSE, export = TRUE,
  path = "/home/condit/data/maps/bci/")
}
\arguments{
\item{spplist}{A table with species codes and Latin names can be submitted as
well, so that the full genus-species is added to plot. This must have
species codes as row names. It should be the CTFS R format species table
(eg, bciex::bci_species). It can be set to NULL if not available, then only the
species code (as submitted) appears on the map.}

\item{plotdim}{The x and y dimensions of the plot.}

\item{topoint}{Interval for topolines, if elevdata are submitted.}

\item{plotside}{The vertical dimension (inches) of the map.}

\item{h, w}{A number giving the height and width.}

\item{export}{\code{export = "no"}, does nothing, so the next graph uses the
default R device. With this option, the size of the graph cannot have been
set larger than the default size, which is usually 7 inches. \code{export = "unix"}, a new graphics window in unix, size \code{h} and width \code{w}, will be
opened allowing you to alter the size of the output. \code{export = "windows"},
or \code{export = "mac"} are similar to \code{export = "unix"} but for windows and
mac. \code{export = format} exports the graph to a specific file format: "pdf",
"png", "jpg", "bmp", "emf".}
}
\description{
Export species maps as png. There will be one for every species chosen. See
pdf.allplot.
}
\seealso{
\code{\link[=map]{map()}}.
}
