% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{polar.to.cartesian}
\alias{polar.to.cartesian}
\title{polar.to.cartesian}
\usage{
polar.to.cartesian(r, theta)
}
\arguments{
\item{r}{= input vector of radii (distance from origin)}

\item{theta}{= input vector of angle from horizontal (radians), same length as r}
}
\description{
Convert polar coordinates to Cartesian. Returns a dataframe of two columns named x and y.
}
\examples{
\dontrun{
polar.to.cartesian(2,pi/3)}




}
