% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.CTFSdb.R
\name{predht.asym}
\alias{predht.asym}
\title{An allometric model predicting an asymptote at large size, used in ...}
\usage{
predht.asym(dbh, param)
}
\description{
An allometric model predicting an asymptote at large size, used in estimating tree height as a function of dbh.

The model uses 3 parameters, submitted as argument param. The matrix form of param allows a different set of parameters to
be submitted for every species. The default parameters given in the function Chave.AGB assume dbh is in cm, as do all the biomass
allometry functions.

dbh: Vector of dbh
param: Either a vector of length 3, or a matrix of 3 columns; if the latter, there must be one row for each dbh
}
\examples{
\dontrun{
htparam=c(41.7,.057,.748)

d=c(1,2,5,10,20,50)

ht=predht.asym(dbh=d,param=htparam)}


}
