% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-utils.R
\name{print_all}
\alias{print_all}
\title{Print all rows of a tibble or tbl_df.}
\usage{
print_all(x)
}
\arguments{
\item{x}{A tibble or tbl_df.}
}
\value{
All rows of a tibble.
}
\description{
Print all rows of a tibble or tbl_df.
}
\examples{
x <- tibble::tibble(long = 1:30)
x  # prints 10 rows
print_all(x)  # prints all rows
}
