% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{psymexp}
\alias{psymexp}
\title{The CDF for the symmetric exponential.psymexp rsymexp  Drawing a ra...}
\usage{
psymexp(x, center, rate)
}
\arguments{
\item{center}{center Center of the distribution.}
}
\description{
The CDF for the symmetric exponential.
}
