% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pweibull.3param}
\alias{pweibull.3param}
\title{pweibull.3param}
\usage{
pweibull.3param(x, x0, shape, scale)
}
\description{
CDF of three-parameter Weibull
(http://www.itl.nist.gov/div898/handbook/apr/section1/apr162.htm)
}
