% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qasymexp}
\alias{qasymexp}
\title{qasymexp}
\usage{
qasymexp(y, rate1, rate2, c)
}
\description{
Quantiles of dasymexp 

y is the vector of desired quantiles; c is the center parameter; rate1 is the rate for the right half, and rate2 the left.
}
