% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qasympower}
\alias{qasympower}
\title{Quantiles from the bilateral power distribution, dasympower. By R. ...}
\usage{
qasympower(y, rate1, rate2, c)
}
\description{
Quantiles from the bilateral power distribution, dasympower. By R. Chisholm.
}
