% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{randomElement}
\alias{randomElement}
\title{Return a random element from a vector randomElement countUnique  Co...}
\usage{
randomElement(data)
}
\arguments{
\item{data}{A vector.}
}
\description{
Return a random element from a vector.
}
