% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcalpha.R
\name{rarefy.diversity}
\alias{rarefy.diversity}
\title{takes a full plot dataframe and calculates species number in a rand...}
\usage{
rarefy.diversity(fulldata, s)
}
\description{
takes a full plot dataframe and calculates species number in a
random subsample of s individuals. All it requires is a column
called sp which has species designation for each individual.
}
