% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.analysis.R
\name{rasterize}
\alias{rasterize}
\title{Calculate count, basal area or agb per quadrat.}
\usage{
rasterize(x, y, z = NULL, gridsize = 20, plotdim = c(1000, 500),
  FUN = sum, graph = FALSE)
}
\arguments{
\item{x, y}{Coordinates for the point pattern.}

\item{z}{Marks for the marked point process, can be basal area, agb, etc.
Needed only for \code{type = "marked"}.}

\item{gridsize}{Side of the square quadrate.}

\item{plotdim}{The x and y dimensions of the plot.}

\item{FUN}{Function to apply to the point pattern when \code{z} is provided. By
defult it sums the values of z per quadrat.}

\item{graph}{Logical. If \code{TRUE} plots the heat map of raster data.}

\item{type}{Type of rasterization. \code{"point"} runs a simple point pattern
counting the number of individuals per quadrat; \code{"marked"} does a marked
point pattern, where a function (\code{FUN}) is applied to the variable \code{z} in
each quadrat, for example a sum of basal areas.}
}
\value{
A list containing a matrix with the raster data.
}
\description{
Function to calculate the count, basal area or agb per quadrat by selecting
quadrats of variable sizes.
}
\examples{
\dontrun{
onesp = subset(bciex::bci12t1mini, sp == "rinosy")  

# plots the density of the sp in the plot  
rast1 = rasterize(
  x = onesp$gx,
  y = onesp$gy,
  gridsize = 5,
  plotdim = c(1000, 500),
  type = 'point',
  graph = TRUE
)
}

}
\author{
Matteo Detto and Tania Brenes
}
