% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvetopo.R
\name{rearrangeSurveyData}
\alias{rearrangeSurveyData}
\title{rearrangeSurveyData}
\usage{
rearrangeSurveyData(inputtable)
}
\arguments{
\item{all}{points found in the input table, with an integer designation assigned to each. The designation is called pt.}

\item{the}{second table matches the input table,
but instead of x-y coordinates for the two points, only columns pt1 and pt2 are included to
indicate the two points between which a sighting was taken.}
}
\description{
Takes a table of survey sightings with columns of x and y locations of two points, and converts it to the format
required by solve.topo. The input table must have columns x1, y1, x2, and y2. The return value is a list consisting of two dataframes:
}
