% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{regress.plot}
\alias{regress.plot}
\title{Regression coefficients, probabilities and graph.}
\usage{
regress.plot(x, y, title = "", graphit = T, add = F, pts = T,
  clr = "blue", ptnames = NULL, xrange = NULL, yrange = NULL,
  xtitle = NULL, ytitle = NULL)
}
\arguments{
\item{graphit}{Logical; if \code{TRUE} adds plot output.}

\item{clr}{A string indicating colour. See argument col in \code{\link[graphics:par]{graphics::par()}}.}
}
\description{
Performs regression in convenient way and returns coefficients and
probabilities in a single vector, and plots a graph.
}
