% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{regression.Bayes}
\alias{regression.Bayes}
\title{regression.Bayes}
\usage{
regression.Bayes(x, y, start.param = NULL, steps = 1100,
  whichuse = 101:1100, showstep = 50)
}
\description{
Regression using the Gibbs sampler, with just one x variable.
}
