% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_statistics.R
\name{regslope.noint}
\alias{regslope.noint}
\title{regslope.noint}
\usage{
regslope.noint(y, x)
}
\description{
Returns slope of regression with no intercept as single scalar (for use with apply).
}
