% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerBayes.R
\name{resetParam}
\alias{resetParam}
\title{resetParam}
\usage{
resetParam(param, rand, savestep, pars)
}
\description{
This starts a new 3D parameter array whose first element is the last element of the current one. The rest of the new one is empty, to hold
the next set of savestep steps from the Gibbs sampler. This only happens after the current one has been saved to a text file.
}
