% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_distributions.R
\name{rgamma.meansd}
\alias{rgamma.meansd}
\title{rgamma.meansd}
\usage{
rgamma.meansd(n, mean, sd)
}
\description{
Random draws of gamma, parameterized with mean and sd instead of shape and scale.
}
