% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{right}
\alias{right}
\title{Returns the rightmost n characters of a stringright rightbut  Retur...}
\usage{
right(s, n, bytetype = "bytes")
}
\description{
Returns the rightmost n characters of a string
}
