% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RipUvK.R
\name{RipUvK}
\alias{RipUvK}
\title{RipUvK}
\usage{
RipUvK(splitdata, plotdim = c(1000, 500), rseq = c(5, 10, 20, 30, 40, 50),
  mindbh = 10, xcol = "gx", ycol = "gy", debug = FALSE, show = FALSE)
}
\arguments{
\item{splitdata}{A complete CTFS plot dataset comprised of a list, each element having a dataframe for one species. There must be columns for x and y coordinates; the names of those two columns are passed as arguments (default is the CTFS standard gx, gy).}

\item{plotdim}{The x and y dimensions of the plot.}

\item{rseq}{The distances defining intervals in which Ripley's statistic is calculated.}

\item{mindbh}{the minimum dbh to include in results.}

\item{xcol}{and ycol the names of the columns holding x and y coordinates.}

\item{debug}{TRUE or FALSE, calls browser during execution to pause program; only set TRUE to debug.}

\item{show}{creates a graph to show one omega value for each species, as a way to track progress.}
}
\examples{
\dontrun{
# If a split database has been created by split.data and stored, CTFSplot will load it.
CTFSplot("bci",6,type="split")
# Otherwise start with a standard R Analytical Table (could be stem also):
CTFSplot("bci",6,type="full")
bci.split6=split.data(bci.full6)
# A quick test, run on only the first 10 species:
rip=RipUvK(splitdata=bci.split6[1:10],plotdim=c(1000,500),rseq=c(10,20,30,40,50,60),mindbh=10,xcol="gx",ycol="gy")
# All the species (takes several minutes):
rip=RipUvK(splitdata=bci.split6,plotdim=c(1000,500),rseq=c(10,20,30,40,50,60),mindbh=10,xcol="gx",ycol="gy")
str(rip$K[[1]])
plot(rip$midpts,rip$omega[2,],ylim=c(0,10))


# calculate K and the (number of conspecific) individuals for each tree within 
# distances of rseq
}



}
