% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rmixnorm}
\alias{rmixnorm}
\title{Random draw on the mixed normal distribution.rmixnorm minum.mixnorm...}
\usage{
rmixnorm(n, mean1, mean2, sd1, sd2, f)
}
\arguments{
\item{n}{Passed to argument \code{n} in \code{\link[stats:runif]{stats::runif()}}.}

\item{mean1, sd1, }{The mean and sd of the first distribution.}

\item{x}{A distribution.}
}
\description{
Random draw on the mixed normal distribution.
}
\seealso{
\code{\link[stats:runif]{stats::runif()}}
}
