% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiesCTFS.R
\name{rndown5}
\alias{rndown5}
\title{Rounds a numeric value to the next lowest multiple of 5.rndown5 cou...}
\usage{
rndown5(s)
}
\arguments{
\item{s}{Passed to \code{x} in \code{\link[base:floor]{base::floor()}}.}
}
\description{
Rounds a numeric value to the next lowest multiple of 5.
}
\seealso{
\code{\link[base:floor]{base::floor()}}.
}
