% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rpois.max}
\alias{rpois.max}
\title{Random draws on dpois.maxrpois.max rpois.trunc  Random draws on dpo...}
\usage{
rpois.max(N, lambda, maxx)
}
\arguments{
\item{N}{Passed to n in \code{\link[stats:rpois]{stats::rpois()}}.}

\item{lambda}{Passed to lambda in \code{\link[stats:rpois]{stats::rpois()}}.}
}
\description{
Random draws on dpois.max
}
\seealso{
\code{\link[stats:rpois]{stats::rpois()}}.
}
