% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{run.growthbin.manyspp}
\alias{run.growthbin.manyspp}
\title{run.growthbin.manyspp}
\usage{
run.growthbin.manyspp(growthdata, size = "dbh", spp = spp20,
  minabund300 = 15, minTotal = 40, dbhunit = "mm",
  sdmodel = linear.model.ctr, startpar = c(0.03, 0.005),
  startsdpar = c(0.04, 0), badsdfunc = NULL, binoption = 1:4,
  noreps = 5000, noburn = 2500, noshow = 500, outputname = "linear.fit",
  path = "", ...)
}
\description{
Run the model to fit growth rate in bins for many species, 1-4 bins. It takes a list of species, extracts
growth rates for each, one at a time, from the table of growth rates, then calls run.growthfit.bin to fit the model for
the 4 bin options. 

Sample species vector from BCI:
}
\examples{
\dontrun{
spp20=c('tri2tu','alsebl','tet2pa','tachve','beilpe','pri2co','quaras','ocotwh','hirttr','gar2in','protpa','protte',
       'eugeoe','virose','guargu','maquco','jac1co','cecrin','cordbi','micoar') 

Creating the complete table of biomass growth for all individuals in a plot: 
agb.growth=extract.growthdata(bci.full5,bci.full6,growthfunc=growth.biomass.indiv,logit='x',
                             rounddown = FALSE,mindbh = 100,dbhunit = 'mm',err.limit = 4,maxgrow = 75) 

Creating a vector of all species names in the agb.growth table. 
allspecies=sort(unique(agb.growth$sp)) 

Fitting the model for all species, 1-4 bins: 
fit=run.growthbin.manyspp(growthdata=agb.growth,size='agb',spp=allspecies,minabund300=15,minTotal=40,startpar=c(.03,.005),startsdpar=c(.04,0))}


}
