% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{se.kurtosis}
\alias{se.kurtosis}
\title{Standard error of kurtosis. Depends only on sample size. se.kurtosi...}
\usage{
se.kurtosis(x)
}
\description{
Standard error of kurtosis. Depends only on sample size.
}
