% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{se.skewness}
\alias{se.skewness}
\alias{se.skewness}
\title{Standard error of skewness. se.skewness kurtosis   Sample kurtosis....}
\usage{
se.skewness(x)

se.skewness(x)
}
\description{
Standard error of skewness.

A major axis regression with parameters fitted by optim. The regression
is the line which minimizes perpendicular distance summed over all points
(and squared).
}
