% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography_imageJ.R
\name{SectionCorrection}
\alias{SectionCorrection}
\title{SectionCorrection}
\usage{
SectionCorrection(pts, subquad, gridsize, subquadsuffix)
}
\arguments{
\item{pts}{the coordinates to be corrected}

\item{subquad}{the subquadrat to be corrected}

\item{gridsize}{size of each individual map}

\item{subquadratsuffix}{the suffixes used to indicate what section of the map the subquadrat refers to. They should be named clockwise from the lower left.}
}
\description{
This function is called by fullplot.imageJ in the case that the maps are subquadrats to correct the coordinates and make them from 0 to 20 depending on
what section of the map the subquadrat belongs to.
}
