% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{segmentPt}
\alias{segmentPt}
\title{Draw a line segment between two points, where each point is a vecto...}
\usage{
segmentPt(pt1, pt2, clr = "black", lwidth = 1)
}
\description{
Draw a line segment between two points, where each point is a vector of x then y coordinates
}
