% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spparea.R
\name{selectrandomquad}
\alias{selectrandomquad}
\alias{selectrandomquad2}
\alias{selectrandomquad3}
\title{Draws rectangular quadrats in a plot at random.}
\usage{
selectrandomquad(size, shape, rep, plotdim = c(1000, 500),
  graphit = FALSE)

selectrandomquad2(size, rep, plotdim = c(1000, 500), graphit = FALSE)

selectrandomquad3(size, shape, rep, plotdim = c(1000, 500),
  graphit = FALSE)
}
\arguments{
\item{size}{A vector of quadrat sizes, referring to the x-dimension of a
rectangular quadrat.}

\item{shape}{The ratio of y to x dimensions of the rectangles; \code{rectdim = 1}
(the default) for squares.}

\item{rep}{The number of replicated random quadrats (per dimension submitted)
to create.}

\item{plotdim}{The x and y dimensions of the plot.}

\item{graphit}{Logical; if \code{TRUE} adds plot output.}
}
\description{
Draw rectangular quadrats in a plot at random. Returns a dataframe of
coordinates defining the corners of all the quadrats.
}
\details{
Quadrats are chosen simply: a point is chosen by randomly drawing one x and
one y coordinates to serve as the lower-left corner. The \code{x} is chosen from a
uniform distribution between 0 and \code{plotdim[1]-size}; \code{y} is chosen
similarly.

All 3 algorithms (\code{\link[=selectrandomquad]{selectrandomquad()}}, \code{\link[=selectrandomquad2]{selectrandomquad2()}} and
\code{\link[=selectrandomquad3]{selectrandomquad3()}}) under sample the plot corners. That bias,
\code{\link[=selectrandomquad2]{selectrandomquad2()}} and \code{\link[=selectrandomquad3]{selectrandomquad3()}} intended to overcome; but
neither does.
}
\section{Functions}{
\itemize{
\item \code{selectrandomquad2}: Its algorithm aims at capturing corners. The
result, however, is not to capture corners any better than
\code{\link[=selectrandomquad]{selectrandomquad()}} does.

Imagine a line running vertically at \code{x = 0} from \code{y = 0} to
\code{y = plotdim[2] - size}, then continues at \code{x = 1} from 0 to
\code{plotdim[2] - size}, etc. It's wrapping analogous to the way quadrat indices
wrap (see \code{?gxgy.to.index}). This line has length
\code{(plotdim[1] - size) * (plotdim[2] - size)}. Draw a random number on that
line, and place the lower left corner of random square at that point. A
position \code{x} on the line is converted to plot coordinates \code{gx}, \code{gy} using
function \code{index.to.gxgy()} with \code{grid = 1}.

\item \code{selectrandomquad3}: The lower left x, y coordinates are drawn at
random from a range extending outside the plot dimensions, then quadrats
are used only if all 4 corners fall inside the plot.
}}

\section{Arguments details}{

\itemize{
\item \code{graphit} Whether to graph the locations of the chosen quadrats on a
plot map.
}
}

