% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{setsize}
\alias{setsize}
\title{An internal function, called by \code{\link[=map]{map()}} to choose size of plotting points.}
\usage{
setsize(n, s)
}
\arguments{
\item{n}{Number of dbh categories}

\item{s}{Abundance of the species.}
}
\description{
An internal function, called by \code{\link[=map]{map()}} to choose size of plotting points.

Note that the user can override these defaults by
submitting sizes.
}
