% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{simple.model}
\alias{simple.model}
\title{simple.model}
\usage{
simple.model(x, param)
}
\description{
A trivial model to return a different value at every x. If param is atomic, then that value is returned for every x. Otherwise, param must be a vector of same size as x, and param is returned.
}
