% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{skewness}
\alias{skewness}
\title{skewness}
\usage{
skewness(x)
}
\description{
Sample skewness. The biased portion is the population skewness; correction is
for finite sample.

D. N. Joanes and C. A. Gill. “Comparing Measures of Sample Skewness and
Kurtosis”. The Statistician 47(1):183–189
}
