% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvetopo.R
\name{solve.topo}
\alias{solve.topo}
\title{Topographic calculation based on stake and the height difference between them.}
\usage{
\method{solve}{topo}(coldata, column1 = "pt1", column2 = "pt2",
  diffcolumn = "htdiff", basept = "1", baseelev = 0, debug = NULL)
}
\details{
Topographic calculation based on stake and the height difference between them.'
solve.topo
}
\examples{
\dontrun{

See topography tutorial}


}
