% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split.data}
\alias{split.data}
\title{split.data}
\usage{
\method{split}{data}(censdata, splitcol = "sp", keepsplitcol = FALSE,
  allsplit = NULL, showOutput = NULL)
}
\description{
Converts a big dataframe into a list of smaller dataframes, grouped using any
column in the database, or any external vector. The variable allsplit can be set to a vector
of data sections to be included; if allsplit includes values not in the data, empty elements
are included in the list, but if allsplit includes fewer values than found in the data, then
missing elements are omitted.

The option keepsplitcol can be set to TRUE
in order to retain in the new dataframes the column on which the data are grouped; otherwise, that
column is removed.
}
