% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund.fit.CTFS.R
\name{spmean.mort.abundGibbs}
\alias{spmean.mort.abundGibbs}
\title{Likelihood function for a species mean (a scalar, one species at a ...}
\usage{
spmean.mort.abundGibbs(spmean.mu, spmean.r, N1, S, time, MuSD, Rsd, type)
}
\description{
Likelihood function for a species mean (a scalar, one species at a time),
given logMu and logSD and the data, N and S (just one species here, so all
parameters are scalars). In the abundance model, the mortality parameter is
involved in the likelihood for population change, and it must thus depend on
the the fitted little r. Only used as a subroutine of the main modeling
function, model.littleR.Gibbs.
}
\seealso{
\code{\link[=spmean.mort.Gibbs]{spmean.mort.Gibbs()}}
}
