% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{standardreg.no.int}
\alias{standardreg.no.int}
\title{Standard regression with no intercept.standardreg.no.int autoregres...}
\usage{
standardreg.no.int(x, y)
}
\description{
Standard regression with no intercept.
}
