% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startCTFS_startCTFS.R
\name{startCTFS}
\alias{startCTFS}
\title{startCTFS}
\usage{
startCTFS(folder)
}
\description{
Function to source all components of CTFS R package. Load all CTFS packages into a folder,
then when opening R, execute this function, passing the folder name.
}
