% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-utils.R
\name{strip_path}
\alias{strip_path}
\title{Stip the bare file name when its sorrounded by references of its path.}
\usage{
strip_path(path)
}
\arguments{
\item{path}{A string of the form "path/file.extention".}
}
\value{
The bare file name.
}
\description{
In a way, this is a kind of reverse of path_to_dir.
}
\examples{
strip_path("man-roxygen/plotdim.R")
}
\keyword{internal}
