% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-utils.R
\name{strip_rd}
\alias{strip_rd}
\title{Remove ending .Rd or .rd.}
\usage{
strip_rd(string)
}
\arguments{
\item{string}{A string}
}
\value{
The string with the .Rd or .rd extension removed.
}
\description{
Remove ending .Rd or .rd.
}
\keyword{internal}
