% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_lmerBayes.R
\name{summaryMCMC}
\alias{summaryMCMC}
\title{summaryMCMC}
\usage{
summaryMCMC(fit, model, error, sdmodel, badparam, paramfile = NULL,
  returnfull = TRUE, ...)
}
\description{
Make summary calculations based on the full Gibbs sampler. The argument fit is an object holding all steps of the sampler, plus data, observations,
and likelihood. However, if parameters were saved along the way to a text file, then the argument paramfile is used to name the file and restore them
into a 3D array. Estimates of confidence limits of all parameters are returned. If returnfull is set TRUE, then the entire 3D array of parameters is
also returned. Full likelihood at the best parameters is calculated and likelihood at each step in sampler are used to calculate DIC.
}
