% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelBayes.R
\name{summaryModelMCMC}
\alias{summaryModelMCMC}
\title{summaryModelMCMC}
\usage{
summaryModelMCMC(fit, model, error, sdmodel, badparam, paramfile = NULL,
  returnfull = TRUE, ...)
}
\description{
Make summary calculations based on the full Gibbs sampler. The argument fit is an object holding all steps of the sampler, plus data, observations,
and likelihood. Estimates of confidence limits of all parameters are returned. Full likelihood at the best parameters is calculated and likelihood 
at each step in sampler are used to calculate DIC.
}
