% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{torus.shift}
\alias{torus.shift}
\title{torus.shift}
\usage{
torus.shift(quaddata, slip.horiz, slip.vert, invert = F, reverse = F,
  plotdim = c(1000, 500), gridsize = 20)
}
\description{
Creates a torus-shifted quadrat topographic dataset. It accepts a quadrat dataset
with elevation, convexity, and slope for each 20x20 m quadrat in a plot. It returns a parallel
dataset that is torus shifted, slip.horiz quadrats left-right and slip.vert quadrats up-down. 
That is, in the new dataset, the topographic information of each quadrat comes from a quadrat
displaced by slip.horiz and slip.vert units away in the original dataset.
}
