% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trim}
\alias{trim}
\title{trim}
\usage{
trim(s)
}
\description{
Trims leading and trailing blanks from a vector of character variables. Multibyte character strings are returned intact.
(extended ascii in R appears as hex values).
}
