% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.analysis.R
\name{wavelet.bivariate}
\alias{wavelet.bivariate}
\title{wavelet.bivariate}
\usage{
wavelet.bivariate(raster1 = NULL, raster2 = NULL, coords1, coords2,
  gridsize = 1, plotdim = c(1000, 500), FUN = NULL, k0 = 8, dj = 0.15,
  graph = TRUE)
}
\description{
Function to calculate the wavelet variance to evaluate the association
between two point patterns using furier transforms.

It accepts a raster data or a point pattern, but the type of data entered has
to be specified in the argument type (xxx see section Warning).

The wavelet variance describes the spatial autocorrelation or aggregation of point distribution.

A wavelet variance greater than 1 indicates scales at which individuals
are aggregated. A wavelet variance less than 1, indicates scales at which
individuals are dis-aggregated. A wavelet variance equal to 1, indicates scales
at which individuals are randomply distribuited (as Poisson process).

A graphical test is implemented on the null hypothesis of comple randomness.

If the wavelet variance is out of the conf bounds the point distribution
is significantly different from a random process.

Dependencies: needs the package 'spatstat'and the CTFSRpackage

Authors: Matteo Detto and Tania Brenes

Output: a list containing vectors for the wavelet variance, the scale of the wavelet variance, and the normalized variance.

coords1  (): a matrix with raster data OR a table with two or three columns that can be used to calculate a raster, the first two columns are the coordinates and the third is the mark. Type must be specified ;
coords2 (): 2nd dataset for the bivariate analyisis;
type ('raster'): the type of data entered, 'raster'if a raster matrix, 'point'for a point pattern, or 'marked'for a marked point pattern;
gridsize : is the quadrat size of the rasterization;
plotdim (c(1000,500)): the dimensions of the plot;

FUN ('sum'): function to apply to the marked point pattern, by default it sums the values as would be used for sum of basal areas or sum of above graound biomass
k0 (8): numeric. smoothing parameter of the wavelet filter (k0 between 5.5-15),
lower values of k0 produce a smoother wavelet variance;
dj  (0.15):  numeric. discretization of the scale axis;
graph (TRUE): logical. plot the wavelet variace ?
}
\section{Warning}{

If the argument type is ignored the function works. But one issue with this
function is that the description mentions the argument \code{type}, but \code{type} is
not part of the function definition not is passed to any other function.
Confusingly, this function calls \code{plot()}, which has argument \code{type} but
seems to have nothing to do with the \code{tpye} referred to in the description of
this function. In the example, \code{type} is kept as a reminder of this issue but
is with a comment.
}

\examples{
\dontrun{
sp.one = subset(bci::bci12full7, sp == "quaras")[, c("gx", "gy")]
sp.two = subsetbci::bci12full7, sp == "cordal")[, c("gx", "gy")]
wv = wavelet.bivariate(
  coords = sp.one,
  coords2 = sp.two,
  #' type = 'point',  # dissabled because it errs, see section Warning
  k0 = 8,
  dj = 0.15,
  graph = TRUE
)
# plots the scale of aggregation
}

}
