% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.analysis.R
\name{wavelet.univariate}
\alias{wavelet.univariate}
\title{Univariate wavelet variance using furier transforms.}
\usage{
wavelet.univariate(raster = NULL, coords, gridsize = 2, plotdim = c(1000,
  500), FUN = NULL, k0 = 8, dj = 0.15, graph = FALSE)
}
\arguments{
\item{raster}{Used if data is entered already as a raster matrix.}

\item{coords}{An alternate to a raster table, a table with two (or three)
columns giving coordinates x, y (and an optional mark) in that order. This
is used to calculate a raster.}

\item{gridsize}{Side of the square quadrate.}

\item{plotdim}{The x and y dimensions of the plot.}

\item{FUN}{Function to apply to the marked point pattern, by default it sums
the values as would be used for sum of basal areas or sum of above graound
biomass}

\item{k0}{Numeric. Smoothing parameter of the wavelet filter. (k0 between
5.5-15), lower values of k0 produce a smoother wavelet variance.}

\item{dj}{Numeric. Discretization of the scale axis.}

\item{graph}{Logical. If TRUE plot the wavelet variace.}
}
\value{
A list containing vectors for the wavelet variance, the scale of the
wavelet variance, the normalized variance, and the confidence intervals.
}
\description{
Function to calculate the univariate wavelet variance using furier
transforms.
}
\details{
It accepts a raster data or a point pattern, which is the default if raster
is not provided.

The wavelet variance describes the spatial autocorrelation or aggregation of
tree distribution.

A wavelet variance greater than 1 indicates scales at which individuals are
aggregated. A wavelet variance less than 1, indicates scales at which
individuals are dis-aggregated. A wavelet variance equal to 1, indicates
scales at which individuals are randomply distribuited (as Poisson process).

A graphical test is implemented on the null hypothesis of complete
randomness.

If the wavelet variance is out of the conf bounds the tree distribution is
significantly different from a random process.

Dependencies: needs the package 'spatstat'and the CTFSRpackage
}
\section{Parameter details}{

\itemize{
\item FUN If not specified (default) it counts points. E.g. can be used to sum
the basal areas or above graound biomass.
}
}

\examples{
\dontrun{
load("bci.full1.rdata") 
rast1 = rasterize(,
  gridsize = 5,
  plotdim = c(100, 500),
  graph = TRUE)
wv = wavelet.univariate(
  coords = bci.full1[, c("gx", "gy")],
  k0 = 8,
  dj = 0.15,
  graph = TRUE
  )
# plots the scale of aggregation
}

}
\seealso{
\code{\link[=wavelet.bivariate]{wavelet.bivariate()}}
}
\author{
Matteo Detto and Tania Brenes
}
