% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{weibull.mean.3param}
\alias{weibull.mean.3param}
\title{Mean of three-parameter Weibullweibull.mean.3param weibull.sd.3para...}
\usage{
weibull.mean.3param(x0, shape, scale)
}
\description{
Mean of three-parameter Weibull
}
