% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{weibull.median.3param}
\alias{weibull.median.3param}
\title{Median of three-parameter Weibullweibull.median.3param weibull.mean...}
\usage{
weibull.median.3param(x0, shape, scale)
}
\description{
Median of three-parameter Weibull
}
