% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{weibull.sd.3param}
\alias{weibull.sd.3param}
\title{SD of three-parameter Weibullweibull.sd.3param dexp.sin  Four-param...}
\usage{
weibull.sd.3param(x0, shape, scale)
}
\description{
SD of three-parameter Weibull
}
