% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{which.allnonNA}
\alias{which.allnonNA}
\title{which.allnonNA}
\usage{
which.allnonNA(x)
}
\description{
Finds all values of a vector which are not NA. Fills out a vector to
length 6 with NAs if there are fewer than 6.
}
