% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_utilities.R
\name{which.maxNAs}
\alias{which.maxNAs}
\title{which.maxNAs}
\usage{
which.maxNAs(v)
}
\description{
An unfortunate bug in which.max: if all elements are NA, it doesn't return anything. This means that for any vector, which.max returns
a vector of length 1 unless all are NAs. This is a silly error, since a program expecting an element of length 1 falls apart otherwise.
}
