% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{which.nonNA}
\alias{which.nonNA}
\title{Returns the first index at which a vector x is not NAwhich.nonNA wh...}
\usage{
which.nonNA(x)
}
\description{
Returns the first index at which a vector x is not NA
}
