% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{which.vmatch}
\alias{which.vmatch}
\title{Finds which subsequent element of a vector matches the first elemen...}
\usage{
which.vmatch(v)
}
\description{
Finds which subsequent element of a vector matches the first element.
}
