% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_growthfit.bin.R
\name{wideEnoughBins}
\alias{wideEnoughBins}
\title{wideEnoughBins}
\usage{
wideEnoughBins(x, b, ...)
}
\description{
Test whether all the bin widths exceed a minimum. 





MINBIN=0.1
MINBINSAMPLE=5
}
