% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-biomass.R
\name{wsgdata_dummy}
\alias{wsgdata_dummy}
\title{Create a dummy wsgdata data set}
\usage{
wsgdata_dummy(censdata, plot = "bci")
}
\arguments{
\item{censdata}{Census data, e.g. \code{bci::bci12full1}.}

\item{plot}{An character string of length 1 giving the plot name where
\code{censdata} comes from.}
}
\value{
A data frame with one row per species in \code{censdata} and 3 variables.
}
\description{
Create a dummy wsgdata data set with random values of species density.
}
\details{
If you lack a real wood density table, you can use this function to create a
dummy one. For example, this may help you to test \code{biomass.CTFSdb()} and
\code{density_ind()}.

See \code{?density_ind()}, particularly notice the following: \emph{"There must be a
table of wood density submitted (wsgdata), and this table must have a column
sp with species names, a column plot, plus the wood density in a column
called wsg (though the name of that column can be changed using the argument
denscol)"}.
}
\examples{
wsgdata_dummy(bci::bci12stem1)
}
