% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{xydistvect}
\alias{xydistvect}
\title{xydistvect}
\usage{
xydistvect(pt1, pt2)
}
\description{
Distance between two x-y coordinates, but accepts each set of coordinates as a vector of length 2,
with the first element the x coordinates, the second y.
}
