% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abund.manycensus}
\alias{abund.manycensus}
\title{abund.manycensus}
\usage{
abund.manycensus(allcns = list(bci.full1, bci.full2, bci.full3), mindbh,
  dbhunit = "mm", type = "abund", excludespp = NULL,
  excludestatus = NULL)
}
\description{
Collect abundances of all species across several censuses. The full R census tables are submitted as a list, as many as desired. The
argument type can be used to choose basal area or agb, or the default for number of individuals. The mindbh to include must be given
as an argument, but it can be NULL. If the latter, trees are counted if they have no dbh, as long as status=A. By default, any tree
ever given code M is not counted in any census, but set excludestatus=NULL to include them.
A character vector of species codes can be submitted as excludespp, for instance those for unidentified trees.
}
\examples{
\dontrun{
N=abund.manycensus(allcns=list(bci.full1,bci.full2,bci.full3,bci.full4),mindbh=10,type='abund',excludespp='uniden',excludestatus='M')
head(N)
colSums(N)
apply(N,2,countspp)
N=abund.manycensus(allcns=list(bci.full5,bci.full6),mindbh=10,type='abund',excludespp=c('uniden','tremxx'),excludestatus=NULL)}

}
