% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_abundance.R
\name{abundance}
\alias{abundance}
\title{abundance}
\usage{
abundance(censdata, type = "abund", alivecode = c("A"), mindbh = NULL,
  dbhunit = "mm", split1 = NULL, split2 = NULL)
}
\arguments{
\item{censdata}{an R Analytical Table for a full plot census, either full or stem}

\item{type}{must be either 'abund', 'ba', or 'agb'}

\item{alivecode}{all codes defining alive; the default 'A'is the standard CTFS designation for living trees or stems}

\item{mindbh}{the minimum diameter above which the counts are done; if NULL, all (living) are included}

\item{dbhunit}{'cm'or 'mm', only used for basal area}

\item{split1}{a vector of categories, one per individual}

\item{split2}{another vector of categories, one per individual}
}
\description{
Calculates total abundance or basal area, dividing data with 1 or 2 categorical variables. 
The categorical variables must be submitted as vectors whose length
matches exactly as the number of rows in the plot data submitted (so one per tree or stem). 
The first vector should be the one with the most categories
(for instances, split1=species, split2=dbhcategory). 
The mean measurement date for all individuals in each category is
also calculated. 
For abundance, set type='abund'(the default), for basal area, set type='ba'. Be sure to use the R Analytical
Stem Table for basal area. For abundance, either the stem table or full table can be used, the former to count stems, the latter trees. 
The object returned is a list, the first element named either abund or ba, the second named meandate. Each is an array of 
one or two dimensions, depending on how many split variables were submitted:
each of the dimensions of the array handles one of the categorical variables.
}
\examples{
\dontrun{
CTFSplot('bci',5:6,'full')
CTFSplot('bci',5:6,'stem')
total=abundance(bci.full5,mindbh=10)
total$abund
total$meandate
totalstem=abundance(bci.stem5,mindbh=10)
BAperSpecies=abundance(bci.stem5,type='ba',mindbh=10,split1=bci.stem5$sp)
head(BAperSpecies$ba)
head(BAperSpecies$meandate)}


}
