% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{Abundance or basal area dividing data with 1 or 2 categorical variables}
\usage{
abundance(censdata, type = "abund", alivecode = c("A"), mindbh = NULL,
  dbhunit = "mm", split1 = NULL, split2 = NULL)
}
\arguments{
\item{censdata}{A table of plot census data.}

\item{type}{Either 'abund' (default) for abundance, 'ba' for basal area, or
'agb'.}

\item{alivecode}{character, codes of the variable `status` that indicate the 
tree is alive. The default 'A' is the standard CTFS designation for living 
trees or stems}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{dbhunit}{'cm' or 'mm', only used for basal area}

\item{split1}{a vector of categories, one per individual}

\item{split2}{another vector of categories, one per individual}
}
\value{
a named list of two elements, either \code{abund} or \code{ba}, and meandate
(mean measurement date for all individuals in each category). Each element
of the list is an array of one or two dimensions, depending on how many
split variables were submitted: each of the dimensions of the array handles
one of the categorical variables.
}
\description{
Calculates total abundance or basal area, dividing data with 1 or 2
categorical variables.
}
\details{
The categorical variables must be submitted as vectors whose length matches
exactly as the number of rows in the plot data submitted (so one per tree or
stem). The first vector should be the one with the most categories (for
instances, \code{split1 = species}, \code{split2 = dbhcategory}).

For basal area, pass a stem table to \code{censdata}; For abundance, use either
the stem or full table to count stems or trees, respectively.
}
\examples{
\dontrun{

CTFSplot('bci',5:6,'full')

CTFSplot('bci',5:6,'stem')
total=abundance(bci.full5,mindbh=10)
total$abund
total$meandate
totalstem=abundance(bci.stem5,mindbh=10)

BAperSpecies=abundance(bci.stem5,type='ba',mindbh=10,split1=bci.stem5$sp)
head(BAperSpecies$ba)
head(BAperSpecies$meandate)}


}
