% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance.spp}
\alias{abundance.spp}
\title{abundance.spp}
\usage{
abundance.spp(censdata, type = "abund", dbhunit = "mm",
  alivecode = c("A"), dbhbreaks = c(10, 100, 300))
}
\arguments{
\item{dbhbreaks}{a vector of dbhs to define divisions of categories; the last category will be >= the final division}
}
\description{
A wrapper to calculate total abundance (or ba or agb) for each species in given dbh categories. The dbh categories
are set with dbhbreaks. See abundance() for description of the other arguments and return value.
}
