% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundanceperquad}
\alias{abundanceperquad}
\title{Finds abundance, basal area, or agb of every species per square qua...}
\usage{
abundanceperquad(censdata, mindbh = 10, plotdim = c(1000, 500),
  gridsize = 100, type = "abund", dbhunit = "mm")
}
\description{
Finds abundance, basal area, or agb of every species per square quadrat of any size; plotdim is the x dimension then y dimension of the plot and
must be set correctly; gridsize is the quadrat dimension. The plot is divided into a checkerboard of non-overlapping, space-filling squares.

If the plot dimensions is not an exact multiple of the quadrat size, then a strip at the upper edge of the plot (north and east if plot
is on cardinal directions) is omitted. For example, if gridsize=40 and plotdim=500, then there are an extra 20 meters at the upper boundary
omitted from the calculations.

See abundance() for description of the other arguments and return value. The array of abundances per quadrat is useful for similarity, counting
species and stems per quadrat, etc.
}
\examples{
\dontrun{

Nperquad=abundanceperquad(bci.full6,plotdim=c(1000,500),gridsize=100,type='abund')
colSums(Nperquad$abund)
apply(Nperquad$abund,2,countspp)
plot(colSums(Nperquad$abund),apply(Nperquad$abund,2,countspp))}


}
