% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundanceperquad}
\alias{abundanceperquad}
\title{Abundance, basal area, or agb of every species by quadrat.}
\usage{
abundanceperquad(censdata, mindbh = 10, plotdim = c(1000, 500),
  gridsize = 100, type = "abund", dbhunit = "mm")
}
\arguments{
\item{censdata}{A table of plot census data.}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than \code{mindbh} are excluded. If \code{NULL}, all living trees are
included.}

\item{plotdim}{The x and y dimensions of the plot.}

\item{gridsize}{Side of the square quadrate.}

\item{type}{Either 'abund' (default) for abundance, 'ba' for basal area, or
'agb'.}

\item{dbhunit}{'cm' or 'mm', only used for basal area}
}
\value{
See \code{\link[=abundance]{abundance()}}.
}
\description{
Finds abundance, basal area, or agb of every species per square quadrat of
any size; plotdim is the x dimension then y dimension of the plot and must be
set correctly; gridsize is the quadrat dimension. The plot is divided into a
checkerboard of non-overlapping, space-filling squares.
}
\details{
If the plot dimensions is not an exact multiple of the quadrat size, then a
strip at the upper edge of the plot (north and east if plot is on cardinal
directions) is omitted. For example, if \code{gridsize = 40} and \code{plotdim = 500},
then there are an extra 20 meters at the upper boundary omitted from the
calculations.

The array of abundances per quadrat is useful for similarity, counting
species and stems per quadrat, etc.
}
\examples{
\dontrun{
Nperquad = abundanceperquad(
  bciex::bci12t6mini,
  plotdim = c(1000, 500),
  gridsize = 100,
  type = 'abund'
)
colSums(Nperquad$abund)
apply(Nperquad$abund, 2, countspp)
plot(colSums(Nperquad$abund), apply / (Nperquad$abund, 2, countspp))
}

}
\seealso{
\code{\link[=abundance]{abundance()}}
}
