% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund.fit.CTFS.R
\name{abundmodel.fit}
\alias{abundmodel.fit}
\title{Simply return the modeled histogram for any set of parameters.}
\usage{
abundmodel.fit(x, fit, modeltype, conf = NULL)
}
\arguments{
\item{fit}{A fitted model.}

\item{modeltype}{Functional forms to fit to the distribution; it can be:
* Gaussian `modeltype = "norm"`, with the quotes
* Asymmetric Gaussian (a different standard deviation on left and right of
the mode) `modeltype = "asymnorm"`, with the quotes
* Laplace (exponential distribution, with mirror image for negative values)
`modeltype = "symexp"`, with the quotes
* Asymmetric Laplace (different rate constant for left and right of the
center) `modeltype = "asymexp"`, with the quotes
* Asymmetric power distribution (different rate constant for left and right
of the center) `modeltype = "asympower", with the quotes.}

\item{conf}{number of alternate fits to graph, as indication of confidence;
if conf = NULL, no confidence lines are added}
}
\description{
Simply return the modeled histogram for any set of parameters.
}
\details{
Used as a subroutine inside \code{\link[=graph.abundmodel]{graph.abundmodel()}}.
}
