% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogchange_abund.fit.CTFS.R
\name{abundmodel.fit}
\alias{abundmodel.fit}
\title{abundmodel.fit}
\usage{
abundmodel.fit(x, fit, modeltype, conf = NULL)
}
\description{
Simply return the modeled histogram for any set of parameters. Used as a subroutine inside graph.abundmodel.
}
