% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.CTFSdb.R
\name{AGB.ind}
\alias{AGB.ind}
\title{Above ground biomass (agb) based on one of the Chave (Oecologia, 2005).}
\usage{
AGB.ind(df, dbhunit = "mm", plot = "bci", wsgdata = wsg.ctfs2,
  forest = "moist", ht.param = NULL, htmodel = predht.asym)
}
\arguments{
\item{df}{A table with dbh and species names}

\item{dbhunit}{'cm' or 'mm', only used for basal area}

\item{plot}{Name of plot as it appears in the names of the R Analytical Tables, in quote marks}

\item{wsgdata}{A wood-density table, see \code{\link[=density_ind]{density_ind()}} and \link{wsgdata_dummy}.}

\item{forest}{A forest type (for most lowland tropical plots, the default
"moist" is recommended.)}

\item{ht.param}{Height parameters to calculate height from diameter for every
tree, using the Chave model with height passed to \code{htmodel}. If NULL
(defauld) \code{AGB.ind()} uses Chave equations without height.}

\item{htmodel}{Height model to use with a non-NULL \code{ht.param}.}
}
\value{
A vector of biomass in kg for every individual in the submitted table
\code{df}.
}
\description{
Compute biomass (agb) based on one of the Chave (Oecologia, 2005) models for
tropical forest types.
}
\examples{
\dontrun{
biomass <- AGB.ind(df = bci.full1)
hist(log(biomass), breaks = 100)
sum(biomass, na.rm = TRUE) / 50
}

}
\seealso{
\code{\link[=density_ind]{density_ind()}}; Chave (Oecologia, 2005).
}
