% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.CTFSdb.R
\name{AGB.tree}
\alias{AGB.tree}
\title{AGB of each tree, grouping all stems of one tree and adding there agbs.}
\usage{
AGB.tree(df, dbhunit = "mm", plot = "bci", wsgdata = wsg.ctfs2,
  forest = "moist", ht.param = NULL, htmodel = predht.asym)
}
\arguments{
\item{df}{The submitted table. Must have dbh, species name (sp), and a treeID
to identify which tree every stem belong to. There must be just one dbh for
each stem.}

\item{dbhunit}{'cm' or 'mm', only used for basal area}

\item{plot}{Character string giving the name of the plot where the data comes
from (as it appears in the names of the R Analytical Tables).}
}
\value{
Adataframe with one row per tree, including the treeID and
total agb per tree. Note that it will have fewer rows than the table
submitted.
}
\description{
Computes AGB of each tree in a table, grouping all stems of one tree and
adding there agbs.
}
\examples{
\dontrun{
biomasstbl = AGB.tree(df = bci::bci12stem1)
dim(bci::bci12stem1)
dim(biomasstbl)
head(biomasstbl)
}

}
\seealso{
This is called by \code{\link[=biomass.CTFSdb]{biomass.CTFSdb()}} in the standard calculation of
biomass for CTFS R tables.
}
