% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{allquadratslopes}
\alias{allquadratslopes}
\title{Calculates the slope of all quadrats in a plot.  @details `allquadr...}
\usage{
allquadratslopes(elev, gridsize = 20, plotdim = c(1000, 500),
  edgecorrect = TRUE)
}
\description{
Calculates the slope of all quadrats in a plot.
}
\details{
\code{allquadratslopes()} goes through all 20x20 m quadrats in a plot and finds
the slope, mean elevation, and convexity of each. Convexity is the mean
elevation of one 20x20 m quadrat relative (minus) the mean of its immediate
neighbors.

Helene Muller-Landau added a section to correct convexity in edge quadrats
}
\examples{
\dontrun{
# The input to elev is very specific; you may need to tweak it.
elev_asis <- bci::bci_elevation
head(elev_asis)
elev_tweaked <- list(col = elev_asis)
head(elev_tweaked)

result <- allquadratslopes(
  elev = elev_tweaked,
  gridsize = 20,
  plotdim = c(1000, 500),
  edgecorrect = TRUE
)
head(result)
str(result)
}
}
\seealso{
\code{\link[=calcslope]{calcslope()}}, \code{\link[=quadslope]{quadslope()}}
}
