% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{are.ptsinside}
\alias{are.ptsinside}
\title{Check if points belong to quadrat.}
\usage{
are.ptsinside(pts, coord)
}
\arguments{
\item{pts}{A matrix.}
}
\value{
A logical vector; TRUE for the points inside.
}
\description{
Checks many points (dataframe pt with x and y) against a single quadrat whose
corners are given by as xlo, ylo, xhi, yhi.
}
\details{
This is same as \code{\link[=insideRectange]{insideRectange()}}, but accepting input as a matrix pts and a
single vector of the four corners of the rectange.
}
\seealso{
\code{\link[=insideRectange]{insideRectange()}}.
}
