% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{attach_if_needed}
\alias{attach_if_needed}
\title{attach_if_needed}
\usage{
attach_if_needed(datafiles)
}
\description{
Attach one or more datafiles,checking first whether the file is already attached. If it is attached,
it is not reattached, and the search position where attached is returned.
}
