% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{ba}
\alias{ba}
\title{Calculates the individual basal areas (in square meters) for all su...}
\usage{
ba(dbh, dbhunit = "mm")
}
\description{
Calculates the individual basal areas (in square meters) for all submitted dbhs. The dbh units must be submitted, either
cm'or 'millimeters'. The return value is a vector of basal area values of same length as the submitted vector of dbhs.
}
