% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{bad.binsdpar}
\alias{bad.binsdpar}
\title{bad.binsdpar}
\usage{
bad.binsdpar(x, param, ...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=wideEnoughBins]{wideEnoughBins()}}.}
}
\description{
bad.binsdpar
}
