% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_abundance.R
\name{basum}
\alias{basum}
\title{basum}
\usage{
basum(dbh, mindbh = 10, dbhunit = "mm")
}
\description{
Returns the basal area summed over all submitted dbhs. NAs can be included, as sum will be completed with na.rm=TRUE.
}
