% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{basum}
\alias{basum}
\title{Returns the basal area summed over all submitted dbhs.}
\usage{
basum(dbh, mindbh = 10, dbhunit = "mm")
}
\arguments{
\item{dbh}{DBH in mm of tree}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{dbhunit}{'cm' or 'mm', only used for basal area}
}
\description{
Returns the basal area summed over all submitted dbhs. NAs can be included,
as sum will be completed with \code{na.rm = TRUE}.
}
