% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{calculateBinModel.AIC}
\alias{calculateBinModel.AIC}
\title{Calculate AIC of the model, using various log(likelihood) estimators.}
\usage{
calculateBinModel.AIC(fit, type = "optim")
}
\arguments{
\item{fit}{A fitted model.}

\item{type}{One of:
\itemize{
\item \code{optim}, the highest likelihood found by optim;
\item \code{mean}, the mean llikelihood from the Gibbs sampler;
\item \code{gibbs}, the maximum llikelihood from the Gibbs sampler;
\item \code{none}, just return the mean Gibbs likelihood.
}}
}
\description{
Calculate AIC of the model, using various log(likelihood) estimators.
}
