% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_growthfit.bin.R
\name{calculateBinModel.bestpred}
\alias{calculateBinModel.bestpred}
\title{calculateBinModel.bestpred}
\usage{
calculateBinModel.bestpred(fit)
}
\description{
Calculate mean predicted value at every x using every one of the Gibbs sampler parameter combinations
(excluding burn in)
}
