% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{calculateBinModel.BIC}
\alias{calculateBinModel.BIC}
\title{calculateBinModel.BIC}
\usage{
calculateBinModel.BIC(fit)
}
\description{
Calculate Bayes Information Criteria using Wikipedia formula





Description: Calculate Deviance Information Criteria, using Wikipedia formula
}
