% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{calculateBinModel.BIC}
\alias{calculateBinModel.BIC}
\title{Calculate Bayes Information Criteria using Wikipedia formula.}
\usage{
calculateBinModel.BIC(fit)
}
\arguments{
\item{fit}{A fitted model.}
}
\description{
Calculate Bayes Information Criteria using Wikipedia formula.
}
