% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{circle}
\alias{circle}
\title{Calculates points on a circle fullcircle  Create a dataframe for a ...}
\usage{
circle(x, center = c(0, 0), radius = 5, half = "top")
}
\description{
Calculates points on a
}
