% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_utilitiesCTFS.R
\name{clean.demography}
\alias{clean.demography}
\title{clean.demography}
\usage{
clean.demography(demogdata, type = "mort", Ncol = NULL, Scol = NULL,
  timecol = NULL, ratecol = NULL, excludespp = NULL)
}
\description{
This takes a CTFS demography table, output by functions mortality, growth, or recruitment,
and removes rows where N==0, or key data are NA. The rownames are assumed to refer to species names, and
some codes can be excluded using the argument excludespp. The four columns
can be submitted by name or number using default column headers, or by
setting type to 'mort'or 'abund'.
It returns a logical vector indicating with TRUE which rows to keep, not the cleaned table itself.
This was formerly in utilities.r.
}
\examples{
\dontrun{

rows.include=clean.demography(demogtable,type='growth',Ncol='N',ratecol='ratecol');
result=demogtable[rows.include,]}





}
