% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{complete.plotmap}
\alias{complete.plotmap}
\title{complete.plotmap}
\usage{
complete.plotmap(cns = bci.full6, spnames = NULL, mindbh = 10,
  export = "no", nospp = 3, plotdim = c(1000, 500), clrlist = c("blue",
  "green", "red", "yellow", "gray"), ptsize = c(0.45, 0.3), xrange = c(0,
  100), yrange = c(0, 100), wd = 1100, ht = 850, side = 6,
  labsize = 1.75, axisdiv = 10, filepath = "/home/condit/data/maps/",
  outfile = "fullplotmap")
}
\arguments{
\item{cns:}{a full census dataset (all species)}

\item{spnames:}{names of species to map, using the mnemonic in the R tables}

\item{mindbh:}{smallest dbh to include}

\item{export:}{set to 'no'to graph to screen, 'pdf'to export to pdf (see define.graphwindow function)}

\item{nospp:}{number of species to overlay; can be NULL or 0 for none}

\item{plotdim:}{x and y plot dimensions}

\item{clrlist:}{colors to use for the species to be overlaid}

\item{ptsize:}{size of points, the first used for the background of all species, the second for the individual species}

\item{xrange:}{minimum and maximum x coordinates of area graphed}

\item{yrange:}{minimum and maximum y coordinates of area graphed}

\item{wd:}{graph width; see map() function}

\item{ht:}{graph height, same units as wd}

\item{side:}{the side in inches of the graph; see map()}

\item{filepath:}{folder to save output}

\item{outfile:}{filename for output}
}
\description{
This creates a map with every individual shown. The area to be mapped can be chosen, allowing maps of small areas.
On top of the map of every individual, individual species can be overlaid in different colors. If nospp is set NULL, no species
are added. If spnames is set NULL, then the most abundant species in the plot are chosen, up to the number nospp.
}
