% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spparea.R
\name{coverage.diag}
\alias{coverage.diag}
\title{Draws a diagonal across a plot.}
\usage{
coverage.diag(randomquads, slope = 1, plotdim = c(1000, 500),
  graphit = FALSE)
}
\arguments{
\item{slope}{A number that controls the direction of the diagonal drawn
accross the plot. To draw the diagonal
\itemize{
\item from lower left to upper right, use 1;
\item from upper left to lower right, use -1;
\item straight across the middle, use 0.
}}

\item{plotdim}{The x and y dimensions of the plot.}

\item{graphit}{Logical; if \code{TRUE} adds plot output.}

\item{randomquads:}{Defines the four corners of the quadrats in xlo, ylo,
xhi, yhi.}
}
\description{
Draws a diagonal across a plot, and determines for every point along the
diagonal what fraction of a series of quadrats it is inside.
}
\details{
This is only used in testing how well random quadrat draws include corners
and edges of a plot.
}
