% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{define.graphwindow}
\alias{define.graphwindow}
\title{Define an export device for a graph, based on the argument export.}
\usage{
define.graphwindow(export = "no", w, h, file)
}
\arguments{
\item{export}{\code{export = "no"}, does nothing, so the next graph uses the
default R device. With this option, the size of the graph cannot have been
set larger than the default size, which is usually 7 inches. \code{export = "unix"}, a new graphics window in unix, size \code{h} and width \code{w}, will be
opened allowing you to alter the size of the output. \code{export = "windows"},
or \code{export = "mac"} are similar to \code{export = "unix"} but for windows and
mac. \code{export = format} exports the graph to a specific file format: "pdf",
"png", "jpg", "bmp", "emf".}

\item{h, w}{A number giving the height and width.}
}
\description{
This defines an export device for a graph, based on the argument export.
}
