% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{define.graphwindow}
\alias{define.graphwindow}
\title{Define an export device for a graph, based on the argument export.}
\usage{
define.graphwindow(export = "no", w, h, file)
}
\arguments{
\item{export}{* If `export = "no"`, does nothing, so the next graph uses the default R
device. With this option, the size of the graph cannot have been set larger
than the default size, which is usually 7 inches. 
* If `export = "unix"`, a new graphics window in unix, size `h` and width `w`,
will be opened allowing you to alter the size of the output. To do the same
in Windows, set `export = "windows"`, or for Mac, `export = "mac"`.
* To export the graph to a file, export can be "pdf", "png", "jpg", "bmp",
"emf".}

\item{h, w}{Size and width}
}
\description{
This defines an export device for a graph, based on the argument export.
}
