% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dgamma.meansd}
\alias{dgamma.meansd}
\title{dgamma.meansd}
\usage{
dgamma.meansd(x, mean, sd, log = FALSE)
}
\description{
Probability distribution of gamma, parameterized with mean and sd instead of shape and scale.
}
