% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dnormrev}
\alias{dnormrev}
\title{Reverse the order of parameters to dnorm.}
\usage{
dnormrev(m, x, s, log = F)
}
\arguments{
\item{m, x, s}{Passed to \code{mean}, \code{x} and \code{sd} in \code{\link[=dnorm]{dnorm()}}.}
}
\description{
This reverses the order of parameters to dnorm, so that outer can be used
with a vector of x, and two vectors for mean and sd (the latter two equal in
length).
}
\seealso{
\code{\link[=dnorm]{dnorm()}}.
}
