% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{doublenormal}
\alias{doublenormal}
\title{Simulate draws from a double normaldoublenormal.}
\usage{
doublenormal(N = 10000, m = 0, sd1 = 1, sd2 = seq(0, 2, by = 0.1))
}
\arguments{
\item{N, m, sd1}{Passed to \code{n}, \code{mean}, \code{sd} in \code{\link[stats:rnorm]{stats::rnorm()}}.}
}
\description{
Simulate draws from a double normal
}
\seealso{
\code{\link[stats:rnorm]{stats::rnorm()}}.
}
