% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dpois.rearrange}
\alias{dpois.rearrange}
\title{dpois.rearrange}
\usage{
dpois.rearrange(v, log = F)
}
\description{
This rearranges dpois so that it works on a single vector, with the first
element being x and the remaining all being used as lambdas.
}
