% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dpois.trunc}
\alias{dpois.trunc}
\title{A zero-truncated Poisson distribution. dpois.trunc dpois.maxtrunc  ...}
\usage{
dpois.trunc(x, lambda, log = FALSE)
}
\arguments{
\item{x}{See \code{x}, \code{lambda} in \code{\link[stats:dpois]{stats::dpois()}}.}

\item{lambda}{See \code{x}, \code{lambda} in \code{\link[stats:dpois]{stats::dpois()}}.}
}
\description{
A zero-truncated Poisson distribution.
}
