% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapresponse.R
\name{draw.axes}
\alias{draw.axes}
\title{Add axes to a graph, with ticks spaced by the vector div.}
\usage{
draw.axes(whichaxs = 1:4, xrange, yrange, div, whichbox = 1:4)
}
\arguments{
\item{xrange}{if \code{NULL}, all \code{x} (or \code{y}) are included, otherwise, the
graph will only include \code{x} (or \code{y}) inside \code{xrange} or (\code{yrange})}

\item{yrange}{if \code{NULL}, all \code{x} (or \code{y}) are included, otherwise, the
graph will only include \code{x} (or \code{y}) inside \code{xrange} or (\code{yrange})}
}
\description{
Add axes to a graph, with ticks spaced by the vector div. This is a
subroutine used in \code{\link[=imageGraph]{imageGraph()}}.
}
