% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_mapresponse.R
\name{draw.axes}
\alias{draw.axes}
\title{draw.axes}
\usage{
draw.axes(whichaxs = 1:4, xrange, yrange, div, whichbox = 1:4)
}
\description{
Add axes to a graph, with ticks spaced by the vector div. This is a subroutine used in imageGraph().
}
