% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapresponse.R
\name{draw.axes}
\alias{draw.axes}
\title{Add axes to a graph, with ticks spaced by the vector div.}
\usage{
draw.axes(whichaxs = 1:4, xrange, yrange, div, whichbox = 1:4)
}
\arguments{
\item{xrange}{Minimum and maximum x and y coordinates of area graphed.
If `NULL`, all `x` (or `y`) are included.}

\item{yrange}{Minimum and maximum x and y coordinates of area graphed.
If `NULL`, all `x` (or `y`) are included.}
}
\description{
Add axes to a graph, with ticks spaced by the vector div. This is a
subroutine used in \code{\link[=imageGraph]{imageGraph()}}.
}
