% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{drawrectangle}
\alias{drawrectangle}
\title{Draw a rectangle given a matrix or dataframe of 4 x-y coordinates.}
\usage{
drawrectangle(corners, add = TRUE, clr = "black", lwidth = 1)
}
\arguments{
\item{add}{Logical. Whether to add elements such as points or line to an
existing graph. if TRUE, there must be an appropriate existing graph,
otherwise, a new graph is drawn.}

\item{clr}{A string indicating colour. See argument col in [graphics::par()].}
}
\description{
Draw a rectangle given a matrix or dataframe of 4 x-y coordinates. The
columns must have x first then y coordinates.
}
