% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dsymexp}
\alias{dsymexp}
\title{dsymexp}
\usage{
dsymexp(x, center, rate, log = FALSE)
}
\description{
Probability distribution for a folded, symmetrical exponential. When x>=center, 
it's just a standard exponential. When x<center, it's the mirror image of same one.
Each must be divided by two, though, in order to integrate to one.
}
