% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{elev.to.list}
\alias{elev.to.list}
\title{Convert a dataframe of coordinates and elevation into a list.}
\usage{
elev.to.list(elevfile, gridsize = 5)
}
\arguments{
\item{elevfile}{A data frame with x, y coordinates and elevation,
bcielev.info.}

\item{gridsize}{scale of elevation values (m)}
}
\value{
\itemize{
\item object 1: the input dataframe
\item object 2: matrix of elevation value sin the orientation of the plot
}
}
\description{
Reads a dataframe with variable names \code{x} (coordinate), \code{y} (coordinate),
\code{elev} (elevation) for a given grid size and converts to a list.
}
