% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{elev.to.list}
\alias{elev.to.list}
\title{elev.to.list}
\usage{
elev.to.list(elevfile, gridsize = 5)
}
\description{
A function which reads a dataframe with x,y,elevation for a given grid
size and converts to a list
note the names of the columns in the dataframe must be:
x y elev

object 1 in list is: the input dataframe
object 2 in list is: matrix of elevation value sin the orientation of the plot
}
