% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{enoughSamplePerBin}
\alias{enoughSamplePerBin}
\title{enoughSamplePerBin}
\usage{
enoughSamplePerBin(x, b, ...)
}
\description{
Test whether the number of elements in a vector x between successive breaks exceeds a minimum. If any
bins have too few, it returns FALSE.
}
