% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_mortality.R
\name{find.climits}
\alias{find.climits}
\title{find.climits}
\usage{
find.climits(N, D, alpha = 0.05, kind = "upper")
}
\arguments{
\item{N, }{number of individuals alive at the outset}

\item{D, }{number of deaths by the end}

\item{alpha, }{the critical probability (default alpha=0.05 gives 95% confidence limits)}

\item{kind, }{either "upper" or "lower"}
}
\description{
Calculates confidence limits around a number of deaths, D, out of N individuals.
It uses the beta distribution as the conjugate of the binomial, so the beta is the posterior of the number
dying. N and D can be vectors or matrices, but must have matching dimensions.
}
\examples{
\dontrun{
find.climits(10,5,kind='lower')}

}
