% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund.fit.CTFS.R
\name{find.xaxis.hist}
\alias{find.xaxis.hist}
\title{find.xaxis.hist}
\usage{
find.xaxis.hist(fit, tinydiv = 0.001, xrange)
}
\description{
Given an abundance fit and x axis range and divisions, return a sequence of x values for drawing the histogram. Used as a subroutine inside graph.abundmodel.
}
