% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{findborderquads}
\alias{findborderquads}
\title{Calculate indices of neighboring quadrats, for a given quadrat inde...}
\usage{
findborderquads(index, dist = 20, gridsize = 20, plotdim = c(1000, 500))
}
\description{
Calculate indices of neighboring quadrats, for a given quadrat index.
}
