% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{fit.pdf}
\alias{fit.pdf}
\title{fit.pdf}
\usage{
fit.pdf(start.param = c(1, 1), data, pdf = dnorm, xrange = c(0, 1),
  badpar = default.badpar)
}
\description{
Fit a random variable x to any submitted probability distribution. The number of start parameters
must match what the pdf needs.
}
