% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fromjulian}
\alias{fromjulian}
\title{fromjulian}
\usage{
fromjulian(j, dateform = "\%m/\%d/\%Y")
}
\description{
Accepts a julian date and returns a character representation of date. See tojulian(). The input
can be vector or array.
}
\examples{
\dontrun{

fromjulian(1000,'\%d\%B\%Y')}





}
