% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_quadfunc.R
\name{getquadratname}
\alias{getquadratname}
\title{getquadratname}
\usage{
getquadratname(x, y, plotdim)
}
\description{
Convert x, y coordinates and plot dimensions into 4-character quadrat names. If x or y are missing, the quadrat=9999.
}
