% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{graph.modeldiag}
\alias{graph.modeldiag}
\title{Graph diagnostics of model.xygraph.modeldiag bootstrap.corr  Runnin...}
\usage{
graph.modeldiag(fit, burn = 500, llikelim = 50, graphpar = 1:3)
}
\description{
Graph diagnostics of model.xy
}
