% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spparea.R
\name{graph.quadrats}
\alias{graph.quadrats}
\title{Plot a series of quadrats with corners xlo, ylo, xhi, yhi.}
\usage{
graph.quadrats(coord, plotdim = c(1000, 500), clr = "black", add = FALSE)
}
\arguments{
\item{plotdim}{The x and y dimensions of the plot.}

\item{clr}{A string indicating colour. See argument col in [graphics::par()].}

\item{add}{Logical. Whether to add elements such as points or line to an
existing graph. if TRUE, there must be an appropriate existing graph,
otherwise, a new graph is drawn.}
}
\description{
Make a graph of a series of quadrats whose corners are given by the rows of
coord: xlo, ylo, xhi, yhi. This is used in illustrating randomly selected
quadrats.
}
