% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{graphFilledBand}
\alias{graphFilledBand}
\title{Fill the area between two curves on a graph. Useful for confidence ...}
\usage{
graphFilledBand(x, y1, y2, fillcol = "gray", add = TRUE,
  linecol = "black", ltype = "solid", lwidth = 1)
}
\arguments{
\item{x}{= The x axis values, as in any R graph}

\item{y1, }{y2 = Two sets of y axis values, each of exactly the same length as x}

\item{fillcol}{= The color filling the area between the two curves}

\item{add}{= TRUE or FALSE, as in other R graphs; if TRUE, there must be an appropriate existing graph, otherwise, a new graph is drawn}

\item{linecol}{= If add==FALSE, lines drawn at y1 and y2 are this color; ignored if add==TRUE}

\item{ltype}{= If add==FALSE, lines drawn at y1 and y2 are this type ('solid','dashed',etc);  ignored if add==TRUE}

\item{lwidth}{= If add==FALSE, lines drawn at y1 and y2 are this thickness (a number);  ignored if add==TRUE}
}
\description{
Fill the area between two curves on a graph. Useful for confidence limits, for example.

Typical use is to draw a graph first with some central y values, then add a confidence band by filling the area between upper and lower confidence limits
(designated by variables lower.y and upper.y in the example below). The central line should then be redrawn over the filled area.
}
\examples{
\dontrun{
plot(x,y,type='l')
graphFilledBand(x,lower.y,upper.y)
lines(x,y)}


}
