% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth.biomass.indiv}
\alias{growth.biomass.indiv}
\title{Like growth.indiv but based on agb growth, not dbh growth.}
\usage{
growth.biomass.indiv(census1, census2, mindbh = 10, steminfo = FALSE,
  dbhunit = "mm", err.limit = 4, maxgrow = 75, rounddown = NULL,
  exclude.stem.change = TRUE, pomcut = 10000)
}
\arguments{
\item{census1}{The R Analytical Table for a single census, either tree or
stem.}

\item{census2}{The matching R Analytical Table for a later census.}

\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than `mindbh` are excluded. If `NULL`, all living trees are
included.}

\item{dbhunit}{'cm' or 'mm', only used for basal area}

\item{maxgrow}{A number. Numbers such as 10000 are high and will return all
measures.}
}
\description{
Like \code{\link[=growth.indiv]{growth.indiv()}} but based on agb growth, not dbh growth. Extreme growth
rates (based on dbh growth) are excluded, but cases where the stemID changed
are not excluded.
}
\details{
Here pomcut is used in a very specific way probably only relevant at BCI. If
the second pom is higher than the first by more than the pomcut, the record
is not used. The function \code{\link[=trim.growth]{trim.growth()}} has already eliminated cases where
the stemID is unchanged and pom changes, so this will only serve for cases
where two different stemIDs have measurements. At BCI, in most cases where
the second pom is lower than the first and the stem changed, it is a
legitimate stem change. But where the second pom is higher, it is really the
same stem measured at a different pom, and with a different stemID because
BCI lacks stem tags.

For most plots, especially with stem tags, the default behavior means changes
in stem allow changes in pom to be included in biomass growth.
}
