% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth.dbh}
\alias{growth.dbh}
\title{growth.dbh}
\usage{
growth.dbh(census1, census2, classbreak = c(10, 100, 300), dbhunit = "mm",
  growthcol = "dbh", err.limit = 4, maxgrow = 75, rounddown = FALSE,
  method = "I", stdev = FALSE, mindbh = 10)
}
\description{
This calculates forest-wide growth in given dbh categories. Arguments as for growth().
}
\examples{
\dontrun{
growth.dbh=growth.eachspp(bci.full5,bci.full6,classbreak=c(10,50,100,300,500))}





}
