% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth.eachspp}
\alias{growth.eachspp}
\title{This calculates growth for each species in given dbh categories. It...}
\usage{
growth.eachspp(census1, census2, classbreak = c(10, 100, 300),
  dbhunit = "mm", mindbh = 10, growthcol = "dbh", err.limit = 4,
  maxgrow = 75, rounddown = FALSE, method = "I", stdev = FALSE)
}
\description{
This calculates growth for each species in given dbh categories. It creates the split
variables then uses growth(). Other arguments are as in growth().
}
\examples{
\dontrun{
growth.result=growth.eachspp(bci.full5,bci.full6,classbreak=c(10,50,100,300,500))}


}
