% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.bin.R
\name{growth.flexbin}
\alias{growth.flexbin}
\title{growth.flexbin}
\usage{
growth.flexbin(growthtable, sizecol = "dbh", nobin = 2, start = NULL,
  startsd = NULL, sdmodel = linear.model.ctr, badsdfunc = NULL,
  method = "Gibbs", rep = 1200, show = 100, burn = 200, ...)
}
\description{
Fitting a regression line through log growth against log dbh, binning on dbh. Growth table has 2 columns,
one named size (ie, dbh), other growth. The function now calls optim first to find a set of parameters, then, if Gibbs
is selected, runs the Gibbs sampler starting with those parameters. Since optim fails if the initial likelihood is -Inf,
it became necessary to check the parameters with bad.binparam before running.
}
