% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth.indiv}
\alias{growth.indiv}
\title{growth.indiv}
\usage{
growth.indiv(census1, census2, rounddown = FALSE, mindbh = 10,
  dbhunit = "mm", err.limit = 4, maxgrow = 75)
}
\description{
This returns a complete table with growth rate of every individual, both relative and dbh-increment. The table
also includes most other key pieces of information for every individual: species, dbh, coordinates. Growth is trimmed with trim.growth,
and growth is returned as NA if the individual is excluded; note, though, that every individual tree is always included in the table, even
if growth=NA. Arguments are the same as in growth().
}
\examples{
\dontrun{
growth.table=growth.indiv(bci.full5,bci.full6)}


}
