% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{gxgy.to.hectindex}
\alias{gxgy.to.hectindex}
\title{Converts GX GY Coordinates to a Hectare Number.}
\usage{
gxgy.to.hectindex(gx, gy, plotdim = c(1000, 500))
}
\arguments{
\item{gx, gy}{Tree x and y coordinate.}

\item{plotdim}{The x and y dimensions of the plot.}
}
\description{
Takes an x, y plot location and identifies the hectare number.
}
