% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{gxgy.to.index}
\alias{gxgy.to.index}
\title{gxgy.to.index}
\usage{
gxgy.to.index(gx, gy, gridsize = 20, plotdim = c(1000, 500))
}
\description{
Assign any location(s) a single index identifying the quadrat. The index runs from 1 to the number of quadrats.
}
