% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{gxgy.to.lxly}
\alias{gxgy.to.lxly}
\title{Given global coordinates and quadrat and plot dimensions, calculate...}
\usage{
gxgy.to.lxly(gx, gy, gridsize = 20, plotdim = c(1000, 500))
}
\description{
Given global coordinates and quadrat and plot dimensions, calculate local x and y, the within-quadrat coordinates
}
