% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{gxgy.to.quad}
\alias{gxgy.to.quad}
\title{Calculate a quadrat name from gy-gy.}
\usage{
gxgy.to.quad(gx, gy, gridsize = 20, plotdim = c(1000, 500), digits = 2,
  start = "zero")
}
\arguments{
\item{gx, gy}{Tree x and y coordinate.}

\item{gridsize}{Side of the square quadrate.}

\item{plotdim}{The x and y dimensions of the plot.}

\item{start}{If \code{start = "zero"}, quadrats start with \code{0000}, otherwise, they
start with \code{0101}.}
}
\description{
Calculate a quadrat name (column number then row number, as a 4-digit
character string) from gy-gy.
}
