% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{gxgy.to.rowcol}
\alias{gxgy.to.rowcol}
\title{Returns row and column for any set of coordinates. Rows and columns...}
\usage{
gxgy.to.rowcol(gx, gy, gridsize = 20, plotdim = c(1000, 500))
}
\description{
Returns row and column for any set of coordinates. Rows and columns both start at 1, not 0.
}
