% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{index.to.gxgy}
\alias{index.to.gxgy}
\title{Calculate the x and y coordinates given the quadrat index, as calcu...}
\usage{
index.to.gxgy(index, gridsize = 20, plotdim = c(1000, 500))
}
\description{
Calculate the x and y coordinates given the quadrat index, as calculated in gygy.to.index.
}
