% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{index.to.rowcol}
\alias{index.to.rowcol}
\title{Calculate the row and column given the quadrat index.}
\usage{
index.to.rowcol(index, gridsize = 20, plotdim = c(1000, 500))
}
\arguments{
\item{index}{Quadrate number, between 0 and 1249 in the standard plot.}

\item{gridsize}{Side of the square quadrate.}

\item{plotdim}{The x and y dimensions of the plot.}
}
\description{
Calculate the row and column given the quadrat index, as calculated in
gygy.to.index. Both row and column start at 1, not 0 as in quadrat naming.
}
