% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{index.to.rowcol}
\alias{index.to.rowcol}
\title{index.to.rowcol}
\usage{
index.to.rowcol(index, gridsize = 20, plotdim = c(1000, 500))
}
\description{
Calculate the row and column given the quadrat index, as calculated in gygy.to.index. Both row and column start at 1, not 0 as in quadrat naming.
}
