% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogChange.R
\name{individual_mort.table}
\alias{individual_mort.table}
\title{Create a table of individual trees and their survival status over t...}
\usage{
individual_mort.table(cnsdata, mindbh = 10, maxdbh = 10000,
  alivecode = c("A", "AB", "AS"), center = 1992)
}
\description{
Create a table of individual trees and their survival status over two censuses, with many species included.
}
\seealso{
individual_grow.table
}
