% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogChange.R
\name{individual_mort.table}
\alias{individual_mort.table}
\title{individual_mort.table}
\usage{
individual_mort.table(cnsdata = list(bci.full1, bci.full2, bci.full3,
  bci.full4, bci.full5, bci.full6, bci.full7), mindbh = 10, maxdbh = 10000,
  alivecode = c("A", "AB", "AS"), center = 1992)
}
\description{
Create a table of individual trees and their survival status over two censuses, with many species included.
}
