% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogChange.R
\name{individual_mort.table}
\alias{individual_mort.table}
\title{Table individual trees and their survival status over two censuses.}
\usage{
individual_mort.table(cnsdata, mindbh = 10, maxdbh = 10000,
  alivecode = c("A", "AB", "AS"), center = 1992)
}
\arguments{
\item{mindbh}{The minimum diameter above which the counts are done. Trees
smaller than \code{mindbh} are excluded. If \code{NULL}, all living trees are
included.}

\item{alivecode}{character, codes of the variable \code{status} that indicate the
tree is alive. The default 'A' is the standard CTFS designation for living
trees or stems}
}
\description{
Create a table of individual trees and their survival status over two
censuses, with many species included.
}
\seealso{
\code{\link[=individual_grow.table]{individual_grow.table()}}
}
