% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{intersection.line.curve}
\alias{intersection.line.curve}
\title{Finds the points where a straight line, given lines as intercept b ...}
\usage{
intersection.line.curve(b, m, curve)
}
\description{
Finds the points where a straight line, given lines as intercept b then slope m, intersects with a curve defined by a sequence of segments. Intersections must be found for the line with every segment, then every one checked with is.between. The intersections are returned as a dataframe of x, y coordinates. The b and m must be atomic. The curve must be a dataframe with columns x, y (or capital X, Y).
having any number of segments.
}
