% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.leap}
\alias{is.leap}
\title{Return a logical indicating which elements of a vector are leap yea...}
\usage{
is.leap(yr, start = 1904, end = 2096)
}
\description{
Return a logical indicating which elements of a vector are leap years.
}
