% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_utilities.R
\name{left}
\alias{left}
\title{left}
\usage{
left(s, n)
}
\description{
A standard left function. Returns the leftmost n characters of a string. If n<0, returns all except the rightmost n. If n==0, returns an
empty string. Arguments can be vectors, but both must be the same length, or n can be a scalar.
}
