% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{logistic.inter}
\alias{logistic.inter}
\title{Logistic function with intercept parameterization (i.e., first para...}
\usage{
logistic.inter(x, param, log = FALSE, ...)
}
\arguments{
\item{...}{Unused.}
}
\description{
Logistic function with intercept parameterization (i.e., first parameter is y
when all \code{x = 0}). The input x are all independent variables, in a matrix
with each column one of the variables. The number of rows is the number of
datapoints. Just one inter, which is the value at all \code{x = 0}, and passed as
\code{param[1]}. Slope parameters follow, one per column of \code{x}.

This is identical to standard.
}
