% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{logistic.multiplicative}
\alias{logistic.multiplicative}
\title{Logistic with a pair of parameters for each x; y=product of all the...}
\usage{
logistic.multiplicative(x, param, log = FALSE)
}
\description{
Logistic with a pair of parameters for each x; y=product of all the logistics. First set of parameters are intercepts, then
an equal number of slopes. If there are additional parameters, they are asymptote and basement.
}
