% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_distributions.R
\name{logistic.power.mode}
\alias{logistic.power.mode}
\title{logistic.power.mode}
\usage{
logistic.power.mode(x, param, log = FALSE, ...)
}
\description{
This is the Gaussian logistic function, where logit is a second-order polynomial of x, but with third parameter the position
of the critical point (peak or trough). Given 3 parameters for standard logistic.power, the mode is at -param[2]/2*param[3]).
Asymptote and basement are allowed. There must be 1+2*nopredictors parameters; the asympotote and basement are only implemented
if extra parameters are passed.
}
