% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{lxly.to.p5}
\alias{lxly.to.p5}
\title{lxly.to.p5}
\usage{
lxly.to.p5(lx, ly, gridsize = 20)
}
\description{
Given local, or  within-quadrat, coordinates for a 20-m quadrat, return the p5x5; lx and ly must be vectors of equal length. Any values outside [0,20) are returned p5=NA.
}
