% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map2species}
\alias{map2species}
\title{Map 2 species to a 2-panel pdf.}
\usage{
map2species(spp, data = list(bci3.spp, korup.spp),
  spplist = list(plotspp$bci, plotspp$korup),
  elev = list(ctfs.elev$bci, ctfs.elev$korup), path = "spp",
  export = pdf, file = pst(spp[1], spp[2], "map.pdf"),
  ptsize = c(0.6, 0.6), topo = c(2, 4), ht = 11, wd = 8.5,
  plotside = 4)
}
\arguments{
\item{spp}{A vector of two species names.}

\item{data}{A list of two split plot data objects.}

\item{spplist}{A list of two different species tables.}

\item{elev}{A list of two different elevation matrices.}

\item{export}{File format to export to.}
}
\description{
Map 2 species to a 2-panel pdf.
}
\details{
To make 2 maps from the same plot, each of the lists should repeated
the same data twice.
}
