% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{maptopo}
\alias{maptopo}
\title{Draws contours on a plot map.}
\usage{
maptopo(elevmat, plotdim, add = FALSE, new = FALSE, export = "no",
  interval = 0, xrange = NULL, yrange = NULL, xaxis = TRUE,
  yaxis = TRUE, axspos = 0.5, labelsize = 1.15, maintitle = "",
  ht = 6, wd = 9, plotside = 4.5, filepath = "", clr = "black",
  bgcolor = "transparent")
}
\arguments{
\item{elevmat}{Matrix of elevation data (see ?\link[bci:bci_elevation]{bci::bci_elevation}).}

\item{plotdim}{The x and y dimensions of the plot.}

\item{add}{Logical. Whether to add points or line to an existing graph. if
TRUE, there must be an appropriate existing graph, otherwise, a new graph
is drawn.}

\item{new}{See \code{\link[graphics:par]{graphics::par()}}.}

\item{export}{`export = "no"`, does nothing, so the next graph uses the
default R device. With this option, the size of the graph cannot have been
set larger than the default size, which is usually 7 inches. `export =
"unix"`, a new graphics window in unix, size `h` and width `w`, will be
opened allowing you to alter the size of the output. `export = "windows"`,
or `export = "mac"` are similar to `export = "unix"` but for windows and
mac. `export = format` exports the graph to a specific file format: "pdf",
"png", "jpg", "bmp", "emf".}

\item{interval}{Contour line intervals.}

\item{xrange}{Minimum and maximum x and y coordinates of area graphed.}

\item{yrange}{Minimum and maximum x and y coordinates of area graphed.}

\item{xaxis}{xxxdocparam}

\item{yaxis}{xxxdocparam}

\item{axspos}{Distance between axis numbers and axis.}

\item{labelsize}{xxxdocparam}

\item{maintitle}{A title to appear at the top of the page, above the species
name.}

\item{ht}{Height and width of the overall graph.
The default work for pdf export or mapping to the screen, and ht and wd are
inches. But if export is png, jpg, emf, height and width are pixels and
need to be 500-1000.}

\item{wd}{Height and width of the overall graph.
The default work for pdf export or mapping to the screen, and ht and wd are
inches. But if export is png, jpg, emf, height and width are pixels and
need to be 500-1000.}

\item{plotside}{The vertical dimension (inches) of the map.}

\item{filepath}{The folder to which map will be exported.}

\item{clr}{xxxdocparam}

\item{bgcolor}{The background color. Defaults to white. For presentation
exports, try bgcolor="transparent".}
}
\description{
Draws contours on a plot map. Elevmat has elevation data in matrix form; plot
dimensions are not needed, instead it calculates them from elevmat and the
gridsize. If add=TRUE, they are added to an existing map. If new=TRUE, a new
screen window is created.
}
\seealso{
\code{\link[=map]{map()}}, \link[bci:bci_elevation]{bci::bci_elevation}.
}
