% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{maptopo}
\alias{maptopo}
\title{Draws contours on a plot map.}
\usage{
maptopo(elevmat, plotdim, add = FALSE, new = FALSE, export = "no",
  interval = 0, xrange = NULL, yrange = NULL, xaxis = TRUE,
  yaxis = TRUE, axspos = 0.5, labelsize = 1.15, maintitle = "",
  ht = 6, wd = 9, plotside = 4.5, filepath = "", clr = "black",
  bgcolor = "transparent")
}
\arguments{
\item{elevmat}{Matrix of elevation data (see ?\link[bci:bci_elevation]{bci::bci_elevation}).}

\item{plotdim}{The x and y dimensions of the plot.}

\item{add}{Logical. Whether to add points or line to an existing graph. if
TRUE, there must be an appropriate existing graph, otherwise, a new graph
is drawn}

\item{new}{See \code{\link[graphics:par]{graphics::par()}}.}

\item{export}{See function define.graphwindow.}

\item{interval}{Contour line intervals.}

\item{xrange}{Minimum and maximum x coordinates and y coordinates to
map. Allows a portion of plot to be drawn. Defaults to the entire plot.}

\item{yrange}{Minimum and maximum x coordinates and y coordinates to
map. Allows a portion of plot to be drawn. Defaults to the entire plot.}

\item{xaxis}{xxxdocparam}

\item{yaxis}{xxxdocparam}

\item{axspos}{Distance between axis numbers and axis.}

\item{labelsize}{xxxdocparam}

\item{maintitle}{A title to appear at the top of the page, above the species name.}

\item{ht}{These are the height and width of the overall graph,
and the vertical dimension (inches) of the map. The default work for pdf
export or mapping to the screen, and ht and wd are inches. But if export is
png, jpg, emf, height and width are pixels and need to be 500-1000.}

\item{wd}{These are the height and width of the overall graph,
and the vertical dimension (inches) of the map. The default work for pdf
export or mapping to the screen, and ht and wd are inches. But if export is
png, jpg, emf, height and width are pixels and need to be 500-1000.}

\item{plotside}{These are the height and width of the overall graph,
and the vertical dimension (inches) of the map. The default work for pdf
export or mapping to the screen, and ht and wd are inches. But if export is
png, jpg, emf, height and width are pixels and need to be 500-1000.}

\item{filepath}{The folder to which map will be exported.}

\item{clr}{xxxdocparam}

\item{bgcolor}{The background color. Defaults to white. For presentation
exports, try bgcolor="transparent".}
}
\description{
Draws contours on a plot map. Elevmat has elevation data in matrix form; plot
dimensions are not needed, instead it calculates them from elevmat and the
gridsize. If add=TRUE, they are added to an existing map. If new=TRUE, a new
screen window is created.
}
\seealso{
\code{\link[=map]{map()}}, \link[bci:bci_elevation]{bci::bci_elevation}.
}
