% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{maptopo}
\alias{maptopo}
\title{maptopo}
\usage{
maptopo(elevmat, plotdim, add = FALSE, new = FALSE, export = "no",
  interval = 0, xrange = NULL, yrange = NULL, xaxis = TRUE,
  yaxis = TRUE, axspos = 0.5, labelsize = 1.15, maintitle = "",
  ht = 6, wd = 9, plotside = 4.5, filepath = "", clr = "black",
  bgcolor = "transparent")
}
\description{
Draws contours on a plot map. Elevmat has elevation data in matrix form; plot
dimensions are not needed, instead it calculates them from elevmat and the
gridsize. If add=TRUE, they are added to an existing map. If new=TRUE, a new screen window is created.

This is called by map(), and see the description of map() for the other arguments. It can be used
on its own though.
}
