% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match.dataframe}
\alias{match.dataframe}
\title{match.dataframe}
\usage{
match.dataframe(x, y)
}
\description{
Matches two dataframes using two or more columns. R's function match() works only on vectors 
(and thus single columns only). The return is a vector of indices, exactly as match() does.
}
