% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mergeParam}
\alias{mergeParam}
\title{Merges a list of parameter matrices into one large matrix. Used for...}
\usage{
mergeParam(p)
}
\description{
Merges a list of parameter matrices into one
large matrix. Used for the parameter output from MCMCmetrop1R, stored
as a list. No longer used; superseded by merge_data.
}
