% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{mortality.dbh}
\alias{mortality.dbh}
\title{mortality.dbh}
\usage{
mortality.dbh(census1, census2, classbreak = c(10, 100, 300),
  alivecode = c("A", "AB", "AS"))
}
\description{
Calculate forest-wide mortality in given dbh categories. See mortality and mortality.eachspp, which have same arguments and same output format.
}
