% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{mortality.eachspp}
\alias{mortality.eachspp}
\title{Calculate mortality for each species in given dbh categories. It se...}
\usage{
mortality.eachspp(census1, census2, classbreak = c(10, 100, 300),
  alivecode = c("A", "AB", "AS"))
}
\description{
Calculate mortality for each species in given dbh categories. It sets the split variables using the species name and
submitted dbh classbreaks and then uses mortality to do the calculation. See argument descriptions for mortality. Return object
is the list from mortality and can be passed to assemble.demography for a convenient format.
}
\examples{
\dontrun{

CTFSplot("bci",5:6)
mort.data=mortality.eachspp(bci.full5,bci.full6)
mort.table1=assemble.demography(mort.data,type="m",whichdbhcat=1)
mort.table2=assemble.demography(mort.data,type="m",whichdbhcat=2)
mort.table3=assemble.demography(mort.data,type="m",whichdbhcat=3)}


}
