% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mvrnormRC}
\alias{mvrnormRC}
\title{mvrnormRC}
\usage{
mvrnormRC(N, Sigma)
}
\description{
Function that takes a variance-covariance matrix and produces normal variates
following it, but with means 0. The R function mvrnorm does this too; this was a
test of the algorithm from Tommaso Zillio. Sigma must be square. N is the number
to draw.
}
