% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborDensityFun.R
\name{NDcount}
\alias{NDcount}
\title{NDcount}
\usage{
NDcount(censdata, r = 20, plotdim = c(1000, 500))
}
\arguments{
\item{censdata}{Is the coordinates dataset, should have the coordinates (gx, gy)}

\item{r}{(20) radius 
@param plotdim  (c(1000,500))  vector with two numbers indicating the xy-plot size;}
}
\description{
Calculates the count of neighbors within radius r with edge effect correction. Is a quick version of NeighborDensities with limited capabilities. 
Dependencies: CalcRingArea from the CTFS package and splancs. 
Output: a single vector with neighbor counts within radius r for each individual in censdata.
}
\examples{
\dontrun{
##' count all the consp neighbors of one species 
> one.sp = subset(bci.full7, sp=="ingasa" & status=="A")
> neighbor.counts <- NDcount(one.sp)}

}
\author{
Tania Brenes
}
