% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{normalproduct}
\alias{normalproduct}
\title{normalproduct}
\usage{
normalproduct(x, lag, mean1, sd1, meanint, meanslope, CV)
}
\description{
A function which returns the product of 2 normal distributions, the first
at x (a vector), the second at lag-x (lag is a scalar). The mean and SD 
of the second normal are linear functions of x, with meanint being the
intercept, meanslope the slope, and CV the coefficient of variation.  
A convolution!
}
