% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthfit.graph.R
\name{overlay.growthbinmodel}
\alias{overlay.growthbinmodel}
\title{overlay.growthbinmodel}
\usage{
overlay.growthbinmodel(fit, bins = 1:4, regclr = "green",
  modelclr = "blue", graphdiv = 15, add = FALSE, newgraph = TRUE,
  export = pdf, outfile = "growth/linearbin.overlay.pdf", h = 8, w = 10)
}
\description{
Show model fits for 1, 2, 3, and 4 bins on each species
}
