% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{parallel.line}
\alias{parallel.line}
\title{Finds the slope and intercept of the line parallel to a line whose ...}
\usage{
parallel.line(b, m, x, y)
}
\description{
Finds the slope and intercept of the line parallel to a line whose slope and intercept are given,
through the points x,y. Note that the intercept is not needed. For any m that are infinite, the intercept is x.
}
