% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_map.R
\name{pdf.allplot}
\alias{pdf.allplot}
\title{pdf.allplot}
\usage{
pdf.allplot(splitdata = bci.split6, spplist = bci.spptable,
  elev = ctfs.elev$bci$mat, plotdim = c(1000, 500), cutoff = c(10, 100,
  300, 3000), topoint = 2, bw = FALSE, topoclr = "gray80", bty = "o",
  plotside = 9, h = 8, w = 10, size = c(0.3, 0.45, 0.65),
  deadtree = FALSE, symbols = NULL, legsize = 0.75,
  legpos = c(plotdim[1]/2, 1.16 * plotdim[2]), export = TRUE,
  singlefile = TRUE, path = "/home/condit/data/maps/bci/")
}
\description{
Export a pdf with one or more species maps. If singlefile=TRUE, 
all maps will be in one big pdf, otherwise, a pdf for every species is created. The file exported will be named with Map.pdf
in the path name given.  
This calls the map() function for all species
in the splitdata list; it is designed for all species from a plot (default is all species from the BCI plot).
See description of map() for details.
}
