% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{perpendicular.distance}
\alias{perpendicular.distance}
\title{Distance from a point to a line (so its the perpendicular distance).}
\usage{
perpendicular.distance(b, m, x, y)
}
\arguments{
\item{b}{Intercept.}

\item{m}{Slope.}

\item{x}{Coordinates.}

\item{y}{Coordinates.}
}
\description{
Distance from a point to a line (so it's the perpendicular distance).
}
\details{
Find distance from a pt (x then y coords) to a line segment given as start
and end points. Either x, y or end points can be vectors, but not both.

It first finds perpendicular distance, then distance to each end point, and
returns the minimum.
}
\section{Arguments details}{

If both \code{b, m} and \code{x, y} are vectors, they must all be same length.

Note check for infinite slope, meaning that the intercept b is the
x-intercept.
}

