% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{perpendicular.line}
\alias{perpendicular.line}
\title{Finds the slope and intercept of the line perpendicular to a line w...}
\usage{
perpendicular.line(b, m, x, y)
}
\description{
Finds the slope and intercept of the line perpendicular to a line whose slope and intercept are given,
through the points x,y. Either x and y can be vectors, or b and m can be vectors, or if all are vectors, must be same length.
}
