% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{png.allplot}
\alias{png.allplot}
\title{png.allplot}
\usage{
png.allplot(splitdata = bci.split6, spplist = plotspp$bci,
  elev = ctfs.elev$bci$mat, plotdim = c(1000, 500), topoint = 2,
  plotside = 9, h = 850, w = 1100, ptsizes = c(0.3, 0.45, 0.65),
  dbhcut = c(10, 100, 300, 3000), dead = FALSE, export = TRUE,
  path = "/home/condit/data/maps/bci/")
}
\description{
Export species maps as png. There will be one for every species chosen. See pdf.allplot.
}
