% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{polar.to.cartesian}
\alias{polar.to.cartesian}
\title{Convert polar coordinates to Cartesian.}
\usage{
polar.to.cartesian(r, theta)
}
\arguments{
\item{r}{A vector of radii (distance from origin).}

\item{theta}{input vector of angle from horizontal (radians), same length as r}
}
\value{
A dataframe of two columns named x and y.
}
\description{
Convert polar coordinates to Cartesian.
}
\examples{
\dontrun{
polar.to.cartesian(2, pi / 3)
}

}
