% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rasympower}
\alias{rasympower}
\title{Random draws from the bilateral power distribution, dasympower.}
\usage{
rasympower(n, rate1, rate2, c)
}
\arguments{
\item{n}{Passed to argument \code{n} in \code{\link[stats:runif]{stats::runif()}}.}
}
\description{
Random draws from the bilateral power distribution, \code{\link[=dasympower]{dasympower()}}.
}
\seealso{
\code{\link[stats:runif]{stats::runif()}}
}
\author{
R. Chisholm.
}
