% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rasympower}
\alias{rasympower}
\title{Random draws from the bilateral power distribution, dasympower. By ...}
\usage{
rasympower(n, rate1, rate2, c)
}
\description{
Random draws from the bilateral power distribution, dasympower. By R. Chisholm.
}
