% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{readelevdata}
\alias{readelevdata}
\title{Read a table of elevation data into a matrix appropriate for mapping.}
\usage{
readelevdata(elevfile, gridsize = 5, text = FALSE)
}
\arguments{
\item{elevfile}{A data frame with x, y coordinates and elevation,
bcielev.info.}

\item{gridsize}{scale of elevation values (m)}

\item{text}{Logical. Defaults to read from a datatrame. Set to TRUE to read a
text table.

@seealso \link[bci:bci_elevation]{bci::bci_elevation}.}
}
\value{
A matrix, as needed for R's contour function.
}
\description{
A function to read a table of elevation data into a matrix appropriate for
mapping. The table must have x then y coordinates, followed by elevation, at
every corner across the plot, using grid of \code{gridsize}.
}
