% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerBayes.R
\name{recalculate.lmerBayesllike}
\alias{recalculate.lmerBayesllike}
\title{Walk through entire chain of parameters to calculate full likelihoo...}
\usage{
recalculate.lmerBayesllike(fit, model, error, bad = NULL, keep = NULL,
  dup = TRUE, ...)
}
\arguments{
\item{fit}{A fitted model.}
}
\description{
Walk through entire chain of parameters to calculate full likelihood at each
step, as was done during the model run. The argument keep defines the
elements to be used, or if NULL, fit$keep is used.
}
