% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_recruitment.R
\name{recruitment.eachspp}
\alias{recruitment.eachspp}
\title{recruitment.eachspp}
\usage{
recruitment.eachspp(census1, census2, mindbh = 10, alivecode = c("A", "AB",
  "AS"))
}
\description{
A wrapper to calculate recruitment for each species.
}
