% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{regress.loglog}
\alias{regress.loglog}
\title{Performs regression and graphs in a convenient way.}
\usage{
regress.loglog(x, y, xlog = TRUE, ylog = TRUE, addone = NULL,
  graphit = TRUE, xrange = NULL, yrange = NULL, add = FALSE, pts = 16,
  lwidth = 1, drawline = "solid", title = "", xtitle = NULL,
  ytitle = NULL, ptnames = NULL, ptsize = 1, clr = "blue",
  lineclr = "red", includeaxis = TRUE)
}
\arguments{
\item{x, y}{Variables on which to perform regression and graph.}

\item{xlog, ylog}{Logical. Set to \code{TRUE} to log-transform x and y variables.}

\item{addone}{Use to handle zeros for log-transformation}

\item{graphit}{Logical; if `TRUE` adds plot output.}

\item{add}{Logical. Whether to add elements such as points or line to an
existing graph. if TRUE, there must be an appropriate existing graph,
otherwise, a new graph is drawn.}

\item{pts}{See argument \code{pch} in \code{\link[graphics:par]{graphics::par()}}.}

\item{lwidth}{Line width; see `width` in ?[graphics::par()].}

\item{ptsize}{Passed to argument `cex` in `graphics::par()`.}

\item{clr}{A string indicating colour. See argument col in [graphics::par()].}

\item{ltype}{Line type; see `ltype` in ?[graphics::par()].}
}
\description{
Performs regression and graphs in a convenient way: with or without
log-transforming x and y variables (the option addone can be included to
handle zeros for log-transformation), with or without manual point labelling,
without or without the best-fit line added, and with many options for colors
and points. add can be a vector of length 2, a constant to be added to every
value of x, y to remove zeroes.
}
\seealso{
\code{\link[graphics:par]{graphics::par()}}.
}
