% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{regression.Bayes}
\alias{regression.Bayes}
\title{Regression using the Gibbs sampler, with just one x variable.}
\usage{
regression.Bayes(x, y, start.param = NULL, steps = 1100,
  whichuse = 101:1100, showstep = 50)
}
\arguments{
\item{steps}{The number of steps to run the Gibbs sampler.}

\item{showstep}{Information is printed to the screen every showstep steps.}
}
\description{
Regression using the Gibbs sampler, with just one x variable.
}
