% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{regsum}
\alias{regsum}
\title{For a polynomial regression, find x or y at y's peak or x-intercept.}
\usage{
regsum(x, y, poly = 1, graphit = F, yrange = c(-1, -1),
  yaxslab = "multiplier", newgraph = F, add = F, pts = T, ptype = 16,
  ltype = "solid")
}
\arguments{
\item{graphit}{Logical; if `TRUE` adds plot output.}
}
\description{
This carries out either first or second order polynomial regression,
finds the x- and y-values at y's peak if its second order,
otherwise the x-intercept.
}
