% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{right}
\alias{right}
\title{Returns the rightmost n characters of a stringright rightbut  Retur...}
\usage{
right(s, n, bytetype = "bytes")
}
\arguments{
\item{s}{Character vector.}

\item{n}{Scalar of vector of the same lenght as \code{s}. If n < 0, returns all
except the rightmost n. If n == 0, returns an empty string.}
}
\description{
Returns the rightmost n characters of a string.
}
\seealso{
\code{\link[=left]{left()}}, \code{\link[stringr:str_sub]{stringr::str_sub()}}.
}
