% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{rowcol.to.index}
\alias{rowcol.to.index}
\title{Takes row and column numbers and identifies the quadrate number (index).}
\usage{
rowcol.to.index(rowno, colno, gridsize = 20, plotdim = c(1000, 500))
}
\arguments{
\item{rowno}{Row number.}

\item{colno}{Column number.}

\item{gridsize}{Side of the square quadrate.}

\item{plotdim}{The x and y dimensions of the plot.}
}
\description{
Takes row and column numbers and identifies the quadrate number (index). The
row and column numbers are based on a \code{gridsize} that by default divides the
plot into 20 by 20 m squares. The g\code{ridsize} can be defined by the user so
other quadrate sizes can be used.
}
