% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadfunc.R
\name{rowcol.to.index}
\alias{rowcol.to.index}
\title{None given.rowcol.to.index gxgy.to.quad  Calculate a quadrat name (...}
\usage{
rowcol.to.index(rowno, colno, gridsize = 20, plotdim = c(1000, 500))
}
\description{
None given.
}
