% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rpois.trunc}
\alias{rpois.trunc}
\title{rpois.trunc}
\usage{
rpois.trunc(N, lambda)
}
\description{
Random draws on dpois.trunc. This is taken unchanged from an answer Peter Dalgaard posted to a list serve in 2005. I checked
by comparing to dpois.trunc and it was spot on.
}
