% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rpower}
\alias{rpower}
\title{Random draws based on the integral.rpower dasympower  A bilateral p...}
\usage{
rpower(n, beta)
}
\arguments{
\item{n}{Passed to argument `n` in [stats::runif()].}
}
\description{
Random draws based on the integral.
}
\seealso{
[stats::runif()]
}
