% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerBayes.R
\name{saveParamFile}
\alias{saveParamFile}
\title{This saves a run of the full parameters into a text file, reducing ...}
\usage{
saveParamFile(param, paramfile, rand, firsttime = FALSE)
}
\description{
This saves a run of the full parameters into a text file, reducing the amount of memory needed.
}
