% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spparea.R
\name{selectrandomquad2}
\alias{selectrandomquad2}
\title{Creates randomly drawn quadrats, using same arguments and producing...}
\usage{
selectrandomquad2(size, rep, plotdim = c(1000, 500), graphit = FALSE)
}
\description{
Creates randomly drawn quadrats, using same arguments and producing same
return value as selectrandomquad, but using a different algorithm aimed at
capturing corners. The result, however, is not to capture corners any better
than selectrandomquad() does.

Imagine a line running vertically at \code{x = 0} from \code{y = 0} to
\code{y = plotdim[2] - size}, then continues at \code{x = 1} from 0 to
\code{plotdim[2] - size}, etc. It's wrapping analogous to the way quadrat indices
wrap (see \code{?gxgy.to.index}). This line has length
\code{(plotdim[1] - size) * (plotdim[2] - size)}.

Draw a random number on that line, and place the lower left corner of random
square at that point.

A position \code{x} on the line is converted to plot coordinates \code{gx}, \code{gy} using
function \code{index.to.gxgy()} with \code{grid = 1}.
}
