% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spparea.R
\name{selectrandomquad3}
\alias{selectrandomquad3}
\title{Creates randomly drawn quadrats, using same arguments and producing...}
\usage{
selectrandomquad3(size, shape, rep, plotdim = c(1000, 500), graphit = FALSE)
}
\arguments{
\item{size}{a vector of quadrat sizes, referring to the x-dimension of a
rectangular quadrat}

\item{shape}{the ratio of y to x dimensions of the rectangles; rectdim=1 (the
default) for squares}

\item{rep}{the number of random quadrats to create, of each size}

\item{plotdim}{the x and y dimensions of the entire plot}

\item{graphit}{whether to graph the locations of the chosen quadrats on a
plot map}
}
\description{
Creates randomly drawn quadrats, using same arguments and producing same return value as selectrandomquad, but using a different algorithm.

The lower left x, y coordinates are drawn at random from a range extending outside the plot dimensions, then quadrats are used only if
all 4 corners fall inside the plot.
}
