% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_map.R
\name{setsize}
\alias{setsize}
\title{setsize}
\usage{
setsize(n, s)
}
\description{
An internal function, called by map() to choose size of plotting points. It uses the number of dbh categories (n) and abundance of the species (s) 
to help determine the size. Note that the user can override these defaults by submitting sizes.
}
