% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvetopo.R
\name{solve_topo}
\alias{solve_topo}
\title{This is based on the problem and solution outlined in my book on pl...}
\usage{
solve_topo(coldata, column1 = "pt1", column2 = "pt2",
  diffcolumn = "htdiff", basept = "1", baseelev = 0, debug = NULL)
}
\description{
This is based on the problem and solution outlined in my book on plot methods
(Condit, 1998). Each of N stakes i has an estimated height \code{E[i]} and a true
height \code{e[i]}.
Pairs of stakes have a height difference \code{d[i, j]}, where
\code{e[i] + d[i, j] = e[j]}, but only estimated height differences \code{D[i, j]} are
known.
The least-squares estimate \code{E[i]} of \code{e[i]} can be written as the mean of the
\code{n[i]} points j for which \code{D[i, j]} was measured:
\code{E[i] = mean(E[j] - D[i, j]) =}
\code{(1 / n[i]) * sum(E[j]) - (1 / n[i]) * sum(D[i, j])}, and
\code{n[i] * E[i] - sum(E[j]) = -sum(D[i, j])}.
The latter produces N equations in N unknowns, but they are exactly singular.
One of the points must be assigned a value, and it's easiest to set
\code{e[1] = 0}.
The effect is to exclude the equation for \code{i = 0}, but all j for which
\code{D[0, j]} was measured have unchanged equations (\code{D[0, j]} are included).
The equations are written in matrix form. The coefficients are a matrix M
whose diagonal elements are the \code{n[i]}, so \code{M[i, i] = n[i]} for all i.
All other entries are \code{-1} where \code{D[i, j]} was measured otherwise zero.
The estimated \code{E[i]} are a vector, and the vector \code{V[i] = -sum(D[i, j])},
where the summation is over j only. Then \code{M *PERCENT* E = V}, \code{Minv} is the
inverse of M, and \code{E = Minv *PERCENT* V}.
}
\details{
In theory, the program will accept duplicate measures of the same pair of
plots. They are treated as replicates with equal weight to all other
estimates. There would be -2 or -3 etc. off the diagonal. The data are in
columns. One column has the label of one stake (column header pt1), the
second has the label of the second stake (column header pt2), and the final
column has the height difference at pt2 minus pt1 (column header diff). The
point labelled basept is assigned elevation baseelev. The last 5 arguments
allow those column headers to be reassigned.

The name solve.topo clashed with an S3 method, so it was replaced by
solve_topo.
}
\examples{
\dontrun{
# See topography tutorial
}

}
