% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spparea.R
\name{spparea.sq}
\alias{spparea.sq}
\title{spparea.sq}
\usage{
spparea.sq(censdata, spcolumn = "sp", size, rectdim = 1, mindbh = NULL,
  plotdim = c(1000, 500), replicates = 10, unidennames = c("**", "UNID",
  "uniden", "UNIDEN"))
}
\arguments{
\item{censdata}{one R Analytical Table, either full or stem}

\item{spcolumn}{name of the column in the table having the species; defaults to 'sp', but can be set to 'genus'for 'family'if desired}

\item{size}{a vector of quadrat sizes, referring to the x-dimension of a rectangular quadrat}

\item{rectdim}{the ratio of y to x dimensions of the rectangles; rectdim=1 (the default) for squares}

\item{mindbh}{the minimum dbh included}

\item{plotdim}{the x and y dimensions of the entire plot}

\item{replicates}{the number of random quadrats to create, of each size}

\item{unidennames}{a vector of species names that should not be included in species counts (see the function unidentified.species()}
}
\description{
Function for calculating the number of species in replicated, randomly-placed
quadrats of various areas. The variable size refers to the dimension (side)
of the square, and can be a vector; replicates is the
number of random draws per dimension. Full plot data are submitted as censdata.

Species are counted as the number of unique values of sp in the R Analytical Table; unidentified species are not counted,
based on a default unidennames; see the function unidentified.species in utilitiesCTFS.r.

The return value is a list of two components. The first is a table giving the mean number (and SD) of individuals and species
in all quadrat sizes submitted. The second is a table giving the number of individuals and species in every random quadrat created.

In addition, a graph of species vs. individuals in every quadrat is created as the program runs.

This can also be used to calculate genus- or family-area curves with use of the spcolumn argument. The censdata table must have
a new column added, for example the genus for every record, then spcolumn can be set to 'genus'.

Note: randomly-placed quadrats produce statistically preferable species-area curves than checkerboards of non-overlapping quadrats. If
required, though, the function abundanceperquad() in abundance.r offers a fast way to count the number of species in checkerboard-type
quadrats of different sizes.
}
\examples{
\dontrun{
speciesPerArea=spparea.sq(bci.full6,size=c(10,20,50),mindbh=10,plotdim=c(1000,500),replicates=5,unidennames=c('unid'))
rowmatch=match(bci.full6$sp,bci.spptable$sp)
bci.full6$Genus=bci.spptable$Genus[rowmatch]
genusPerArea=spparea.sq(bci.full6,spcolumn='Genus',size=c(10,20,50),mindbh=10,plotdim=c(1000,500),replicates=5,unidennames=c('unid'))}


}
