% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerBayes.R, R/modelBayes.R
\name{summaryMCMC}
\alias{summaryMCMC}
\alias{summaryModelMCMC}
\alias{summaryModelMCMC}
\title{Summarize calculations based on the full Gibbs sampler.}
\usage{
summaryMCMC(fit, model, error, sdmodel, badparam, paramfile = NULL,
  returnfull = TRUE, ...)

summaryModelMCMC(fit, model, error, sdmodel, badparam, paramfile = NULL,
  returnfull = TRUE, ...)
}
\arguments{
\item{fit}{An object holding all steps of the sampler, plus data,
observations, and likelihood.}

\item{paramfile}{If parameters were saved along the way to a text file, then
this argument is used to name the file and restore them into a 3D array.}

\item{returnfull}{If TRUE, the entire 3D array of parameters is also
returned. Full likelihood at the best parameters is calculated and likelihood
at each step in sampler are used to calculate DIC.}
}
\value{
Estimates of confidence limits of all parameters are returned.
}
\description{
Make summary calculations based on the full Gibbs sampler.
}
\section{Functions}{
\itemize{
\item \code{summaryModelMCMC}: See \code{\link[=summaryMCMC]{summaryMCMC()}}.
}}

