% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TextToRdata}
\alias{TextToRdata}
\title{Reads a tab-delimited text file and save as rdata.}
\usage{
TextToRdata(infile, outfile, object, NAs = "NA")
}
\arguments{
\item{outfile}{Filename for output.}
}
\description{
Reads a tab-delimited text file and save as rdata.
}
