% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tojulian}
\alias{tojulian}
\title{Convert calendar dates to Julian dates.}
\usage{
tojulian(x, dateform = "\%m/\%d/\%Y")
}
\arguments{
\item{x}{A character representation of a date.}

\item{dateform}{Format for calendar date, see \code{\link[base:strptime]{base::strptime()}} for format
information.}
}
\value{
A julian date, the number of days since 1 Jan 1960; a julian is an
integer and can be graphed or operated as such, though it displays as a
date.
}
\description{
Accepts any character representation of a date and a description of the
format. The submitted dates can be a vector or a matrix.
}
\examples{
\dontrun{
tojulian(c('23Oct2010','29Mar1956'),'\%d\%b\%Y')
}

}
\seealso{
\code{\link[base:strptime]{base::strptime()}}, \code{\link[base:julian]{base::julian()}}.
}
