% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.analysis.R
\name{wavelet.allsp}
\alias{wavelet.allsp}
\title{wavelet.allsp}
\usage{
wavelet.allsp(censdata, plotdim = c(1000, 500), gridsize = 2.5,
  mindbh = NULL)
}
\description{
Function to calculate the wavelet variance curves for all species in one plot using quadrat indexation from CTFS package.
}
\examples{
\dontrun{
load("bci.full1.rdata") 
wavelet.variances = wavelet.allsp(censdata, plotdim=c(1000,500))}


}
